% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{bindfiles}
\alias{bindfiles}
\title{Bind Files into Dataframe}
\usage{
bindfiles(files)
}
\arguments{
\item{files}{Dataframe}
}
\description{
This function imports and binds multiple files into a single 
data.frame. Files must be inserted with absolute roots filenames.
}
