% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Sampling Scores}
\usage{
calibrate(score, train, target, train_sample, target_sample)
}
\arguments{
\item{score}{Vector. Probability predictions of the model output}

\item{train}{Integer. Total row count in the training dataset}

\item{target}{Integer. Total row count of the target class 
in the training dataset}

\item{train_sample}{Integer. Total row count in the training 
dataset after sampling}

\item{target_sample}{Integer. Total row count of the target 
class in the training dataset after sampling}
}
\description{
This function lets the user calibrate a model's predictions when 
under or over sampling methods were applied when training it.
}
