% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{categoryCounter}
\alias{categoryCounter}
\title{Count Categories on a Dataframe}
\usage{
categoryCounter(df)
}
\arguments{
\item{df}{Categorical Vector}
}
\description{
This function lets the user count unique values in a categorical 
dataframe
}
