% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{chatgpt_ask}
\alias{chatgpt_ask}
\title{ChatGPT API Interaction with R}
\usage{
chatgpt_ask(
  ask,
  secret_key = get_credentials()$openai$secret_key,
  url = "https://api.openai.com/v1/chat/completions",
  model = "gpt-3.5-turbo",
  quiet = FALSE
)
}
\arguments{
\item{ask}{Character. Redacted prompt to ask ChatGPT.}

\item{secret_key}{Character. Secret Key. Get yours in:
\href{https://platform.openai.com}{platform.openai.com}.}

\item{url}{Character. Base URL for OpenAI's ChatGPT API.}

\item{model}{Character. OpenAI model to use.}

\item{quiet}{Boolean. Keep quiet? If not, show informative messages.}
}
\value{
(Invisible) list. Content returned from API POST.
}
\description{
This function lets the user ask ChatGPT via its API, and returns
the rendered reply.
}
\examples{
\dontrun{
api_key <- get_credentials()$openai$secret_key
chatgpt_ask("Can you write an R function to plot a dummy histogram?", api_key)
}
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_process}()},
\code{\link{fb_report_check}()},
\code{\link{fb_rf}()},
\code{\link{fb_token}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}
}
\concept{API}
