% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{cleanText}
\alias{cleanText}
\title{Clean text}
\usage{
cleanText(text, spaces = TRUE)
}
\arguments{
\item{text}{Character Vector}

\item{spaces}{Boolean. Keep spaces?}
}
\description{
This function lets the user clean text into getting only alphanumeric 
characters and no accents/symbols on letters.
}
