% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer}
\usage{
dalex_explainer(df, model, model_name = "lares::h2o_automl", pred = NA)
}
\description{
DALEX library function to create an explainer object
}
