% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_local}
\alias{dalex_local}
\title{DALEX Local}
\usage{
dalex_local(explainer, observation = NA, row = 1, type = "break_down")
}
\arguments{
\item{explainer}{Object. Result from h2o_explainer function}

\item{observation}{Data.frame. If you want to use an observation
that was not in the original explainer function, add here. Else, use row}

\item{row}{Dataframe. Row number from the data.frame used in explainer.}

\item{type}{Character. The type of variable attributions.
Either shap, oscillations, break_down or break_down_interactions.}
}
\value{
List. Containing observation, breakdown results, and breakdown plot.
}
\description{
DALEX function for local interpretations
}
\seealso{
Other Interpretability:
\code{\link[=dalex_residuals]{dalex_residuals()}},
\code{\link[=dalex_variable]{dalex_variable()}},
\code{\link[=h2o_explainer]{h2o_explainer()}}
}
\concept{Interpretability}
