% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{date_cuts}
\alias{date_cuts}
\title{Convert Date into Year + Cut}
\usage{
date_cuts(date = Sys.Date(), type = "Q")
}
\arguments{
\item{date}{Date. Date we wish to transform}

\item{type}{Character. Any of the following: B (2 months),
Q (3 months), T (4 months), H (6 months)}
}
\value{
Vector with date cut for each \code{date}
}
\description{
This function returns categorical values for any date(s) using year
cuts such as bimonths, quarters, terms, and halves.
}
\examples{
date_cuts(Sys.Date(), type = "Q")
date_cuts(Sys.Date(), type = "H")
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}
}
\concept{Data Wrangling}
