% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{db_file}
\alias{db_file}
\title{Download Dropbox File by File's Name}
\usage{
db_file(filename, token_rds = NA, xlsx = TRUE)
}
\arguments{
\item{filename}{String. File's name}

\item{token_rds}{String. Tokens directory, must be a RDS file.}

\item{xlsx}{Boolean. Is it an Excel file? Will be returned as a list
for each tab.}
}
\description{
This function lets the user download a file from Dropbox, specifying
its name and using a previously created token.
}
