% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{db_upload}
\alias{db_upload}
\title{Upload Local Files to Dropbox}
\usage{
db_upload(
  filename,
  dir,
  delete_file = FALSE,
  token_dir = NA,
  token_name = "token_pers.rds"
)
}
\arguments{
\item{filename}{String. Local file's name to upload.}

\item{dir}{String. Directory you wish to upload the file to.}

\item{delete_file}{Boolean. Delete local file after uploading?}

\item{token_dir}{Character. RDS with token local directory. You may set to
NA if you already set your credentials (see \code{get_creds()})}

\item{token_name}{Character. RDS file name with your token's data.}
}
\value{
\code{TRUE} when successfully uploads file.
}
\description{
This function lets the user upload a local file to Dropbox,
using a previously created token or with interactive window.
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}

Other Credentials:
\code{\link[=db_download]{db_download()}},
\code{\link[=encrypt_file]{encrypt_file()}},
\code{\link[=get_credentials]{get_credentials()}},
\code{\link[=get_tweets]{get_tweets()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=queryDB]{queryDB()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=slackSend]{slackSend()}},
\code{\link[=stocks_file]{stocks_file()}}

Other Dropbox:
\code{\link[=db_download]{db_download()}}
}
\concept{Credentials}
\concept{Dropbox}
\concept{Tools}
