% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{df_tree}
\alias{df_tree}
\title{Recursive Partitioning and Regression Trees}
\usage{
df_tree(df, target, max = 3, min = 20, cp = 0, size = 0.7,
  ohse = TRUE, plot = TRUE, ...)
}
\arguments{
\item{df}{Data frame}

\item{target}{Variable}

\item{max}{Integer. Maximal depth of the tree}

\item{min}{Integer. The minimum number of observations that must 
exist in a node in order for a split to be attempted}

\item{cp}{Numeric. Complexity parameter}

\item{size}{Numeric. Textsize of plot}

\item{ohse}{Boolean. Auto generate One Hot Smart Encoding?}

\item{plot}{Boolean. Return a plot? If not, rpart object}

\item{...}{rpart.plot custom parameters}
}
\description{
Fit and plot a rpart model for exploratory purposes using rpart
and rpart.plot libraries. Idea from explore library.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{df_str}}, \code{\link{distr}},
  \code{\link{freqs}}, \code{\link{gain_lift}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{textCloud}}, \code{\link{textTokenizer}},
  \code{\link{trendsRelated}}
}
\concept{Exploratory}
