% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dfl}
\alias{dfl}
\title{Dataset: Random Transactional Data}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(dfl)
}
\description{
Data randomly generated for examples.
}
\examples{
data(dfl)
head(dfl)
}
\seealso{
Other Dataset: 
\code{\link{dfr}},
\code{\link{dft}}
}
\concept{Dataset}
\keyword{datasets}
