% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{dist2d}
\alias{dist2d}
\title{Distance from specific point to line}
\usage{
dist2d(x, a = c(0, 0), b = c(1, 1))
}
\arguments{
\item{x}{Vector. Coordinates of the point from which we want to
measure the distance}

\item{a}{Vector. Coordinates of 1st point over the line}

\item{b}{Vector. Coordinates of 2st point over the line}
}
\value{
Numeric value result
}
\description{
This function lets the user calculate the mathematical linear distance
Between a specific point and a line (given geometrical 3 points)
}
\examples{
dist2d(x = c(5, 2))
dist2d(x = c(5, 2), a = c(0, 0), b = c(0, 1))
dist2d(x = c(5, 2), a = c(0, 0), b = c(1, 0))
}
\seealso{
Other Calculus:
\code{\link[=corr]{corr()}},
\code{\link[=corr_cross]{corr_cross()}},
\code{\link[=model_metrics]{model_metrics()}},
\code{\link[=quants]{quants()}}
}
\concept{Calculus}
