% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{rmse}
\alias{rmse}
\alias{mae}
\alias{mse}
\alias{mape}
\alias{rsq}
\alias{rsqa}
\title{Root Mean Squared Error (RMSE)}
\usage{
rmse(tag, score)

mae(tag, score)

mse(tag, score)

mape(tag, score)

rsq(tag, score)

rsqa(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate Root Mean Squared Error

This function lets the user calculate Mean Absolute Error

This function lets the user calculate Mean Squared Error

This function lets the user calculate Mean Squared Error

This function lets the user calculate r squared

This function lets the user calculate adjusted r squared
}
\seealso{
Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{errors}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}
}
\concept{Calculus}
