% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{etf_sector_plot}
\alias{etf_sector_plot}
\title{Portfolio's Sector Distribution (ETFs)}
\usage{
etf_sector_plot(portfolio_perf, save = FALSE)
}
\arguments{
\item{portfolio_perf}{Dataframe. Output of the portfolio_performance function}

\item{save}{Boolean. Export plot as an image?}
}
\description{
This function lets the user plot his portfolio's distribution, 
specifically ETF's sectors
}
\seealso{
Other Investment: \code{\link{etf_sector}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
