% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{export_results}
\alias{export_results}
\title{Export h2o_automl's Results}
\usage{
export_results(
  results,
  thresh = 10,
  which = c("txt", "csv", "rds", "binary", "mojo", "plots", "dev", "production"),
  note = NA,
  subdir = NA,
  save = TRUE,
  seed = 0
)
}
\arguments{
\item{results}{\code{h2o_automl} or \code{h2o} model}

\item{thresh}{Integer. Threshold for selecting binary or regression
models: this number is the threshold of unique values we should
have in 'tag' (more than: regression; less than: classification)}

\item{which}{Character vector. Select which file format to export:
Possible values: txt, csv, rds, binary, mojo, plots. You might also
use dev (txt, csv, rds) or production (binary, mojo) or simply don't use
parameter to export everything}

\item{note}{Character. Add a note to the txt file. Useful when lots of
models are trained and saved to remember which one is which one}

\item{subdir}{Character. In which directory do you wish to save
the results?}

\item{save}{Boolean. Do you wish to save/export results?}

\item{seed}{Numeric. For reproducible results and random splits.}
}
\value{
No return value, called for side effects.
}
\description{
Export RDS, TXT, POJO, MOJO and all results from \code{h2o_automl()}.
}
\seealso{
Other Machine Learning:
\code{\link[=ROC]{ROC()}},
\code{\link[=conf_mat]{conf_mat()}},
\code{\link[=gain_lift]{gain_lift()}},
\code{\link[=h2o_automl]{h2o_automl()}},
\code{\link[=h2o_predict_MOJO]{h2o_predict_MOJO()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=impute]{impute()}},
\code{\link[=iter_seeds]{iter_seeds()}},
\code{\link[=lasso_vars]{lasso_vars()}},
\code{\link[=model_metrics]{model_metrics()}},
\code{\link[=model_preprocess]{model_preprocess()}},
\code{\link[=msplit]{msplit()}}

Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Machine Learning}
\concept{Tools}
