% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_post}
\alias{fb_post}
\title{Get Facebook's Post Comments (API Graph)}
\usage{
fb_post(token, post_id, limit = 5000)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{post_id}{Character vector. Post id(s)}

\item{limit}{Integer. Query limit}
}
\description{
Connect to an API Graph's token and get posts comments given the
post(s) id.
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
