% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_process}
\alias{fb_process}
\title{Paginate and Process Facebook's API Results}
\usage{
fb_process(input, paginate = TRUE, sleep = 0, quiet = FALSE, ...)
}
\arguments{
\item{input}{GET's output object (response) or link (character).}

\item{paginate}{Boolean or integer. Run through all paginations? If set
to \code{FALSE}, only the first one will be processed. If set to any other
integer value, will process the first N paginations.}

\item{sleep}{Numeric value. How much should each loop wait until until running
the next pagination query?}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}

\item{...}{Additional parameters.}
}
\value{
data.frame with un-nested processed results or NULL if no results found.
}
\description{
Process and paginate raw results from Facebook's API, result of
querying the API with \code{httr::GET} or by passing an API link.
}
\seealso{
Other API:
\code{\link[=bring_api]{bring_api()}},
\code{\link[=fb_accounts]{fb_accounts()}},
\code{\link[=fb_ads]{fb_ads()}},
\code{\link[=fb_creatives]{fb_creatives()}},
\code{\link[=fb_insights]{fb_insights()}},
\code{\link[=fb_report_check]{fb_report_check()}},
\code{\link[=fb_rf]{fb_rf()}},
\code{\link[=fb_token]{fb_token()}},
\code{\link[=gemini_ask]{gemini_ask()}},
\code{\link[=gpt_ask]{gpt_ask()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=slackSend]{slackSend()}}

Other Meta:
\code{\link[=fb_accounts]{fb_accounts()}},
\code{\link[=fb_ads]{fb_ads()}},
\code{\link[=fb_creatives]{fb_creatives()}},
\code{\link[=fb_insights]{fb_insights()}},
\code{\link[=fb_report_check]{fb_report_check()}},
\code{\link[=fb_rf]{fb_rf()}},
\code{\link[=fb_token]{fb_token()}}
}
\concept{API}
\concept{Meta}
