% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_token}
\alias{fb_token}
\title{Facebook's Long-Life User API Token}
\usage{
fb_token(app_id, app_secret, token, api_version = NULL)
}
\arguments{
\item{app_id, app_secret}{Character. Application ID and Secret.}

\item{token}{Character. User token, created with
\href{https://developers.facebook.com/tools/explorer}{API Graph}
or with this same \code{fb_token()}'s token.}

\item{api_version}{Character. Facebook API version.}
}
\value{
Character. String with token requested. If successful, it'll contain
an attribute called "expiration" with date and time of expiration.
}
\description{
Using a 1-hour generic user token you can generate a 60 day token.
You will need to have an App ID and App secret, and a valid token.
Generate a new valid User Token with the
\href{https://developers.facebook.com/tools/explorer}{API Graph}.
}
\details{
More info: \href{https://developers.facebook.com/documentation/facebook-login/guides/access-tokens/get-long-lived/}{Long-Lived Access Tokens}
}
\seealso{
Other API:
\code{\link[=bring_api]{bring_api()}},
\code{\link[=fb_accounts]{fb_accounts()}},
\code{\link[=fb_ads]{fb_ads()}},
\code{\link[=fb_creatives]{fb_creatives()}},
\code{\link[=fb_insights]{fb_insights()}},
\code{\link[=fb_process]{fb_process()}},
\code{\link[=fb_report_check]{fb_report_check()}},
\code{\link[=fb_rf]{fb_rf()}},
\code{\link[=gemini_ask]{gemini_ask()}},
\code{\link[=gpt_ask]{gpt_ask()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=slackSend]{slackSend()}}

Other Meta:
\code{\link[=fb_accounts]{fb_accounts()}},
\code{\link[=fb_ads]{fb_ads()}},
\code{\link[=fb_creatives]{fb_creatives()}},
\code{\link[=fb_insights]{fb_insights()}},
\code{\link[=fb_process]{fb_process()}},
\code{\link[=fb_report_check]{fb_report_check()}},
\code{\link[=fb_rf]{fb_rf()}}
}
\concept{API}
\concept{Meta}
