% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{file_type}
\alias{file_type}
\title{Get file extensions without file names}
\usage{
file_type(filepath)
}
\arguments{
\item{filepath}{Character vector. File path(s) to get file extensions or types.}
}
\description{
Get file extensions without file names
}
\examples{
file_type("file.aux")
file_type("temp/file.R")
file_type("/temp/temp3/music.mp3")
}
