% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{files_functions}
\alias{files_functions}
\title{List all functions used in R script files by package}
\usage{
files_functions(filename, abc = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{Character. Path to an R file (or directory)
containing R code files.}

\item{abc}{Boolean. List functions alphabetically.
If FALSE, will list in order of frequency.}

\item{quiet}{Boolean. Keep quiet? If not, print messages and
\code{statusbar}.}
}
\value{
data.frame. Each row is a function and columns stating
number of appearances, percentage, packages, and files searched.
}
\description{
Parses all functions called by an R script and then lists
them by package. Wrapper for 'getParseData'. May be of great
use for those developing a package to help see what
namespace 'importsFrom' calls will be required.
}
\examples{
\dontrun{
# Choose an R script file with functions
rfile <- file.choose()
files_functions(rfile)
}
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Tools}
