% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten lists into data.frame}
\usage{
flatten_list(x)
}
\arguments{
\item{x}{List}
}
\description{
Flatten a list with possible multiple lists within into a single
data.frame with valid column names.
}
