% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{format_text}
\alias{format_text}
\title{Format a string text as markdown/HTML}
\usage{
format_text(text, color = "black", size = 20, bold = FALSE)
}
\arguments{
\item{text}{Character}

\item{color}{Character}

\item{size}{Numeric}

\item{bold}{Boolean}
}
\description{
Format a string text as markdown/HTML
}
\examples{
\dontrun{
col1 <- "grey"
col2 <- "orange"
pt <- data.frame(
  label = paste0(
    format_text(123, color = col2, size = 120, bold = TRUE), "<br/>",
    format_text("of children had a", col1),"<br/>",
    format_text("traditional stay-at-home mom", color = col2, bold = TRUE), "<br/>",
    format_text(paste0("in 2012, compared to ", 321," in 1970"), color = col1)))
  ggplot(pt, aes(x = 0, y = 0))  +
  ggtext::geom_richtext(
    aes(label = label), 
    hjust = 0, 
    label.color = NA, 
    lineheight = 1.5) +
  xlim(0, 0.01) +
  theme_void()
}
}
