% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{formatNum}
\alias{formatNum}
\title{Nicely Format Numerical Values}
\usage{
formatNum(x, decimals = 2, type = 1, scientific = FALSE)
}
\arguments{
\item{x}{Numerical Vector}

\item{decimals}{Integer. Amount of decimals to display}

\item{type}{Integer. 1 for International standards. 2 for 
American Standards.}

\item{scientific}{Boolean. Scientific notation}
}
\description{
This function lets the user format numerical values nicely
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{dateformat}}, \code{\link{left}},
  \code{\link{missingness}}, \code{\link{normalize}},
  \code{\link{numericalonly}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{removenacols}}, \code{\link{removenarows}},
  \code{\link{replaceall}}, \code{\link{right}},
  \code{\link{vector2text}}, \code{\link{year_month}},
  \code{\link{year_week}}
}
\concept{Data Wrangling}
