% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{formatTime}
\alias{formatTime}
\title{Auto Detect Time-Date Format}
\usage{
formatTime(vector)
}
\arguments{
\item{vector}{Vector containing dates or timestamps}
}
\description{
This function tries to detect which format is your date value and returns
a valid time class or same values if non-found.
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{cleanText}}, \code{\link{date_feats}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{holidays}}, \code{\link{impute}},
  \code{\link{left}}, \code{\link{normalize}},
  \code{\link{numericalonly}}, \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
