% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_facecounter}
\alias{fr_facecounter}
\title{Facial Recognition: Face Counter}
\usage{
fr_facecounter(img)
}
\arguments{
\item{img}{Character. Image file}
}
\description{
This function lets the user detect and count faces in an image.
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
