% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_getresult}
\alias{fr_getresult}
\title{Facial Recognition: Get REsults}
\usage{
fr_getresult(imgUrl, print = T)
}
\arguments{
\item{imgUrl}{Character. Image's URL}

\item{print}{Boolean. Print results}
}
\description{
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
