% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_live}
\alias{fr_live}
\title{Facial Recognition: Live Cam}
\usage{
fr_live()
}
\description{
This function lets the user establish a live recoring with its front laptop cam. 
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
