% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_plotface}
\alias{fr_plotface}
\title{Facial Recognition: Plot Face}
\usage{
fr_plotface(img_url, df)
}
\arguments{
\item{img_url}{Character. Image's URL}

\item{df}{Dataframe}
}
\description{
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
