% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_sendimage}
\alias{fr_sendimage}
\title{Facial Recognition: Send Image}
\usage{
fr_sendimage(type, body, endpoint, key)
}
\arguments{
\item{type}{Integer}

\item{body}{Character. Body}

\item{endpoint}{Character. Endpoint URL}

\item{key}{Character. Key}
}
\description{
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
