% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facial_recognition.R
\name{fr_takeselfie}
\alias{fr_takeselfie}
\title{Facial Recognition: Take a Selfie}
\usage{
fr_takeselfie(save = FALSE, print = FALSE)
}
\arguments{
\item{save}{Boolean. Save output as an image}

\item{print}{Boolean. Print output}
}
\description{
This function lets the user take a selfie with its front laptop cam. 
This tools doesn't work well with RStudio sometimes. Try running it in your Terminal.
}
