% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{freqs_df}
\alias{freqs_df}
\title{Plot for All Frequencies on Dataframe}
\usage{
freqs_df(
  df,
  max = 0.9,
  min = 0,
  novar = TRUE,
  plot = FALSE,
  top = 30,
  quiet = FALSE,
  save = FALSE,
  subdir = NA
)
}
\arguments{
\item{df}{Data.frame}

\item{max}{Numeric. Top variance threshold. Range: (0-1].
These variables will be excluded}

\item{min}{Numeric. Minimum variance threshold. Range: [0-1).
These values will be grouped into a high frequency (HF) value}

\item{novar}{Boolean. Remove no variance columns?}

\item{plot}{Boolean. Do you want to see a plot? Three variables tops}

\item{top}{Integer. Plot most relevant (less categories) variables}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to
save the plot to?}
}
\value{
Plot when \code{plot=TRUE} and data.frame with grouped frequency results
when \code{plot=FALSE}.
}
\description{
This function lets the user analize data by visualizing the
frequency of each value of each column from a whole data frame.
}
\examples{
data(dft) # Titanic dataset
freqs_df(dft)
freqs_df(dft, plot = TRUE)
}
\seealso{
Other Frequency:
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}}

Other Exploratory:
\code{\link[=corr_var]{corr_var()}},
\code{\link[=crosstab]{crosstab()}},
\code{\link[=df_str]{df_str()}},
\code{\link[=distr]{distr()}},
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}},
\code{\link[=lasso_vars]{lasso_vars()}},
\code{\link[=missingness]{missingness()}},
\code{\link[=plot_cats]{plot_cats()}},
\code{\link[=plot_df]{plot_df()}},
\code{\link[=plot_nums]{plot_nums()}},
\code{\link[=tree_var]{tree_var()}}

Other Visualization:
\code{\link[=distr]{distr()}},
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}},
\code{\link[=noPlot]{noPlot()}},
\code{\link[=plot_chord]{plot_chord()}},
\code{\link[=plot_survey]{plot_survey()}},
\code{\link[=plot_timeline]{plot_timeline()}},
\code{\link[=tree_var]{tree_var()}}
}
\concept{Exploratory}
\concept{Frequency}
\concept{Visualization}
