% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{geoAddress}
\alias{geoAddress}
\title{Get Google's Geodata given the Addresses}
\usage{
geoAddress(
  address,
  country = "Argentina",
  index = NA,
  creds = NA,
  which = "api_01"
)
}
\arguments{
\item{address}{Character Vector. Addresses you wish to query}

\item{country}{Character. Default Country if known}

\item{index}{Character Vector. If you wish to keep an id on each 
address, set this values}

\item{creds}{Character. Credential's user (see \code{get_creds()})}

\item{which}{Character. Which of the APIs should be used?}
}
\description{
This function lets the user obtain Google's Geodata on given addresses
}
\seealso{
Other Geographical: 
\code{\link{geoGrid}()},
\code{\link{geoMap}()},
\code{\link{geoStratum}()}

Other Credentials: 
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{f1_contacts}()},
\code{\link{get_credentials}()},
\code{\link{get_tweets}()},
\code{\link{mailSend}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()},
\code{\link{stocks_file}()},
\code{\link{stocks_report}()}

Other Google: 
\code{\link{filesGD}()},
\code{\link{queryGA}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()},
\code{\link{trendsTime}()},
\code{\link{writeGS}()}
}
\concept{Credentials}
\concept{Geographical}
\concept{Google}
