% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{geodataAddress}
\alias{geodataAddress}
\title{Get Google's Geodata given the Addresses}
\usage{
geodataAddress(address, country = "Colombia", index = NA, creds = NA)
}
\arguments{
\item{address}{Character Vector. Addresses you wish to query}

\item{country}{Character. Default Country if known}

\item{index}{Character Vector. If you wish to keep an id on each address, set this values}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user obtain Google's Geodata on given addresses
}
