% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{geoStratum}
\alias{geoStratum}
\title{Get Colombia's Stratum given the Coordinates}
\usage{
geoStratum(lon, lat, label = NA)
}
\arguments{
\item{lon}{Numeric Vector. Longitudes}

\item{lat}{Numeric Vector. Latitudes}

\item{label}{Character Vector. If you wish to keep an id on each address, set this values}
}
\description{
This function lets the user obtain Colombia's stratum given the coordinates of an address
}
