% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{get_stocks}
\alias{get_stocks}
\title{Get Personal Portfolio's Data}
\usage{
get_stocks(filename = NA,
  token_dir = "~/Dropbox (Personal)/Documentos/Docs/Data", auto = TRUE)
}
\arguments{
\item{filename}{Characeter. Import a local Excel file}

\item{token_dir}{Character. Where is my personal token for Dropbox connection?}

\item{auto}{Boolean. Automatically user my local personal file?}
}
\description{
This function lets me download my personal Excel with my Portfolio data
}
