% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{get_stocks_hist}
\alias{get_stocks_hist}
\title{Download Stocks Historical Data}
\usage{
get_stocks_hist(symbols = NA, from = NA, today = TRUE, tax = 30,
  verbose = TRUE)
}
\arguments{
\item{symbols}{Character Vector. List of symbols to download historical data. 
Example: c('VTI','TSLA')}

\item{from}{Date. Since when do you wish to download historical data}

\item{today}{Boolean. Do you wish to additionaly download today's quote?}

\item{tax}{Numeric. Percentage for dividends real return. Range from 0 to 99}

\item{verbose}{Boolean. Print results and progress while downloading?}
}
\description{
This function lets the user download stocks historical data
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
