% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tweets}
\alias{get_tweets}
\title{Get Tweets}
\usage{
get_tweets(q, n = 10000, creds = NA)
}
\arguments{
\item{q}{Query. Check for ?rtweet::search_tweets()}

\item{n}{Integer. Total of tweets to return}

\item{creds}{Character. Credential's user (see \code{get_creds()})}
}
\value{
data.frame with API response results.
}
\description{
This function downloads tweets with personal credentials
}
\seealso{
Other Credentials:
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=encrypt_file]{encrypt_file()}},
\code{\link[=get_credentials]{get_credentials()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=queryDB]{queryDB()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=slackSend]{slackSend()}},
\code{\link[=stocks_file]{stocks_file()}}
}
\concept{Credentials}
\concept{Twitter}
