% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{gg_pie}
\alias{gg_pie}
\title{Density plot for discrete and continuous values}
\usage{
gg_pie(df, var, table = FALSE, ...)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Variable to group, count and plot}

\item{table}{Boolean. Print results as table?}

\item{...}{Further parameters passed to freqs()}
}
\description{
This function plots discrete and continuous values results
}
\examples{
\dontrun{
data(dft) # Titanic dataset
gg_pie(dft, Survived)
gg_pie(dft, Pclass, table = TRUE)
gg_pie(dft, SibSp, table = TRUE, abc = TRUE)
}
}
