% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS}
\alias{readGS}
\alias{readGS4}
\alias{writeGS}
\alias{writeGS4}
\title{Google Sheets Reading and Writing (API v4)}
\usage{
readGS(
  title,
  sheet = "Hoja 1",
  range = NULL,
  drop_nas = TRUE,
  json = NULL,
  email = NULL,
  api_key = NULL,
  server = FALSE,
  ...
)

writeGS(
  data,
  title,
  sheet = "Hoja 1",
  range = "A1",
  reformat = FALSE,
  append = FALSE,
  json = NULL,
  email = NULL,
  api_key = NULL,
  server = FALSE,
  ...
)
}
\arguments{
\item{title}{Character. Title of Google Drive file. Uses regular
expressions so you may fetch with patterns instead of names.}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{drop_nas}{Boolean. Remove columns and rows that contain only NAs?}

\item{json}{Character. JSON filename with service auth}

\item{email, api_key}{Character. If you have multiple pre-authorized
accounts in your machine, you may non-interactively select
which one you wish to use by email and/or api_key.}

\item{server}{Boolean. Force interacting auth process?}

\item{...}{Additional parameters passed to \code{read_sheet()}.}

\item{data}{Object (value, vector, data.frame, list).}

\item{reformat}{Boolean. Reformat the affected cells?}

\item{append}{Boolean.}
}
\value{
For reading, data.frame with the results of your Google Sheets file
based on its \code{title}, specificially the \code{sheet} and \code{range}
requested. For writing, no return value.
}
\description{
Read and write data from Google Sheets knowing the file's title.
You may use a single value from a cell or a data.frame from a cell range.
}
\seealso{
Other Scrapper:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=gtrends_related]{gtrends_related()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=mp3_get]{mp3_get()}},
\code{\link[=splot_summary]{splot_summary()}},
\code{\link[=stocks_quote]{stocks_quote()}}

Other Google:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=gtrends_related]{gtrends_related()}},
\code{\link[=queryGA]{queryGA()}}
}
\concept{Google}
\concept{Scrapper}
