% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_trends.R
\name{gtrends_related}
\alias{gtrends_related}
\alias{trendsRelated}
\alias{gtrends_time}
\alias{trendsTime}
\title{Google Trends: Related and Time Plots}
\usage{
gtrends_related(gtrend, top = NA, title = NA, note = NA, exclude = NULL)

gtrends_time(gtrend, title = NA)
}
\arguments{
\item{gtrend}{List. Result from \code{gtrendsR::gtrends(keyword, geo, time)}}

\item{top}{Integer. Filter top n results only.}

\item{title}{Character. Custom title for the plot.}

\item{note}{Character. Add a note to the plot if needed.}

\item{exclude}{Character vector. Which observations do you wish to exclude?}
}
\value{
plot for Google Trend's results input \code{gtrend}.

Plot for Google Trend's results input \code{gtrend}.
}
\description{
This function creates a plot with Google Trend's related topics
and queries, and let the user compare different keywords.

This function creates a plot with google trend's data on timelines
and let the user compare different keywords.
}
\seealso{
Other Scrapper:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=mp3_get]{mp3_get()}},
\code{\link[=readGS]{readGS()}},
\code{\link[=splot_summary]{splot_summary()}},
\code{\link[=stocks_quote]{stocks_quote()}}

Other Google:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=readGS]{readGS()}}
}
\concept{Google}
\concept{Scrapper}
