% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{grepl_anywhere}
\alias{grepl_anywhere}
\title{Pattern Matching and Replacement Anywhere with Blanks}
\usage{
grepl_anywhere(vector, pattern, blank = "_")
}
\arguments{
\item{vector}{Character vector}

\item{pattern}{Character. Character string containing a 
semi-regular expression which uses the following logic:
"a_b" means any character that contains "a" followed by 
something followed by "b", anywhere in the string.}

\item{blank}{Character. String to use between regular expressions.}
}
\description{
Match and replace pattern with blanks within each element of a 
character vector, allowing counted characters between patterns.
}
\examples{
x <- c("aaaa", "bbbb", "abab", "aabb", "a", "ab")
grepl_anywhere(x, "ab")
grepl_anywhere(x, "_ab")
grepl_anywhere(x, "a_a")
grepl_anywhere(x, "c")
}
