% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_predict_API}
\alias{h2o_predict_API}
\title{H2O Predict using API Service}
\usage{
h2o_predict_API(df, api)
}
\arguments{
\item{df}{Dataframe/Vector. Data to insert into the model}

\item{api}{Character. API's URL}
}
\description{
This function lets the user get the score from an API service
}
