% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_predict_binary}
\alias{h2o_predict_binary}
\title{H2O Predict using Binary file}
\usage{
h2o_predict_binary(df, model_path, sample = NA)
}
\arguments{
\item{df}{Dataframe. Data to insert into the model}

\item{model_path}{Character. Relative model_path directory}

\item{sample}{Integer. How many rows should the function predict?}
}
\description{
This function lets the user predict using the h2o binary file.
Note that it works with the files generated when using the 
function lares::export_results(). Recommendation: use the 
h2o_predict_MOJO() function when possible - it let's you change
h2o's version without problem.
}
