% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_predict_model}
\alias{h2o_predict_model}
\title{H2O Predict using H2O Model Object}
\usage{
h2o_predict_model(df, model)
}
\arguments{
\item{df}{Dataframe/Vector. Data to insert into the model}

\item{model}{H2o Object. Model}
}
\description{
This function lets the user get scores from a H2O Model Object
}
