% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{h2o_update}
\alias{h2o_update}
\title{Install latest version of H2O}
\usage{
h2o_update(run = TRUE)
}
\arguments{
\item{run}{Boolean. Do you want to run and start an H2O cluster?}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
