% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{holidays}
\alias{holidays}
\title{Holidays in your Country}
\usage{
holidays(
  countries = "Venezuela",
  years = year(Sys.Date()),
  quiet = FALSE,
  include_regions = FALSE
)
}
\arguments{
\item{countries}{Character or vector. For which country(ies) should the
holidays be imported? Accepts country names or ISO 3166-1 alpha-2 codes.}

\item{years}{Character or vector. For which year(s) do you wish to import
holiday dates?}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}

\item{include_regions}{Boolean. Default FALSE. If TRUE, for countries with
internal subdivisions, it will provide details on which sub-state the found
holidays apply.}
}
\value{
\code{data.frame} with holidays data for given \code{countries} and \code{years}.
}
\description{
This function lets the user automatically retrieve public holiday dates
for any country supported by the Nager.Date API. Accepts country names
(e.g., "Portugal") or ISO 3166-1 alpha-2 codes (e.g., "PT").
Thanks to \href{https://nagerholidays.com/}{Nager.Date}!
}
\examples{
\donttest{
holidays(countries = "Argentina")
year <- as.integer(format(Sys.Date(), format = "\%Y"))
holidays(countries = c("Spain", "Venezuela"), years = year)
holidays(countries = "Germany", include_regions = TRUE)
holidays(countries = "PT") # Also accepts ISO country codes
}
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}

Other Feature Engineering:
\code{\link[=date_feats]{date_feats()}},
\code{\link[=ohse]{ohse()}}

Other Scrapper:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=gtrends_related]{gtrends_related()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=mp3_get]{mp3_get()}},
\code{\link[=readGS]{readGS()}},
\code{\link[=splot_summary]{splot_summary()}},
\code{\link[=stocks_quote]{stocks_quote()}}

Other One Hot Encoding:
\code{\link[=date_feats]{date_feats()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}}
}
\concept{Data Wrangling}
\concept{Feature Engineering}
\concept{One Hot Encoding}
\concept{Scrapper}
