% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot.R
\name{hs_contacts}
\alias{hs_contacts}
\title{Hubspot Contacts}
\usage{
hs_contacts(limit = 1000, creds = NA)
}
\arguments{
\item{limit}{Integer. Limit how many contacts you wish to bring}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function brings ALL contacts with Hubspot's HAPI key
}
