% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot.R
\name{hs_deals}
\alias{hs_deals}
\title{Hubspot Deals}
\usage{
hs_deals(limit = 10000, creds = NA)
}
\arguments{
\item{limit}{Integer. Limit how many contacts you wish to bring}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function brings ALL deals with Hubspot's HAPI key
}
