% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robyn.R
\name{hyps_builder}
\alias{hyps_builder}
\title{Robyn: Generate default hyperparameters}
\usage{
hyps_builder(
  paid_media_spends,
  media_type = "default",
  lagged = FALSE,
  adstock = "geometric",
  date_type = "weekly"
)
}
\arguments{
\item{paid_media_spends}{Character vector. Paid media variables names.}

\item{media_type}{Character vector. Must be ength 1 or same as
\code{paid_media_spends}. Pick, for every \code{paid_media_spends} value,
what type of media it is: "online" or "offline".}

\item{lagged}{Boolean vector. Must be ength 1 or same as
\code{paid_media_spends}. Pick, for every \code{paid_media_spends} value,
if you wish to have a lagged effect. Only valid for Weibull adstock.}

\item{adstock}{Character. Pick one of: "geometric" or "weibull".}

\item{date_type}{Character. Pick one of: "daily", "weekly, or "monthly".
Only valid to transform thetas when using geometric adstock.}
}
\value{
list with default hyperparameters ranges.
}
\description{
Generate a list with hyperparameter defaults values, ready to be
passed to \code{Robyn::robyn_inputs()}.
}
\examples{
hyps_builder(
  paid_media_spends = c(
    "branded_search_spend",
    "nonbranded_search_spend",
    "facebook_spend",
    "print_spend",
    "ooh_spend",
    "tv_spend",
    "radio_spend"),
  media_type = c(
    "online", "online", "online",
    "offline", "offline", "offline", "offline"),
  adstock = "geometric",
  date_type = "weekly")
}
\concept{Robyn}
