% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares.R
\name{install_recommended}
\alias{install_recommended}
\title{Install/Update Additional Recommended Libraries}
\usage{
install_recommended(progress = TRUE)
}
\arguments{
\item{progress}{Boolean. Show status bar?}
}
\description{
All needed libraries to use (most) lares are already a dependency. 
There are some functions that many people won't event know exist 
that will require other additional libraries. Also, this may be 
used as a Docker way of installing useful libraries on an new instance.
}
