% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{isURL}
\alias{isURL}
\title{Regular Expression for URLs}
\usage{
isURL(x, ...)
}
\arguments{
\item{x}{Character vector}

\item{...}{Additional parameters passed to \code{grepl}}
}
\description{
Regular Expression for URLs
}
\examples{
isURL("test")
isURL("google.com")
isURL("http://google.com")
isURL("test@mail.com")
isURL("https://stackoverflow.com/help")
}
