% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares.R
\docType{package}
\name{lares}
\alias{lares-package}
\alias{lares}
\title{Lean Analytics and Robust Exploration Sidekick}
\description{
R library for better/faster analytics, visualization, data mining, and machine learning tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laresbernardo/lares}
  \item \url{https://laresbernardo.github.io/lares/}
  \item Report bugs at \url{https://github.com/laresbernardo/lares/issues}
}

}
\author{
Bernardo Lares (lares@bervos.org)
}
