% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{lares_logo}
\alias{lares_logo}
\title{Print lares R library Logo}
\usage{
lares_logo(version = TRUE)
}
\arguments{
\item{version}{Boolean. Include R and lares version?}
}
\description{
Used "ASCII Art Generator" from manytools.org to convert logo to ASCII.
\href{https://manytools.org/hacker-tools/convert-images-to-ascii-art/}{Visit}.
}
\examples{
lares_logo()
}
