% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{left}
\alias{left}
\title{Left: First n characters}
\usage{
left(string, n = 1)
}
\arguments{
\item{string}{String or Vector}

\item{n}{Integer. How many characters from left to right?}
}
\description{
This function lets the user extract the first n characters of a
string or vector of strings.
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{cleanText}}, \code{\link{date_feats}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{holidays}}, \code{\link{impute}},
  \code{\link{normalize}}, \code{\link{numericalonly}},
  \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
