% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_profile}
\alias{li_profile}
\title{Get My Personal LinkedIn Data}
\usage{
li_profile(token = NA)
}
\arguments{
\item{token}{Object. OAuth Authentication: li_auth()'s output}
}
\description{
This function brings a list with your personal LinkedIn data
}
\seealso{
Other Social Networks: 
\code{\link{get_tweets}()},
\code{\link{li_auth}()}
}
\concept{Social Networks}
