% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{list_cats}
\alias{list_cats}
\title{List categorical values for data.frame}
\usage{
list_cats(df, ..., abc = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{...}{Variables to segment counters}

\item{abc}{Boolean. Sort alphabetically?}
}
\value{
List. Length same as number of categorical columns, each with a
frequency data.frame using \code{freqs()}.
}
\description{
Make a list with all categorical values and
}
\examples{
data(dft) # Titanic dataset
df <- dft[, 1:5]
head(df)
list_cats(df)
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Tools}
