% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error (MAPE)}
\usage{
mape(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate Mean Squared Error
}
