% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{matrixwd}
\alias{matrixwd}
\title{MatrixDS Auto-Working Directory for Shiny}
\usage{
matrixwd(dir)
}
\arguments{
\item{dir}{Character. Project's folder}
}
\description{
This function automatically detects which of the Shiny paths 
should be used when running a Shiny App on MatrixDS (internal
RStudio session or external production version). It also
works for a local file so the script is the same.
}
