% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_roc}
\alias{mplot_roc}
\title{ROC Curve Plot}
\usage{
mplot_roc(
  tag,
  score,
  multis = NA,
  sample = 1000,
  model_name = NA,
  subtitle = NA,
  interval = 0.2,
  squared = TRUE,
  plotly = FALSE,
  save = FALSE,
  subdir = NA,
  file_name = "viz_roc.png"
)
}
\arguments{
\item{tag}{Vector. Real known label.}

\item{score}{Vector. Predicted value or model's result.}

\item{multis}{Data.frame. Containing columns with each category probability
or score (only used when more than 2 categories coexist).}

\item{sample}{Integer. Number of samples to use for rendering plot.}

\item{model_name}{Character. Model's name}

\item{subtitle}{Character. Subtitle to show in plot}

\item{interval}{Numeric. Interval for breaks in plot}

\item{squared}{Boolean. Keep proportions?}

\item{plotly}{Boolean. Use plotly for plot's output for an interactive plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\value{
Plot with ROC curve and AUC performance results.
}
\description{
This function plots ROC Curves with AUC values with 95\% confidence
range. It also works for multi-categorical models.
}
\examples{
Sys.unsetenv("LARES_FONT") # Temporal
data(dfr) # Results for AutoML Predictions
lapply(dfr[c(1, 2)], head)

# ROC Curve for Binomial Model
mplot_roc(dfr$class2$tag, dfr$class2$scores,
  model_name = "Titanic Survived Model"
)

# ROC Curves for Multi-Categorical Model
mplot_roc(dfr$class3$tag, dfr$class3$score,
  multis = subset(dfr$class3, select = -c(tag, score)),
  squared = FALSE,
  model_name = "Titanic Class Model"
)
}
\seealso{
Other ML Visualization:
\code{\link[=mplot_conf]{mplot_conf()}},
\code{\link[=mplot_cuts]{mplot_cuts()}},
\code{\link[=mplot_cuts_error]{mplot_cuts_error()}},
\code{\link[=mplot_density]{mplot_density()}},
\code{\link[=mplot_full]{mplot_full()}},
\code{\link[=mplot_gain]{mplot_gain()}},
\code{\link[=mplot_importance]{mplot_importance()}},
\code{\link[=mplot_lineal]{mplot_lineal()}},
\code{\link[=mplot_metrics]{mplot_metrics()}},
\code{\link[=mplot_response]{mplot_response()}},
\code{\link[=mplot_splits]{mplot_splits()}},
\code{\link[=mplot_topcats]{mplot_topcats()}}
}
\concept{ML Visualization}
