% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{msplit}
\alias{msplit}
\title{Split a dataframe for training and testing sets}
\usage{
msplit(df, size = 0.7, seed = 0, print = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{size}{Numeric. Split rate value, between 0 and 1. If set to
1, the train and test set will be the same.}

\item{seed}{Integer. Seed for random split}

\item{print}{Boolean. Print summary results?}
}
\value{
List with both datasets, summary, and split rate.
}
\description{
This function splits automatically a dataframe into train and
test datasets. You can define a seed to get the same results
every time, but has a default value. You can prevent it from
printing the split counter result.
}
\examples{
data(dft) # Titanic dataset
splits <- msplit(dft, size = 0.7, seed = 123)
names(splits)
}
\seealso{
Other Machine Learning:
\code{\link[=ROC]{ROC()}},
\code{\link[=conf_mat]{conf_mat()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=gain_lift]{gain_lift()}},
\code{\link[=h2o_automl]{h2o_automl()}},
\code{\link[=h2o_predict_MOJO]{h2o_predict_MOJO()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=impute]{impute()}},
\code{\link[=iter_seeds]{iter_seeds()}},
\code{\link[=lasso_vars]{lasso_vars()}},
\code{\link[=model_metrics]{model_metrics()}},
\code{\link[=model_preprocess]{model_preprocess()}}

Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Machine Learning}
\concept{Tools}
