% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{normalize}
\alias{normalize}
\title{Normalize Vector}
\usage{
normalize(x, range = c(0, 1), ...)
}
\arguments{
\item{x}{Numeric vector. The numbers to be normalized.}

\item{range}{A numeric vector of length 2 specifying the desired range
for normalization. Default is c(0, 1).}

\item{...}{Additional parameters.}
}
\value{
A numeric vector with normalized \code{x} values.
}
\description{
This function normalizes numerical values into a specified range,
defaulting to the 0 to 1 range.
}
\examples{
x <- c(0, 1, 4, 7.5, 10)
normalize(x)
normalize(x, range = c(-1, 1))
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}
}
\concept{Data Wrangling}
