% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{numericalonly}
\alias{numericalonly}
\title{Filter only Numerical Values and}
\usage{
numericalonly(df, dropnacols = TRUE, logs = FALSE, natransform = NA)
}
\arguments{
\item{df}{Data.frame}

\item{dropnacols}{Boolean. Drop columns with only NA values?}

\item{logs}{Boolean. Calculate log(x)+1 for numerical columns?}

\item{natransform}{String. "mean" or 0 to impute NA values. If
set to NA no calculation will run.}
}
\description{
This function lets the user remove all rows that have some or
all values as NAs
}
