% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_detection.R
\name{object_detection}
\alias{object_detection}
\title{Image's Objects Detection using YOLO}
\usage{
object_detection(filename, threshold = 0.2, type = "faster")
}
\arguments{
\item{filename}{Character. Name of the file to analize}

\item{threshold}{Numeric. Range from 0 to 1. Detection threshold}

\item{type}{Character. Select 'faster' or 'better' for YOLO method}
}
\description{
This function brings ALL contacts with Hubspot's HAPI key. To install
the needed library please run devtools::install_github("bnosac/image", 
subdir = "image.darknet", build_vignettes = TRUE)
}
