% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{ohe}
\alias{ohe}
\title{One Hot Encoding (Dummy Variables)}
\usage{
ohe(df, redundant = FALSE, trim = 0, summary = TRUE,
  variance = 0.95, limit = 10, other_label = "OTHER", sep = "_")
}
\arguments{
\item{df}{Dataframe}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. It the
column only has two different values, should we keep both new columns?}

\item{trim}{Integer. Trim names until the nth character}

\item{summary}{Boolean. Print a summary of the operations?}

\item{variance}{Numeric. Drop columns with more than n variance. Range: 0-1}

\item{limit}{Integer. Limit OHE to n columns. Automatically, the
function will select the n most frequent values}

\item{other_label}{Character. With which text do you wish to replace 
the filtered values with?}

\item{sep}{Character. Separator's string}
}
\description{
This function lets the user automatically transform a dataframe with
categorical columns into numerical by one hot encoding technic.
}
