% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{pass}
\alias{pass}
\title{Pass Through a dplyr's Pipeline}
\usage{
pass(df, fun)
}
\arguments{
\item{df}{Dataframe}

\item{fun}{Function. What function do you wish to run? For example:
pass(. \%>\% ncol \%>\% print)}
}
\description{
This function lets the user print, save or do something inside a
pipeline without affecting the output or pipeline.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\concept{Tools}
