% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{plot.h2o_automl}
\alias{plot.h2o_automl}
\title{Plotting function for h2o_automl}
\usage{
\method{plot}{h2o_automl}(x, ...)
}
\arguments{
\item{x}{h2o_automl object.}

\item{...}{Additional arguments}
}
\description{
Plotting function for \code{h2o_automl}.
}
