% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{plot_cats}
\alias{plot_cats}
\title{Plot All Categorical Features (Frequencies)}
\usage{
plot_cats(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Plot. Result of \code{df} categorical features.
}
\description{
This function filters categorical columns and plots the frequency
for each value on every feature.
}
\seealso{
Other Exploratory:
\code{\link[=corr_var]{corr_var()}},
\code{\link[=crosstab]{crosstab()}},
\code{\link[=df_str]{df_str()}},
\code{\link[=distr]{distr()}},
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_df]{freqs_df()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}},
\code{\link[=lasso_vars]{lasso_vars()}},
\code{\link[=missingness]{missingness()}},
\code{\link[=plot_df]{plot_df()}},
\code{\link[=plot_nums]{plot_nums()}},
\code{\link[=tree_var]{tree_var()}}
}
\concept{Exploratory}
