% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{plot_df}
\alias{plot_df}
\title{Plot Summary of Numerical and Categorical Features}
\usage{
plot_df(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Plot. Result of \code{df} categorical and numerical features.
}
\description{
This function plots all columns frequencies and boxplots, for
categorical and numerical respectively.
}
\seealso{
Other Exploratory:
\code{\link[=corr_var]{corr_var()}},
\code{\link[=crosstab]{crosstab()}},
\code{\link[=df_str]{df_str()}},
\code{\link[=distr]{distr()}},
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_df]{freqs_df()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}},
\code{\link[=lasso_vars]{lasso_vars()}},
\code{\link[=missingness]{missingness()}},
\code{\link[=plot_cats]{plot_cats()}},
\code{\link[=plot_nums]{plot_nums()}},
\code{\link[=tree_var]{tree_var()}}
}
\concept{Exploratory}
