% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{plot_survey}
\alias{plot_survey}
\title{Visualize Survey Results}
\usage{
plot_survey(answers, ignore = 1, title = NA, subtitle = NA)
}
\arguments{
\item{answers}{Dataframe. Answers. Each row a different person.
Each column a different answer.}

\item{ignore}{Numeric Vector. Which columns are NOT answers?}

\item{title}{Character. Title for your plot}

\item{subtitle}{Character. Subtitle for your plot.}
}
\value{
ggplot2 object
}
\description{
This function lets the user plot a survey's result.
}
\seealso{
Other Visualization:
\code{\link[=distr]{distr()}},
\code{\link[=freqs]{freqs()}},
\code{\link[=freqs_df]{freqs_df()}},
\code{\link[=freqs_list]{freqs_list()}},
\code{\link[=freqs_plot]{freqs_plot()}},
\code{\link[=noPlot]{noPlot()}},
\code{\link[=plot_chord]{plot_chord()}},
\code{\link[=plot_timeline]{plot_timeline()}},
\code{\link[=tree_var]{tree_var()}}
}
\concept{Visualization}
