% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{plotMap}
\alias{plotMap}
\title{Plot Map or Shapefile}
\usage{
plotMap(map, fix_coords = FALSE)
}
\arguments{
\item{map}{SpatialPolygonsDataFrame or .shp directory}

\item{fix_coords}{Boolean. Transform and fix coordinates system?}
}
\description{
This function returns a simple plot from a SpatialPolygonsDataFrame
imported or a .shp file
}
