% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_daily}
\alias{portfolio_daily}
\title{Portfolio Daily Performance}
\usage{
portfolio_daily(data, dailys, cash_fix = 0)
}
\arguments{
\item{data}{Dataframe. Result from get_stocks()}

\item{dailys}{Dataframe. Result from get_stocks_hist()}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value 
to your cash balance}
}
\description{
This function lets the user calculate daily portfolio performance
}
\seealso{
Other Investment: \code{\link{etf_sector}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
