% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_daily_plot}
\alias{portfolio_daily_plot}
\title{Portfolio Daily Plot}
\usage{
portfolio_daily_plot(stocks_perf, save = FALSE)
}
\arguments{
\item{stocks_perf}{Dataframe. Output of the stocks_performance function}

\item{save}{Boolean. Export plot as an image?}
}
\description{
This function lets the user plot his portfolio daily change
}
