% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_distr_plot}
\alias{portfolio_distr_plot}
\title{Portfolio's Distribution}
\usage{
portfolio_distr_plot(portfolio_perf, daily)
}
\arguments{
\item{portfolio_perf}{Dataframe. Output of the portfolio_performance function}

\item{daily}{Dataframe. Daily data}
}
\description{
This function lets the user plot his portfolio's distribution
}
