% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_performance}
\alias{portfolio_performance}
\title{Portfolio Overall Performance}
\usage{
portfolio_performance(portfolio, daily)
}
\arguments{
\item{portfolio}{Dataframe. Structure: "Symbol", "Stocks", "StockIniValue", "InvPerc", "Type", "Trans", "StartDate"}

\item{daily}{Dataframe. Daily data}
}
\description{
This function lets the user calculate portfolio performance
}
