% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{print_coloured}
\alias{print_coloured}
\title{Print Coloured Messages}
\usage{
print_coloured(txt, colour = c("BLUE", "YELLOW", "GREY"), cat = TRUE)
}
\arguments{
\item{txt}{Character. Text to print.}

\item{colour}{Character. Any of: grey, red, green, yellow, blue, or purple.}

\item{cat}{Boolean. Print with cat? If not, raw string}
}
\value{
Depends on \code{cat}: NULL if TRUE or character string if FALSE.
}
\description{
Print Coloured Messages
}
\examples{
opts <- c("GREY", "RED", "GREEN", "YELLOW", "BLUE", "PURPLE")
for (colour in opts) print_coloured(paste("Colour:", colour, "\n"), colour)
}
