% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{quants}
\alias{quants}
\title{Calculate cuts by quantiles}
\usage{
quants(values, splits = 10, return = "labels", n = 2)
}
\arguments{
\item{values}{Vector. Values to calculate quantile cuts}

\item{splits}{Integer. How many cuts should split the values?}

\item{return}{Character. Return "summary" or "labels"}

\item{n}{Integer. Determines the number of digits used in
formatting the break numbers.}
}
\value{
Factor vector or data.frame. Depending on \code{return} input:
\itemize{
  \item \code{labels} a factor ordered vector with each observation's quantile
  \item \code{summary} a data.frame with information on each quantile cut
}
}
\description{
This function lets the user quickly calculate cuts for quantiles
and discretize numerical values into categorical values.
}
\examples{
data(dft) # Titanic dataset
quants(dft$Age, splits = 5, "summary")
quants(dft$Age, splits = 5, "labels")[1:10]
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}

Other Calculus:
\code{\link[=corr]{corr()}},
\code{\link[=corr_cross]{corr_cross()}},
\code{\link[=dist2d]{dist2d()}},
\code{\link[=model_metrics]{model_metrics()}}
}
\concept{Calculus}
\concept{Data Wrangling}
