% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryDummy}
\alias{queryDummy}
\title{PostgreSQL Queries on Dummy Database (read only)}
\usage{
queryDummy(query, creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user query our Dummy Database
}
