% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryProduc}
\alias{queryProduc}
\title{PostgreSQL Queries on Production Database (read-write)}
\usage{
queryProduc(query, creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user query our Production Database
}
