% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{quiet}
\alias{quiet}
\title{Quiet prints and verbose noise}
\usage{
quiet(fx, quiet = TRUE)
}
\arguments{
\item{fx}{Function to quiet}

\item{quiet}{Quiet outputs? If not, skip quietness.}
}
\value{
Same as \code{fx} but with no messages or prints.
}
\description{
This function silences (verbose) output prints. Thanks to Hadley Wickham
for bringing the idea.
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Tools}
