% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{quo_val}
\alias{quo_val}
\title{Quote object's name (deparse + substitute)}
\usage{
quo_val(x, env = parent.frame())
}
\arguments{
\item{x}{(Unquoted) Object}

\item{env}{Environment where object lives}
}
\value{
Character
}
\description{
Quote object's name (deparse + substitute)
}
