% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbind.R
\name{rbind_full}
\alias{rbind_full}
\title{Smart rbind}
\usage{
rbind_full(..., list, fill = NA, sep = ":")
}
\arguments{
\item{...}{Dataframes. Dataframes to combine}

\item{list}{List. Containing dataframes to combine}

\item{fill}{Value. To use when 'filling' missing columns}

\item{sep}{Character. String used to separate column names when 
pasting them together.}
}
\description{
Efficient rbind of data frames, even if the column names don't match.
Code based on gtools::smartbind
}
