% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS4}
\alias{readGS4}
\title{Google Sheets Reading (API v4)}
\usage{
readGS4(title, sheet = "Hoja 1", range = NULL, creds = NA)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{creds}{Character. JSON filename with service auth}
}
\description{
This function lets the user read any Google Sheet's data
}
\seealso{
Other Connection: \code{\link{get_tweets}},
  \code{\link{queryGA}}, \code{\link{readGS}},
  \code{\link{writeGS}}
}
\concept{Connection}
