% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{removenacols}
\alias{removenacols}
\title{Remove/Drop Columns in which ALL or SOME values are NAs}
\usage{
removenacols(df, all = TRUE)
}
\arguments{
\item{df}{Data.frame}

\item{all}{Boolean. Remove columns which contains ONLY NA values.
If set to FALSE, columns which contains at least one NA will be removed}
}
\description{
This function lets the user remove all columns that have some or
all values as NAs
}
