% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{removenarows}
\alias{removenarows}
\title{Remove/Drop Rows in which ALL or SOME values are NAs}
\usage{
removenarows(df, all = TRUE)
}
\arguments{
\item{df}{Data.frame}

\item{all}{Boolean. Remove rows which contains ONLY NA values.
If set to FALSE, rows which contains at least one NA will be removed}
}
\description{
This function lets the user remove all rows that have some or
all values as NAs
}
