% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{right}
\alias{right}
\title{Right: Last n characters}
\usage{
right(string, n)
}
\arguments{
\item{string}{String or Vector}

\item{n}{Integer. How many characters from right to left?}
}
\description{
This function lets the user extract the last n characters of a
string or vector of strings.
}
