% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{rsq}
\alias{rsq}
\title{Calculate R Squared}
\usage{
rsq(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate r squared
}
