% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scrabble_dictionary}
\alias{scrabble_dictionary}
\title{Scrabble: Dictionaries}
\usage{
scrabble_dictionary(language)
}
\arguments{
\item{language}{Character. Any of "en","es","de","fr"}
}
\description{
Download words from 4 different languages: English, Spanish, 
German, and French. Thanks to "lorenbrichter" for sharing them.
}
\examples{
\dontrun{
# For Spanish words
dictionary <- scrabble_dictionary("es")
}
}
\seealso{
Other Scrabble: 
\code{\link{scrabble_score}()},
\code{\link{scrabble_words}()}
}
\concept{Scrabble}
