% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scrabble_points}
\alias{scrabble_points}
\title{Scrabble: Tiles Points}
\usage{
scrabble_points(language)
}
\arguments{
\item{language}{Character. Any of "en","es".}
}
\value{
data.frame with tiles and scores for each letter.
}
\description{
Dataframe for every letter and points given a language.
}
\examples{
scrabble_points("es")
scrabble_points("en")
# Not yet available
scrabble_points("fr")
}
\seealso{
Other Scrabble: 
\code{\link{scrabble_dictionary}()},
\code{\link{scrabble_score}()},
\code{\link{scrabble_words}()}
}
\concept{Scrabble}
