% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scrabble_score}
\alias{scrabble_score}
\title{Scrabble: Scores}
\usage{
scrabble_score(words, scores)
}
\arguments{
\item{words}{Character vector. Words to score}

\item{scores}{Dataframe. Must contain two columns: "tiles" with every
letter of the alphabet and "scores" for each letter's score.}
}
\description{
Get score for any word or list of words. You may set manually depending
on the rules and languages you are playing with. Check the examples
for Spanish and English values when I played Words With Friends.
}
\examples{
# For Spanish words
es_scores <- data.frame(
  tiles = c(tolower(LETTERS)[1:14],"ñ",tolower(LETTERS)[15:length(LETTERS)]),
  scores = c(1,3,2,2,1,4,3,4,1,8,10,1,3,1,8,1,3,5,1,1,1,2,4,10,10,5,10))
# For English words
en_scores <- data.frame(
  tiles = tolower(LETTERS),
  scores = c(1,4,4,2,1,4,3,3,1,10,5,2,4,2,1,4,10,1,1,1,2,5,4,8,3,10))
words <- c("Bernardo", "whiskey", "R is great")
# Score values for each language's rules
scrabble_score(words, es_scores)
scrabble_score(words, en_scores)
}
\seealso{
Other Scrabble: 
\code{\link{scrabble_dictionary}()},
\code{\link{scrabble_words}()}
}
\concept{Scrabble}
