% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scrabble_words}
\alias{scrabble_words}
\title{Scrabble: Highest score words finder}
\usage{
scrabble_words(
  tiles,
  words,
  scores,
  free = 0,
  force_start = "",
  force_end = "",
  force_str = "",
  force_n = NA,
  quiet = FALSE
)
}
\arguments{
\item{tiles}{Character. The letters you wish to consider}

\item{words}{Dataframe. Words from \code{scrabble_dictionary()}}

\item{scores}{Dataframe. Must contain two columns: "tiles" with every
letter of the alphabet and "scores" for each letter's score.}

\item{free}{Integer. How many free blank tiles you have?}

\item{force_start, force_end}{Character. Force words to start or end with
a pattern of letters and position. Examples: "S" or "SO" or "__S_O"...}

\item{force_str}{Character vector. Force words to contain strings.}

\item{force_n}{Integer. Force words to be n characters long.}

\item{quiet}{Boolean. Do not print words as they are being searched.}
}
\description{
Find highest score words given a set of letters, rules, and 
language to win at Scrabble! You just have to find the best 
place to post your tiles.
}
\examples{
\dontrun{
# Dictionary for Spanish Words
es_words <- scrabble_dictionary("es")

# scores for Spanish words
es_scores <- data.frame(
  tiles = c(tolower(LETTERS)[1:14],"ñ",tolower(LETTERS)[15:length(LETTERS)]),
  scores = c(1,3,2,2,1,4,3,4,1,8,10,1,3,1,8,1,3,5,1,1,1,2,4,10,10,5,10))
  
scrabble_words(tiles = "holase",
               words = es_words$words,
               scores = es_scores,
               free = 1,
               force_start = "_o_a")
# A tibble: 96 x 2
  word    scores
  <chr>    <int>
1 hozase     18
2 hozas      17
3 hoza       16
4 hojas      15
5 hoja       14
6 lozas      14
7 solaz      14
8 xolas      14
9 hoyase     13
10 loza      13
# … with 86 more rows
}
}
\seealso{
Other Scrabble: 
\code{\link{scrabble_dictionary}()},
\code{\link{scrabble_score}()}
}
\concept{Scrabble}
