% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{seq_surnames}
\alias{seq_surnames}
\title{Family Tree Surnames Sequence}
\usage{
seq_surnames(n = 1)
}
\arguments{
\item{n}{Integer. Number of generations. Notice this will generate sequences
of 2^n integer values.}
}
\value{
Integer vector.
}
\description{
Family Tree Surnames Sequence
}
\examples{
seq_surnames(1) # Expected: 1
seq_surnames(2) # Expected: 1, 2
seq_surnames(3) # Expected: 1, 5, 3, 7, 2, 6, 4, 8
seq_surnames(4) # Expected: 1, 9, 5, 13, 3, 11, 7, 15, 2, 10, 6, 14, 4, 12, 8, 16
}
