% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{set_colnames}
\alias{set_colnames}
\alias{set_names}
\title{Set Columns/Rows Names within Pipes}
\usage{
set_colnames(x, value)

set_names(x, value)
}
\arguments{
\item{x}{data.frame or vector.}

\item{value}{Character vector. Names to by added. Same length as \code{x}.}
}
\description{
Borrowed from \code{magrittr}.
}
