% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slackSend}
\alias{slackSend}
\title{Send Slack Message (Webhook)}
\usage{
slackSend(text, title = "", pretext = "", hook = NA, creds = NA)
}
\arguments{
\item{text, title, pretext}{Character. Content on you Slack message.}

\item{hook}{Character. Web hook URL. Ths value will be overwritten by
creds if correctly used.}

\item{creds}{Character. Credential's dir (see \code{get_creds()}). Set
hook URL into the "slack" list in your YML file. Will use first value.}
}
\value{
Invisible POST response
}
\description{
This function send a Slack message using its Webhooks.
}
\details{
For more help, you can follow the
\href{https://api.slack.com/messaging/webhooks#posting_with_webhooks}{Sending messages using Incoming Webhooks}
original documentarion.
}
\examples{
\dontrun{
slackSend(text = "This is a message", title = "TEST", pretext = Sys.info()["user"])
}
}
\seealso{
Other API:
\code{\link[=bring_api]{bring_api()}},
\code{\link[=fb_accounts]{fb_accounts()}},
\code{\link[=fb_ads]{fb_ads()}},
\code{\link[=fb_creatives]{fb_creatives()}},
\code{\link[=fb_insights]{fb_insights()}},
\code{\link[=fb_process]{fb_process()}},
\code{\link[=fb_report_check]{fb_report_check()}},
\code{\link[=fb_rf]{fb_rf()}},
\code{\link[=fb_token]{fb_token()}},
\code{\link[=gemini_ask]{gemini_ask()}},
\code{\link[=gpt_ask]{gpt_ask()}},
\code{\link[=queryGA]{queryGA()}}

Other Credentials:
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=encrypt_file]{encrypt_file()}},
\code{\link[=get_credentials]{get_credentials()}},
\code{\link[=get_tweets]{get_tweets()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=queryDB]{queryDB()}},
\code{\link[=queryGA]{queryGA()}},
\code{\link[=stocks_file]{stocks_file()}}
}
\concept{API}
\concept{Credentials}
