% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{splot_divs}
\alias{splot_divs}
\title{Portfolio Plots: Dividends per Year and Quarter}
\usage{
splot_divs(p, type = 1)
}
\arguments{
\item{p}{Dataframe. Result from \code{daily_portfolio()}}

\item{type}{Integer. Typo of plot. 1 for incomes.}
}
\value{
ggplot object
}
\description{
This function plots a portfolio's historical dividends incomes
grouped by quarter an year.
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}

Other Investment Plots: 
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()}
}
\concept{Investment}
\concept{Investment Plots}
