% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{splot_etf}
\alias{splot_etf}
\title{Portfolio's Sector Distribution (ETFs)}
\usage{
splot_etf(s, save = FALSE)
}
\arguments{
\item{s}{Dataframe. Result from daily_stocks()}

\item{save}{Boolean. Save plot into a local file?}
}
\description{
This function lets the user plot his portfolio's distribution, 
specifically ETF's sectors
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_report}()}
}
\concept{Investment}
