% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{splot_growth}
\alias{splot_growth}
\title{Portfolio Plots: Growth (Cash + Invested)}
\usage{
splot_growth(p, save = FALSE)
}
\arguments{
\item{p}{Dataframe. Result from \code{daily_portfolio()}}

\item{save}{Boolean. Save plot into a local file?}
}
\value{
ggplot object
}
\description{
This function plots your portfolio's growth, in cash and investment,
since inception.
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}

Other Investment Plots: 
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()}
}
\concept{Investment}
\concept{Investment Plots}
