% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{splot_summary}
\alias{splot_summary}
\title{Portfolio Plots: Total Summary}
\usage{
splot_summary(p, s, save = FALSE)
}
\arguments{
\item{p}{Dataframe. Result from \code{daily_portfolio()}}

\item{s}{Dataframe. Result from \code{daily_stocks()}}

\item{save}{Boolean. Save plot into a local file?}
}
\value{
ggplot object
}
\description{
This function plots a summary for the whole portfolio, showing
how much have you invested, how much has each ticker changed, etc.
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}

Other Investment Plots: 
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_types}()}
}
\concept{Investment}
\concept{Investment Plots}
