% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{splot_types}
\alias{splot_types}
\title{Portfolio Plots: Types of Stocks}
\usage{
splot_types(s, save = FALSE)
}
\arguments{
\item{s}{Dataframe. Result from \code{daily_stocks()}}

\item{save}{Boolean. Save plot into a local file?}
}
\value{
ggplot object
}
\description{
This function lets the user plot types or categories of tickers.
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}

Other Investment Plots: 
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()}
}
\concept{Investment}
\concept{Investment Plots}
