% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{spread_list}
\alias{spread_list}
\title{Spread list column into new columns}
\usage{
spread_list(df, col, str = NA, replace = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{col}{Variable name.}

\item{str}{Character. Start column names with. If set to `NA`, original
name of column will be used.}

\item{replace}{Boolean. Replace original values (delete column)}
}
\description{
Spread an existing list column into new columns on a data.frame. Note
that every element on every observation must have a name for the function
to do its work. Original column will be automatically suppressed but you
can set the `replace` argument to avoid it.
}
