% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statusbar.R
\name{statusbar}
\alias{statusbar}
\title{Progressive Status Bar (Loading)}
\usage{
statusbar(
  run = 1,
  max.run = 100,
  label = run,
  msg = "",
  type = Sys.getenv("LARES_STATUSBAR"),
  start_time = NA,
  multiples = 1,
  alarm = FALSE
)
}
\arguments{
\item{run}{Iterator. for loop or an integer with the current loop number.
Start with 1 preferibly}

\item{max.run}{Number. Maximum number of loops}

\item{label}{String. With additionaly information to be printed
at the end of the line. The default is \code{run}.}

\item{msg}{Character. Finish message.}

\item{type}{Character. Loading type style: equal, domino, sword, filled.}

\item{start_time}{POSIXct. Start time to consider. If NA, then
when first iteration starts will be set as start time. Useful
for when first iteration is showed as done but started a few
seconds/minutes ago.}

\item{multiples}{Integer. Only print when multiples of N (to avoid)
wasting resources on fast and lots of iterations.}

\item{alarm}{Boolean. Ping (sound) when done. Requires \code{beepr}.}
}
\value{
No return value, called for side effects.
}
\description{
This function lets the user view a progressbar for a 'for' loop.
}
\examples{
for (i in 1:9) {
  statusbar(i, 9, multiples = 2)
  Sys.sleep(0.3)
}
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=updateLares]{updateLares()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Tools}
