% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_daily_plot}
\alias{stocks_daily_plot}
\title{Stocks Daily Plot}
\usage{
stocks_daily_plot(portfolio, daily)
}
\arguments{
\item{portfolio}{Dataframe. Output of the stocks_performance function}

\item{daily}{Dataframe. Daily data}
}
\description{
This function lets the user plot stocks daily change
}
