% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_file}
\alias{stocks_file}
\title{Get Personal Portfolio's Data}
\usage{
stocks_file(filename = NA,
  token_dir = "~/Dropbox (Personal)/Documentos/Docs/Data", auto = TRUE,
  sheets = c("Portafolio", "Fondos", "Transacciones"), keep_old = TRUE)
}
\arguments{
\item{filename}{Characeter. Import a local Excel file}

\item{token_dir}{Character. Where is my personal token for Dropbox connection?}

\item{auto}{Boolean. Automatically user my local personal file?}

\item{sheets}{Character Vector. Names of each sheet containing Portfolio summary,
Cash, and Transactions information}

\item{keep_old}{Boolean. Include sold tickers eventhough not currently in portfolio?}
}
\description{
This function lets me download my personal Excel with my Portfolio 
data, locally or from Dropbox. You can also use it for local files
}
\seealso{
Other Investment: \code{\link{daily_portfolio}},
  \code{\link{daily_stocks}}, \code{\link{etf_sector}},
  \code{\link{splot_change}}, \code{\link{splot_etf}},
  \code{\link{splot_growth}}, \code{\link{splot_roi}},
  \code{\link{splot_summary}}, \code{\link{splot_types}},
  \code{\link{stocks_hist}}, \code{\link{stocks_obj}},
  \code{\link{stocks_report}}
}
\concept{Investment}
