% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_quote}
\alias{stocks_quote}
\alias{stocks_hist}
\alias{plot.stocks_hist}
\title{Download Stocks Historical and Current Values}
\usage{
stocks_quote(symbols, ...)

stocks_hist(
  symbols = c("VTI", "META"),
  from = Sys.Date() - 365,
  to = Sys.Date(),
  today = TRUE,
  tax = 15,
  parg = FALSE,
  cache = TRUE,
  quiet = FALSE,
  ...
)

\method{plot}{stocks_hist}(x, type = 1, ...)
}
\arguments{
\item{symbols}{Character Vector. List of symbols to download historical data.}

\item{...}{Additional parameters.}

\item{from, to}{Date. Dates for range. If not set, 1 year will be downloaded.
Do use more than 4 days or will be over-written.}

\item{today}{Boolean. Do you wish to add today's live quote? This will happen
only if to value is the same as today's date}

\item{tax}{Numeric. How much [0-99] of your dividends are gone with taxes?}

\item{parg}{Boolean. Personal argument. Used to personalize stuff, in this
case, taxes changed from A to B in given date (hard-coded)}

\item{cache}{Boolean. Use daily cache if available?}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}

\item{x}{stocks_hist object}

\item{type}{Integer. Select type of plot.}
}
\value{
data.frame with Symbol, Type of stock, Quote time, current value,
Daily Change, Market, and Symbol Name.
}
\description{
\code{stocks_quote()} lets the user download stocks live data.

\code{stocks_hist()} lets the user download stocks historical data.
}
\examples{
\dontrun{
# Multiple quotes at the same time
stocks_quote(c("VTI", "VOO", "TSLA"))
}
\dontrun{
df <- stocks_hist(symbols = c("VTI", "META", "FIW"), from = Sys.Date() - 180)
print(head(df))
plot(df)
}
}
\seealso{
Other Investment:
\code{\link[=etf_sector]{etf_sector()}},
\code{\link[=splot_summary]{splot_summary()}},
\code{\link[=stocks_file]{stocks_file()}}

Other Scrapper:
\code{\link[=filesGD]{filesGD()}},
\code{\link[=gtrends_related]{gtrends_related()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=mp3_get]{mp3_get()}},
\code{\link[=readGS]{readGS()}},
\code{\link[=splot_summary]{splot_summary()}}
}
\concept{Investment}
\concept{Scrapper}
