% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_hist_fix}
\alias{stocks_hist_fix}
\title{Fix Historical Data on Stocks}
\usage{
stocks_hist_fix(dailys, dividends, transactions, expenses = 7)
}
\arguments{
\item{dailys}{Dataframe. Daily values. Structure: "Date", "Symbol", 
"Open", "High", "Low", "Close", "Volume", "Adjusted"}

\item{dividends}{Dataframe. Dividends. Structure: "Symbol", "Date", 
"Div", "DivReal"}

\item{transactions}{Dataframe. Transactions. Structure: "ID", "Inv", 
"CODE", "Symbol", "Date", "Quant", "Value", "Amount", "Description"}

\item{expenses}{Numeric. How much does that bank or broker charges per
transaction? Absolute value.}
}
\description{
This function lets the user fix downloaded stock data into a usefull 
format output
}
\seealso{
Other Investment: \code{\link{etf_sector}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_html}}, \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
