% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_obj}
\alias{stocks_obj}
\title{Portfolio's Calculations and Plots}
\usage{
stocks_obj(data = stocks_file(creds = creds), cash_fix = 0, tax = 30,
  sectors = TRUE)
}
\arguments{
\item{data}{List. Containing the following dataframes: portfolio,
transactions, cash. They have to follow the original xlsx format}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value 
to your cash balance}

\item{tax}{Numeric. How much [0-99] of your dividends are gone with taxes?}

\item{sectors}{Boolean. Return sectors segmentation for ETFs?}

\item{creds}{Character. Where is my personal token for Dropbox connection?}
}
\description{
This function lets the user create his portfolio's calculations and
plots for further study.
}
\seealso{
Other Investment: \code{\link{daily_portfolio}},
  \code{\link{daily_stocks}}, \code{\link{etf_sector}},
  \code{\link{splot_change}}, \code{\link{splot_etf}},
  \code{\link{splot_growth}}, \code{\link{splot_roi}},
  \code{\link{splot_summary}}, \code{\link{splot_types}},
  \code{\link{stocks_file}}, \code{\link{stocks_hist}},
  \code{\link{stocks_report}}
}
\concept{Investment}
