% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_quote}
\alias{stocks_quote}
\title{Download Stocks Historical Data}
\usage{
stocks_quote(ticks)
}
\arguments{
\item{ticks}{Character Vector. Symbols/Tickers to quote in real time}
}
\description{
This function lets the user download stocks historical data
}
\examples{
\dontrun{
# One quote
one <- stocks_quote("VTI")
# Multiple quotes
mult <- stocks_quote(c("VTI","VOO","TSLA"))
}
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_report}()}
}
\concept{Investment}
