% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_total_plot}
\alias{stocks_total_plot}
\title{Stocks Total Performance Plot}
\usage{
stocks_total_plot(stocks_perf, portfolio_perf, daily, trans, cash,
  save = FALSE)
}
\arguments{
\item{stocks_perf}{Dataframe. Output of the stocks_performance function}

\item{portfolio_perf}{Dataframe. Output of the portfolio_performance function}

\item{daily}{Dataframe. Daily data}

\item{trans}{Dataframe. Transactions data}

\item{cash}{Dataframe. Cash data}

\item{save}{Boolean. Export plot as an image?}
}
\description{
This function lets the user plot his stocks total performance
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks_hist}},
  \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}}
}
\concept{Investment}
