% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{target_set}
\alias{target_set}
\title{Set Target Value in Target Variable}
\usage{
target_set(tag, score, target = "auto", quiet = FALSE)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{target}{Value. Which is your target positive value? If
set to 'auto', the target with largest mean(score) will be
selected. Change the value to overwrite. Only used when binary
categorical model.}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}
}
\value{
List. Contains original data.frame \code{df} and
\code{which} with the target variable.
}
\description{
This function detects or forces the target value when predicting
a categorical binary model. This is an auxiliary function.
}
