% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textFeats}
\alias{textFeats}
\title{Create features out of text}
\usage{
textFeats(text, auto = TRUE, contains = NA, prc = FALSE)
}
\arguments{
\item{text}{Character vector}

\item{auto}{Boolean. Auto create some useful parameters?}

\item{contains}{Character vector. Which columns do you wish to add
with a contains (counter) string validator?}

\item{prc}{Boolean. Also add percentage of each column compared with length?}
}
\value{
data.frame with additional features based on \code{text}.
}
\description{
This function creates a data.frame with features based on a text vector
}
\examples{
textFeats("Bernardo Lares")
textFeats("Bernardo Lares 123!", prc = TRUE)
textFeats("I'm 100\% Lares...", contains = c("Lares", "lares"))
textFeats(c("GREAT library!!", "Have you tried this 2?", "Happy faces :D :-)"))
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}

Other Text Mining:
\code{\link[=cleanText]{cleanText()}},
\code{\link[=ngrams]{ngrams()}},
\code{\link[=remove_stopwords]{remove_stopwords()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=sentimentBreakdown]{sentimentBreakdown()}},
\code{\link[=textCloud]{textCloud()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=topics_rake]{topics_rake()}}
}
\concept{Data Wrangling}
\concept{Text Mining}
