% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textTokenizer}
\alias{textTokenizer}
\title{Tokenize Vectors into Words}
\usage{
textTokenizer(
  text,
  exclude = NULL,
  lang = NULL,
  min_word_freq = 5,
  min_word_len = 2,
  keep_spaces = FALSE,
  lowercase = TRUE,
  remove_numbers = TRUE,
  remove_punct = TRUE,
  remove_lettt = TRUE,
  laughs = TRUE,
  utf = TRUE,
  df = FALSE,
  h2o = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{text}{Character vector. Sentences or texts you wish to tokenize.}

\item{exclude}{Character vector. Which words do you wish to exclude?}

\item{lang}{Character. Language in text (used for stop words). Example:
"spanish" or "english". Set to \code{NA} to ignore.}

\item{min_word_freq}{Integer. This will discard words that appear
less than <int> times. Defaults to 2. Set to \code{NA} to ignore.}

\item{min_word_len}{Integer. This will discard words that have
less than <int> characters. Defaults to 5. Set to \code{NA} to ignore.}

\item{keep_spaces}{Boolean. If you wish to keep spaces in each line
to keep unique compound words, separated with spaces, set to TRUE.
For example, 'one two' will be set as 'one_two' and treated as
a single word.}

\item{lowercase, remove_numbers, remove_punct}{Boolean.}

\item{remove_lettt}{Boolean. Repeated letters (more than 3 consecutive).}

\item{laughs}{Boolean. Try to unify all laughs texts.}

\item{utf}{Boolean. Transform all characters to UTF (no accents and crazy symbols)}

\item{df}{Boolean. Return a dataframe with a one-hot-encoding kind of
results? Each word is a column and returns if word is contained.}

\item{h2o}{Boolean. Return \code{H2OFrame}?}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}
}
\value{
data.frame. Tokenized words with counters.
}
\description{
This function transforms texts into words, calculate frequencies,
supress stop words in a given language.
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}

Other Text Mining:
\code{\link[=cleanText]{cleanText()}},
\code{\link[=ngrams]{ngrams()}},
\code{\link[=remove_stopwords]{remove_stopwords()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=sentimentBreakdown]{sentimentBreakdown()}},
\code{\link[=textCloud]{textCloud()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=topics_rake]{topics_rake()}}
}
\concept{Data Wrangling}
\concept{Text Mining}
