% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lares2.R
\name{theme_lares2}
\alias{theme_lares2}
\title{lares Theme for ggplot2}
\usage{
theme_lares2(font = "Arial Narrow", base_size = 12.5,
  main_colour = "darkorange3", hard_colour = "black",
  soft_colour = "grey30", bg_colour = "white", legend = "right",
  mg = 15, pal = 0)
}
\arguments{
\item{font, base_size}{Character and numeric. Base font family and size}

\item{main_colour, hard_colour, soft_colour, bg_colour}{Character. Main colours for your theme}

\item{legend}{Character. Legend position: top, right, bottom, left}

\item{mg}{Numeric. External margin}

\item{pal}{Integer. 1 for fill and colour palette, 2 for only colour palette,
3 for personal labels-colour palette. 0 or else for nothing.}
}
\description{
Based on hrbrthemes' theme_ipsum and customized for lares' use.
}
\section{Why Arial Narrow?}{

First and foremost, Arial Narrow is generally installed by default or readily
available on any modern system, so it's "free"-ish; plus, it is a condensed font
with solid default kerning pairs and geometric numbers.
}

\seealso{
Other Visualization: \code{\link{corr_plot}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{textCloud}},
  \code{\link{theme_lares}}, \code{\link{tree_var}}
}
\concept{Visualization}
