% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{time_forecast}
\alias{time_forecast}
\title{Simple Forecast}
\usage{
time_forecast(time, values, n_future = 15, use_last = TRUE,
  plot = TRUE, automl = FALSE,
  project = "Simple Forecast using Machine Learning")
}
\arguments{
\item{time}{POSIX. Vector with dates or time values}

\item{values}{Numeric. Vector with numerical values}

\item{n_future}{Integer. How many steps do you wish to forecast?}

\item{use_last}{Boolena. Use last observation?}

\item{plot}{Boolean. If you wish to plot your results}

\item{automl}{Boolean. Use lares::h2o_automl()}

\item{project}{Character. Name of your forecast project for plot title}
}
\description{
This function lets the user create a forecast setting a time series
and a numerical value.
}
