% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_trends.R
\name{trendsRelated}
\alias{trendsRelated}
\title{Google Trends: Related Plot}
\usage{
trendsRelated(gtrend, top = NA, title = NA, note = NA,
  exclude = c())
}
\arguments{
\item{gtrend}{List. Result from gtrends(keyword, geo, time)}

\item{top}{Integer. Filter top n results only}

\item{title}{Character. Custom title for the plot}

\item{note}{Character. Add a note to the plot if needed}

\item{exclude}{Character vector. Which observations do you wish to exclude?}
}
\description{
This function creates a plot with google trend's related topics 
and queries, and let the user compare different keywords.
}
