% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_trends.R
\name{trendsTime}
\alias{trendsTime}
\title{Google Trends: Timelines Plot}
\usage{
trendsTime(gtrend, title = NA)
}
\arguments{
\item{gtrend}{List. Result from gtrends(keyword, geo, time)}

\item{title}{Character. Custom title for the plot}
}
\description{
This function creates a plot with google trend's data on timelines
and let the user compare different keywords.
}
\seealso{
Other Google Trends: 
\code{\link{trendsRelated}()}

Other Google: 
\code{\link{geoAddress}()},
\code{\link{queryGA}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()}
}
\concept{Google}
\concept{Google Trends}
