% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{trim_mp3}
\alias{trim_mp3}
\title{Trim MP3 Audio File}
\usage{
trim_mp3(
  file,
  start_time = 0,
  end_time = NA,
  overwrite = FALSE,
  ext = "mp3",
  quiet = FALSE
)
}
\arguments{
\item{file}{Character. File name to trim.}

\item{start_time}{Numeric. Start and end time
to trim the audio output in seconds.}

\item{end_time}{Numeric. Start and end time
to trim the audio output in seconds.}

\item{overwrite}{Boolean. Overwrite original file?}

\item{ext}{Character. File extension/type.}

\item{quiet}{Boolean. Keep quiet? If not, print messages.}
}
\description{
This function trims MP3 files given a start and/or end numeric
timestamp. Requires \code{ffmpeg} installed in your machine.
}
\seealso{
Other Audio: 
\code{\link{get_mp3}()}
}
\concept{Audio}
