% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateLares}
\alias{updateLares}
\title{Update the library (dev or CRAN version)}
\usage{
updateLares(force = FALSE, dev = TRUE, all = FALSE, local = FALSE, fb = FALSE)
}
\arguments{
\item{force}{Boolean. Force install.}

\item{dev}{Boolean. Developer version (Github)? If not, CRAN version.}

\item{all}{Boolean. Install other recommended libraries? Kinda Docker install!}

\item{local}{Boolean. Install package with local files? (or Github repo).}

\item{fb}{Boolean. From FB instance? Personal internal use.}
}
\value{
No return value, called for side effects.
}
\description{
This auxiliary function lets the user update \code{lares} to latest
CRAN or developer version.
}
\examples{
\dontrun{
# Update to CRAN version
updateLares(dev = FALSE)

# Update to development version from GitHub
updateLares(dev = TRUE)

# Install with all recommended packages
updateLares(all = TRUE)
}
}
\seealso{
Other Tools:
\code{\link[=autoline]{autoline()}},
\code{\link[=bind_files]{bind_files()}},
\code{\link[=bring_api]{bring_api()}},
\code{\link[=cal_split]{cal_split()}},
\code{\link[=chr2num]{chr2num()}},
\code{\link[=db_download]{db_download()}},
\code{\link[=db_upload]{db_upload()}},
\code{\link[=dont_sleep]{dont_sleep()}},
\code{\link[=export_plot]{export_plot()}},
\code{\link[=export_results]{export_results()}},
\code{\link[=files_functions]{files_functions()}},
\code{\link[=font_exists]{font_exists()}},
\code{\link[=formatColoured]{formatColoured()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=glued]{glued()}},
\code{\link[=grepm]{grepm()}},
\code{\link[=h2o_selectmodel]{h2o_selectmodel()}},
\code{\link[=haveInternet]{haveInternet()}},
\code{\link[=image_metadata]{image_metadata()}},
\code{\link[=importxlsx]{importxlsx()}},
\code{\link[=ip_data]{ip_data()}},
\code{\link[=json2vector]{json2vector()}},
\code{\link[=list_cats]{list_cats()}},
\code{\link[=listfiles]{listfiles()}},
\code{\link[=mail_send]{mail_send()}},
\code{\link[=markdown2df]{markdown2df()}},
\code{\link[=move_files]{move_files()}},
\code{\link[=msplit]{msplit()}},
\code{\link[=myip]{myip()}},
\code{\link[=quiet]{quiet()}},
\code{\link[=read.file]{read.file()}},
\code{\link[=statusbar]{statusbar()}},
\code{\link[=tic]{tic()}},
\code{\link[=try_require]{try_require()}},
\code{\link[=warnifnot]{warnifnot()}},
\code{\link[=what_size]{what_size()}}
}
\concept{Tools}
