% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{vector2text}
\alias{vector2text}
\alias{v2t}
\title{Convert a vector into a comma separated text}
\usage{
vector2text(vector, sep = ", ", quotes = TRUE, force_single = FALSE, and = "")

v2t(vector, sep = ", ", quotes = TRUE, force_single = FALSE, and = "")
}
\arguments{
\item{vector}{Vector. Vector with more than 1 observation.}

\item{sep}{Character. String text wished to insert between values.}

\item{quotes}{Boolean. Bring simple quotes for each observation.}

\item{force_single}{Boolean. Force single quotes by replacing \code{\"}.}

\item{and}{Character. Add 'and' or something before last observation.
Not boolean variable so it can be used on other languages. Note that
the last comma will be suppressed if \code{Sys.getenv("LARES_NUMFORMAT")}
is set to \code{1} and you have less than 3 values.}
}
\value{
Vector pasting \code{vector} values into a single string
}
\description{
Convert a vector into a comma separated text
}
\examples{
vector2text(LETTERS[1:5])
vector2text(c(1:5), quotes = FALSE)
vector2text(c(1:5), quotes = FALSE, sep = "-")
vector2text(c(1:5), and = "and also")
vector2text(c("Text", "R's"), force_single = TRUE)
# Shorter function with same purpose
v2t(LETTERS[1:5])
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=year_month]{year_month()}},
\code{\link[=zerovar]{zerovar()}}
}
\concept{Data Wrangling}
