% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.R
\name{x2y}
\alias{x2y}
\alias{x2y_calc}
\alias{plot.x2y}
\title{Ranked Predictive Power Cross-Features (x2y metric)}
\usage{
x2y(df, target = NULL, plot = FALSE, top = 20, quiet = "auto", ...)

x2y_calc(x, y, confidence = FALSE, bootstraps = 10)

\method{plot}{x2y}(x, ...)
}
\arguments{
\item{df}{data.frame}

\item{target}{Character. If you are only interested in the \code{x2y}
values between particular variable(s) in \code{df}, set
name(s) of the variable(s) you are interested in. Keep \code{NULL}
to calculate for every variable (column).}

\item{plot}{Boolean. Return a plot? If not, only a data.frame with calculated
results will be returned.}

\item{top}{Integer. Show/plot only top N predictive cross-features. Set
to \code{NULL} to return all.}

\item{quiet}{Boolean. Keep quiet? If not, show progress bar.}

\item{...}{Additional parameters passed to \code{x2y_calc()}}

\item{x, y}{Vectors. Categorical or numerical vectors of same length.}

\item{confidence}{Boolean. Calculate 95% confidence intervals estimated
with n \code{bootstraps}.}

\item{bootstraps}{Integer. If \code{confidence=TRUE}, how many bootstraps?
The more iterations we run the more precise the \code{confidence}internal.}
}
\description{
The porcentual reduction in error when we go from a baseline model to
a predictive model measures the strength of the relationship between 
features x and y. This metric, x2y, measures the ability of x to 
predict y. Based on Rama Ramakrishnan's post: An Alternative to the
Correlation Coefficient That Works For Numeric and Categorical Variables.
}
\examples{
\dontrun{
x2y(dft, c("Survived","Age"))
x2y(dft, "Fare", confidence = TRUE, bootstraps = 10)
}
}
