% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{year_month}
\alias{year_month}
\alias{year_quarter}
\alias{year_week}
\title{Convert Date into Year-Month, Year-Quarter or Year-Week Format}
\usage{
year_month(date)

year_quarter(date)

year_week(date)
}
\arguments{
\item{date}{Date vector. Date to transform format.}
}
\value{
Vector with dates reformatted
}
\description{
This function lets the user convert a date into YYYY-MM, YYYY-QX,
or YYYY-WW format easily.
}
\examples{
year_month(Sys.Date())
year_quarter(Sys.Date())
year_week(Sys.Date())
}
\seealso{
Other Data Wrangling:
\code{\link[=balance_data]{balance_data()}},
\code{\link[=categ_reducer]{categ_reducer()}},
\code{\link[=cleanText]{cleanText()}},
\code{\link[=date_cuts]{date_cuts()}},
\code{\link[=date_feats]{date_feats()}},
\code{\link[=file_name]{file_name()}},
\code{\link[=formatHTML]{formatHTML()}},
\code{\link[=holidays]{holidays()}},
\code{\link[=impute]{impute()}},
\code{\link[=left]{left()}},
\code{\link[=normalize]{normalize()}},
\code{\link[=num_abbr]{num_abbr()}},
\code{\link[=ohe_commas]{ohe_commas()}},
\code{\link[=ohse]{ohse()}},
\code{\link[=quants]{quants()}},
\code{\link[=removenacols]{removenacols()}},
\code{\link[=replaceall]{replaceall()}},
\code{\link[=replacefactor]{replacefactor()}},
\code{\link[=textFeats]{textFeats()}},
\code{\link[=textTokenizer]{textTokenizer()}},
\code{\link[=vector2text]{vector2text()}},
\code{\link[=zerovar]{zerovar()}}
}
\concept{Data Wrangling}
