% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{year_week}
\alias{year_week}
\title{Convert year week format YYYY-WW}
\usage{
year_week(date)
}
\arguments{
\item{date}{Date. Date we wish to transform}
}
\description{
This function lets the user convert a date into YYYY-WW format
}
