% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{auto_preprocess}
\alias{auto_preprocess}
\title{Auto-Preprocessing data for a Machine Learning Model}
\usage{
auto_preprocess(df, num2fac = 7, print = FALSE, return = "results")
}
\arguments{
\item{df}{Dataframe. Dataframe to format, transform and input}

\item{num2fac}{Integer. Threshold for unique values on a numerical variable 
to transform into factor}

\item{return}{Character. Select if you wish the preprocessed data 'results' or 
recipe 'function' result}
}
\description{
This function helps us preprocess data automatically.
}
