% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{auto_preprocess}
\alias{auto_preprocess}
\title{Auto-Preprocessing data for a Machine Learning Model}
\usage{
auto_preprocess(df, num2fac = 7, impute = FALSE, ohe = FALSE,
  scale = FALSE, logs = FALSE, print = FALSE)
}
\arguments{
\item{df}{Dataframe. Dataframe to format, transform and input}

\item{num2fac}{Integer. Threshold for unique values on a 
numerical variable to transform into factor}

\item{ohe}{Boolean. One hot encoding on variables with 3 or
more cateogries?}

\item{scale}{Boolean. Scale the data (normalized)}

\item{print}{Boolean. Print results summary}
}
\description{
This function helps us preprocess data automatically.
}
