% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{balance_data}
\alias{balance_data}
\title{Balance Binary Data by Resampling: Under-Over Sampling}
\usage{
balance_data(df, variable, rate = 1, seed = 0)
}
\arguments{
\item{df}{Vector or Dataframe. Contains different variables in each 
column, separated by a specific character}

\item{variable}{Character. Which binary variable should we use to resample df}

\item{rate}{Numeric. How many X for every Y we need? Default: 1. If there are
more than 2 unique values, rate will represent percentage for number of rows}

\item{seed}{Numeric. Seed to replicate and obtain same values}
}
\description{
This function lets the user balance a given data.frame by resampling
with a given relation rate and a binary feature.
}
\examples{
\dontrun{
data(dft)
df <- balance_data(dft, "Survived", rate = 1, seed = 123)
# Resampled from: 549 x 342
# Reducing size for: FALSE
# Into: 342 x 342
freqs(df, Survived)
#    Survived     n     p  pcum order
#    <lgl>    <int> <dbl> <dbl> <int>
# 1 FALSE      342    50    50     1
# 2 TRUE       342    50   100     2
}
}
\seealso{
Other Data Wrangling: 
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohse}()},
\code{\link{one_hot_encoding_commas}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
