% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{bind_files}
\alias{bind_files}
\title{Bind Files into Dataframe}
\usage{
bind_files(files)
}
\arguments{
\item{files}{Character vector. Files names.}
}
\value{
data.frame with data joined from all \code{files} passed.
}
\description{
This function imports and binds multiple files into a single
data.frame. Files must be inserted with absolute roots files names.
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{files_functions}()},
\code{\link{font_exists}()},
\code{\link{formatColoured}()},
\code{\link{formatHTML}()},
\code{\link{get_credentials}()},
\code{\link{glued}()},
\code{\link{grepm}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{list_cats}()},
\code{\link{listfiles}()},
\code{\link{mail_send}()},
\code{\link{move_files}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{warnifnot}()},
\code{\link{what_size}()}
}
\concept{Tools}
