% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{bindfiles}
\alias{bindfiles}
\title{Bind Files into Dataframe}
\usage{
bindfiles(files)
}
\arguments{
\item{files}{Dataframe}
}
\description{
This function imports and binds multiple files into a single 
data.frame. Files must be inserted with absolute roots filenames.
}
\seealso{
Other Tools: \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{statusbar}},
  \code{\link{zerovar}}
}
\concept{Tools}
