% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{bindfiles}
\alias{bindfiles}
\title{Bind Files into Dataframe}
\usage{
bindfiles(files)
}
\arguments{
\item{files}{Character vector. Files names.}
}
\value{
data.frame with data joined from all \code{files} passed.
}
\description{
This function imports and binds multiple files into a single 
data.frame. Files must be inserted with absolute roots files names.
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}
}
\concept{Tools}
