% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{bring_api}
\alias{bring_api}
\title{Get API (JSON) and Transform into data.frame}
\usage{
bring_api(url)
}
\arguments{
\item{url}{Character. API's URL}
}
\description{
This function lets the user bring API data as JSON format and transform it 
into data.frame. Designed initially for Hubspot but may work on other API
}
