% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{bring_api}
\alias{bring_api}
\title{Get API (JSON) and Transform into data.frame}
\usage{
bring_api(url, status = TRUE)
}
\arguments{
\item{url}{Character. API's URL}

\item{status}{Boolean. Display status message?}
}
\description{
This function lets the user bring API data as JSON format and transform it 
into data.frame. Designed initially for Hubspot but may work on other API
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{statusbar}},
  \code{\link{zerovar}}
}
\concept{Tools}
