% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{cal_split}
\alias{cal_split}
\title{Split ICS File}
\usage{
cal_split(file, n_splits = 5, prefix = "calendar_split_", dir = NULL)
}
\arguments{
\item{file}{Character. Path to the ICS file to split.}

\item{n_splits}{Integer. Number of files to split into.}

\item{prefix}{Character. Prefix for the output filenames.}

\item{dir}{Character. Directory to save the output files.
Defaults to same directory as input file.}
}
\value{
Invisible list. The paths of the created files.
}
\description{
This function splits a large ICS calendar file into smaller files.
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bind_files}()},
\code{\link{bring_api}()},
\code{\link{chr2num}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{dont_sleep}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{files_functions}()},
\code{\link{font_exists}()},
\code{\link{formatColoured}()},
\code{\link{formatHTML}()},
\code{\link{glued}()},
\code{\link{grepm}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{list_cats}()},
\code{\link{listfiles}()},
\code{\link{mail_send}()},
\code{\link{markdown2df}()},
\code{\link{move_files}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{warnifnot}()},
\code{\link{what_size}()}
}
\concept{Tools}
