% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Sampling Scores}
\usage{
calibrate(score, train, target, train_sample, target_sample)
}
\arguments{
\item{score}{Vector. Probability predictions of the model output}

\item{train}{Integer. Total row count in the training dataset}

\item{target}{Integer. Total row count of the target class 
in the training dataset}

\item{train_sample}{Integer. Total row count in the training 
dataset after sampling}

\item{target_sample}{Integer. Total row count of the target 
class in the training dataset after sampling}
}
\description{
This function lets the user calibrate a model's predictions when 
under or over sampling methods were applied when training it.
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{date_feats}}, \code{\link{dateformat}},
  \code{\link{formatNum}}, \code{\link{holidays}},
  \code{\link{left}}, \code{\link{normalize}},
  \code{\link{numericalonly}}, \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
