% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{categ_reducer}
\alias{categ_reducer}
\title{Reduce categorical values}
\usage{
categ_reducer(vector, nmin = 0, pmin = 0, pcummax = 100, top = NA,
  other_label = "other")
}
\description{
Reduce categorical values
}
