% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{categ_reducer}
\alias{categ_reducer}
\title{Reduce categorical values}
\usage{
categ_reducer(
  df,
  ...,
  nmin = 0,
  pmin = 0,
  pcummax = 100,
  top = NA,
  other_label = "other"
)
}
\arguments{
\item{df}{Categorical Vector}

\item{...}{Variables. Which variable do you wish to reduce?}

\item{nmin}{Integer. Number of minimum times a value is repeated}

\item{pmin}{Numerical. Porcentage of minimum times a value is repeated}

\item{pcummax}{Numerical. Top cumulative porcentage of most 
repeated values}

\item{top}{Integer. Keep the n most frequently repeated values}

\item{other_label}{Character. With which text do you wish to replace 
the filtered values with?}
}
\description{
This function lets the user reduce categorical values in a vector. 
It is tidyverse friendly for use on pipelines
}
\examples{
data(dft) # Titanic dataset
categ_reducer(dft, Embarked, top = 2) \%>\% freqs(Embarked)
categ_reducer(dft, Ticket, nmin = 7, other_label = "Other Ticket") \%>\% freqs(Ticket)
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
