% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{categoryCounter}
\alias{categoryCounter}
\title{Count Categories on a Dataframe}
\usage{
categoryCounter(df)
}
\arguments{
\item{df}{Categorical Vector}
}
\description{
This function lets the user count unique values in a categorical 
dataframe
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{cleanText}}, \code{\link{date_feats}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{holidays}}, \code{\link{impute}},
  \code{\link{left}}, \code{\link{normalize}},
  \code{\link{numericalonly}}, \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
