% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{categoryCounter}
\alias{categoryCounter}
\title{Count all categories on factor variables}
\usage{
categoryCounter(df)
}
\description{
This function lets the user count all distinct categories on factor variables
}
