% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_attr}
\alias{check_attr}
\title{Attribute checker}
\usage{
check_attr(object, attr, check, stop = TRUE)
}
\arguments{
\item{object}{Object of any kind}

\item{attr}{Character. Attribute to check}

\item{check}{Character. Attribute value}

\item{stop}{Boolean. Stop if doesn't check?}
}
\value{
No return value, called for side effects.
}
\description{
This function checks if an object has a specific attribute and
stops if not
}
