% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{check_opts}
\alias{check_opts}
\title{Validate options within vector}
\usage{
check_opts(inputs, options, type = "all", not = "stop", quiet = TRUE)
}
\arguments{
\item{inputs}{Vector character}

\item{options}{Vector character}

\item{type}{Character. Options: all, any}

\item{not}{Character. Options: stop, message, print, return}

\item{quiet}{Boolean. Keep quiet? If not, returns TRUE when checked}
}
\description{
This function validates if inputs match all/any of your options
and return error/message with possible options to use.
}
