% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{cleanText}
\alias{cleanText}
\title{Clean text}
\usage{
cleanText(text, spaces = TRUE, lower = TRUE)
}
\arguments{
\item{text}{Character Vector}

\item{spaces}{Boolean. Keep spaces?}

\item{lower}{Boolean. Transform all to lower case?}
}
\description{
This function lets the user clean text into getting only alphanumeric 
characters and no accents/symbols on letters.
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{dateformat}},
  \code{\link{formatNum}}, \code{\link{left}},
  \code{\link{missingness}}, \code{\link{normalize}},
  \code{\link{numericalonly}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{removenacols}}, \code{\link{removenarows}},
  \code{\link{replaceall}}, \code{\link{right}},
  \code{\link{vector2text}}, \code{\link{year_month}},
  \code{\link{year_week}}
}
\concept{Data Wrangling}
