% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{cleanText}
\alias{cleanText}
\title{Clean text}
\usage{
cleanText(text)
}
\arguments{
\item{text.}{Character Vector}
}
\description{
This function lets the user clean text into getting only alphanumeric 
characters and no accents/symbols on letters.
}
