% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusterKmeans}
\alias{clusterKmeans}
\title{K-Means Clustering Automated}
\usage{
clusterKmeans(df, force_n = NA, limit = 20, ohse = TRUE,
  comb = c(1, 2))
}
\arguments{
\item{df}{Dataframe}

\item{force_n}{Integer. Number of clusters}

\item{limit}{Integer. How many clusters should be considered?}

\item{ohse}{Boolean. Do you wish to automatically run one hot
encoding to non-numerical columns?}

\item{comb}{Vector. Which columns do you wish to plot? Select which 2
by name or position.}
}
\description{
This function lets the user cluster a whole data.frame automatically.
If needed, one hot encoding will be applied to categorical values.
}
