% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusterOptimalK}
\alias{clusterOptimalK}
\title{Visualize K-Means Clusters Methods}
\usage{
clusterOptimalK(
  df,
  method = c("wss", "silhouette", "gap_stat"),
  drop_na = TRUE,
  ohse = TRUE,
  norm = TRUE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{df}{Dataframe}

\item{method}{Character vector.}

\item{drop_na}{Boolean. Should NA rows be removed?}

\item{ohse}{Boolean. Do you wish to automatically run one hot
encoding to non-numerical columns?}

\item{norm}{Boolean. Should the data be normalized?}

\item{quiet}{Boolean. Keep quiet? If not, print messages}

\item{...}{Additional parameters passed to \code{clusterKmeans}}
}
\description{
Visualize cluster data for assorted values of k and methods such as
WSS, Silhouette and Gap statistic. See \code{fviz_nbclust} for more.
}
\examples{
\dontrun{
data("iris")
df <- subset(iris, select = c(-Species))

# Calculate and plot
clusterOptimalK(df)
}
}
\seealso{
Other Clusters: 
\code{\link{clusterKmeans}()},
\code{\link{clusterVisualK}()}
}
\concept{Clusters}
