% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{clusterVisualK}
\alias{clusterVisualK}
\title{Visualize K-Means Clusters for Several K}
\usage{
clusterVisualK(df, ks = 1:6, quiet = FALSE, plot = TRUE, ...)
}
\arguments{
\item{df}{Dataframe}

\item{ks}{Integer vector. Which k should be tested?}

\item{quiet}{Boolean. Keep quiet? If not, print messages}

\item{plot}{Boolean. Plot outcome?}

\item{...}{Additional parameters passed to \code{clusterKmeans}}
}
\description{
Visualize cluster data for assorted values of k.
}
\examples{
options("lares.font" = NA) # Temporal
data("iris")
df <- subset(iris, select = c(-Species))

# Calculate and plot
result <- clusterVisualK(df)

# You can use the data generated as well
lapply(result$data, function(x) head(x$cluster))
}
\seealso{
Other Clusters: 
\code{\link{clusterKmeans}()},
\code{\link{clusterOptimalK}()}
}
\concept{Clusters}
