% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{conf_mat}
\alias{conf_mat}
\title{Confussion Matrix}
\usage{
conf_mat(tag, score, thresh = 0.5)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{thresh}{Numeric. Value which splits the results for the 
confusion matrix when binary.}
}
\description{
This function calculates a Confussion Matrix using crosstab for
2 or more categories. You can either set the score and threshold
or the labels you wish to cross with.
}
