% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr}
\alias{corr}
\title{Correlation table}
\usage{
corr(df, method = "pearson", plot = FALSE)
}
\arguments{
\item{df}{Dataframe. It doesn't matter if it's got non-numerical
columns: they will be filtered!}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}
}
\description{
This function correlates a whole dataframe, filtering automatically
all numerical values.
}
