% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr}
\alias{corr}
\title{Correlation table}
\usage{
corr(
  df,
  method = "pearson",
  pvalue = FALSE,
  ignore = NA,
  dummy = TRUE,
  limit = 10,
  dates = FALSE,
  redundant = FALSE,
  logs = FALSE,
  top = NA
)
}
\arguments{
\item{df}{Dataframe. It doesn't matter if it's got non-numerical
columns: they will be filtered!}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}

\item{pvalue}{Boolean. Returns a list, with correlations and statistical 
significance (p-value) for each value}

\item{ignore}{Character vector. Which columns do you wish to exlude?}

\item{dummy}{Boolean. Should One Hot Encoding be applied to categorical columns?}

\item{limit}{Integer. Limit one hot encoding to the n most frequent 
values of each column. Set to \code{NA} to ignore argument.}

\item{dates}{Boolean. Do you want the function to create more features
out of the date/time columns?}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. It the
column only has two different values, should we keep both new columns?}

\item{logs}{Boolean. Automatically calculate log(values) for numerical
variables (not binaries)}

\item{top}{Integer. Select top N most relevant variables? Filtered 
and sorted by mean of each variable's correlations}
}
\description{
This function correlates a whole dataframe, filtering automatically
all numerical values.
}
\examples{
data(dft) # Titanic dataset
df <- dft[,2:5]
corr(df)
corr(df, ignore = "Pclass")
corr(df, redundant = TRUE)
corr(df, method = "spearman")
corr(df, pvalue = TRUE)
}
\seealso{
Other Calculus: 
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{model_metrics}()},
\code{\link{quants}()}

Other Correlations: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()}
}
\concept{Calculus}
\concept{Correlations}
