% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr}
\alias{corr}
\title{Correlation table}
\usage{
corr(df, method = "pearson", dummy = TRUE, dates = FALSE,
  redundant = TRUE, logs = FALSE, plot = FALSE, top = NA)
}
\arguments{
\item{df}{Dataframe. It doesn't matter if it's got non-numerical
columns: they will be filtered!}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}

\item{dummy}{Boolean. Should One Hot Encoding be applied to categorical columns?}

\item{dates}{Boolean. Do you want the function to create more features
out of the date/time columns?}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. It the
column only has two different values, should we keep both new columns?}

\item{logs}{Boolean. Automatically calculate log(values) for numerical
variables (not binaries)}

\item{plot}{Boolean. Do you wish to see a plot?}

\item{top}{Integer. Select top N most relevant variables? Filtered 
and sorted by mean of each variable's correlations}
}
\description{
This function correlates a whole dataframe, filtering automatically
all numerical values.
}
\seealso{
Other Calculus: \code{\link{ROC}}, \code{\link{conf_mat}},
  \code{\link{deg2num}}, \code{\link{dist2d}},
  \code{\link{errors}}, \code{\link{loglossBinary}},
  \code{\link{mae}}, \code{\link{mape}},
  \code{\link{model_metrics}}, \code{\link{mse}},
  \code{\link{quants}}, \code{\link{rmse}},
  \code{\link{rsqa}}, \code{\link{rsq}}

Other Correlations: \code{\link{corr_plot}},
  \code{\link{corr_var}}
}
\concept{Calculus}
\concept{Correlations}
