% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_cross}
\alias{corr_cross}
\title{Correlation Cross-Table}
\usage{
corr_cross(df, plot = TRUE, type = 1, max = 1, top = 25,
  local = 1, ignore = NA, contains = NA, rm.na = FALSE,
  dummy = TRUE)
}
\arguments{
\item{df}{Dataframe.}

\item{plot}{Boolean. Show and return a plot?}

\item{type}{Integer. Plot type. 1 is for overall rank. 2 is for local rank.}

\item{max}{Numeric. Maximum correlation permited (from 0 to 1)}

\item{top}{Integer. Return top n results only. Only valid when type = 1}

\item{local}{Integer. Show top n local correlations. Only valid when type = 2}

\item{ignore}{Character vector. Which columns do you wish to exlude?}

\item{contains}{Character vector. Filter cross-correlations with variables
that contains certain strings (using any value if vector used)}

\item{rm.na}{Boolean. Remove NAs?}

\item{dummy}{Boolean. Should One Hot Encoding be applied to categorical columns?}
}
\description{
This function creates a correlation full study and returns a rank
of the highest correlation variables obtained in a cross-table.
}
\seealso{
Other Correlations: \code{\link{corr_plot}},
  \code{\link{corr_var}}, \code{\link{corr}}

Other Exploratory: \code{\link{corr_var}},
  \code{\link{crosstab}}, \code{\link{df_str}},
  \code{\link{distr}}, \code{\link{freqs_df}},
  \code{\link{freqs}}, \code{\link{gain_lift}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{plot_cats}}, \code{\link{plot_df}},
  \code{\link{plot_nums}}, \code{\link{tree_var}},
  \code{\link{trendsRelated}}
}
\concept{Correlations}
\concept{Exploratory}
