% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_cross}
\alias{corr_cross}
\title{Correlation Cross-Table}
\usage{
corr_cross(df, plot = TRUE, max_pvalue = 1, type = 1, max = 1,
  top = 25, local = 1, ignore = NA, contains = NA, grid = FALSE,
  rm.na = FALSE, dummy = TRUE, redundant = FALSE,
  method = "pearson")
}
\arguments{
\item{df}{Dataframe.}

\item{plot}{Boolean. Show and return a plot?}

\item{max_pvalue}{Numeric. Filter non-significant variables. Range (0, 1]}

\item{type}{Integer. Plot type. 1 is for overall rank. 2 is for local rank.}

\item{max}{Numeric. Maximum correlation permited (from 0 to 1)}

\item{top}{Integer. Return top n results only. Only valid when type = 1}

\item{local}{Integer. Label top n local correlations. Only valid when type = 2}

\item{ignore}{Character vector. Which columns do you wish to exlude?}

\item{contains}{Character vector and Boolean. Filter cross-correlations 
with variables that contains certain strings (using any value if vector used).
This will automatically invert redundant default value.}

\item{grid}{Boolean. Separate into grids?}

\item{rm.na}{Boolean. Remove NAs?}

\item{dummy}{Boolean. Should One Hot Encoding be applied to categorical columns?}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. It the
column only has two different values, should we keep both new columns?}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}
}
\description{
This function creates a correlation full study and returns a rank
of the highest correlation variables obtained in a cross-table.
}
\seealso{
Other Correlations: \code{\link{corr_plot}},
  \code{\link{corr_var}}, \code{\link{corr}}

Other Exploratory: \code{\link{corr_var}},
  \code{\link{crosstab}}, \code{\link{df_str}},
  \code{\link{distr}}, \code{\link{freqs_df}},
  \code{\link{freqs}}, \code{\link{gain_lift}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{plot_cats}}, \code{\link{plot_df}},
  \code{\link{plot_nums}}, \code{\link{summer}},
  \code{\link{tree_var}}, \code{\link{trendsRelated}}
}
\concept{Correlations}
\concept{Exploratory}
