% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_plot}
\alias{corr_plot}
\title{Correlation plot}
\usage{
corr_plot(df, ignore = NA, method = "pearson", order = "FPC",
  type = "square", logs = FALSE)
}
\arguments{
\item{df}{Dataframe.}

\item{ignore}{Character vector. Which columns do you wish to exlude?}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}

\item{order}{Character. The ordering method of the correlation matrix.
Any of: c("original", "AOE", "FPC", "hclust", "alphabet")}

\item{type}{Character. The visualization method of correlation matrix 
to be used. 
Any of c("circle", "square", "ellipse", "number", "pie", "shade" 
and "color")}

\item{logs}{Boolean. Automatically calculate log(values) for numerical
variables (not binaries) and plot them}
}
\description{
This function correlates a whole dataframe with a single feature.
}
\seealso{
Other Visualization: \code{\link{distr}},
  \code{\link{freqs_df}}, \code{\link{freqs}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{textCloud}},
  \code{\link{theme_lares2}}, \code{\link{theme_lares}},
  \code{\link{tree_var}}

Other Correlations: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{corr}}
}
\concept{Correlations}
\concept{Visualization}
