% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_var}
\alias{corr_var}
\title{Correlation between variable and dataframe}
\usage{
corr_var(df, ..., method = "pearson", trim = 0, clean = FALSE,
  plot = TRUE, logs = FALSE, dates = TRUE, top = NA,
  ceiling = 100, zeroes = FALSE, save = FALSE, subdir = NA,
  file_name = "viz_corrvar.png")
}
\arguments{
\item{df}{Dataframe.}

\item{...}{Object. Name of the variable to correlate}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}

\item{trim}{Integer. Trim words until the nth character for 
categorical values (applies for both, target and values)}

\item{clean}{Boolean. Use lares::cleanText for categorical values (applies 
for both, target and values)}

\item{plot}{Boolean. Do you wish to plot the result? If set to TRUE, the
function will return only the plot and not the result's data}

\item{logs}{Boolean. Automatically calculate log(values) for numerical
variables (not binaries)}

\item{dates}{Boolean. Do you want the function to create more features
out of the date/time columns?}

\item{top}{Integer. If you want to plot the top correlations, 
define how many}

\item{ceiling}{Numeric. Remove all correlations above... Range: (0-100]}

\item{zeroes}{Do you wish to keep zeroes in correlations too?}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to 
save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function correlates a whole dataframe with a single feature.
}
