% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_var}
\alias{corr_var}
\title{Correlation between variable and dataframe}
\usage{
corr_var(df, var, method = "pearson", plot = TRUE)
}
\arguments{
\item{df}{Dataframe.}

\item{var}{Character. Name of the variable to correlate}

\item{method}{Character. Any of: c("pearson", "kendall", "spearman")}
}
\description{
This function correlates a whole dataframe with a single feature.
}
