% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Cross tabulation}
\usage{
crosstab(x, y, weight = 1, prow = FALSE, pcol = FALSE,
  pall = FALSE, decimals = 2, keep_nas = TRUE)
}
\arguments{
\item{x, y}{Vectors. For dependent and independent values}

\item{weight}{Vector. An optional vector for a weighted cross tabulation}

\item{prow, pcol, pall}{Boolean. Calculate percent values for rows, columns,
or the whole table, respectively.}

\item{decimals}{Integer. How many decimals should be returned?}

\item{keep_nas}{Boolean. Keep NAs and count them as well?}
}
\description{
A cross-tabulation function with output similar to STATA.
}
