% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Weighted Cross Tabulation}
\usage{
crosstab(df, ..., prow = FALSE, pcol = FALSE, pall = FALSE,
  decimals = 2, keep_nas = TRUE, total = TRUE, order = TRUE)
}
\arguments{
\item{df}{Data.frame.}

\item{...}{Variables. Dependent and independent variables. If needed,
third value should be the weight variable.}

\item{prow, pcol, pall}{Boolean. Calculate percent values for rows, columns,
or the whole table, respectively.}

\item{decimals}{Integer. How many decimals should be returned?}

\item{keep_nas}{Boolean. Keep NAs and count them as well?}

\item{total}{Boolean. Return total values column?}

\item{order}{Boolean. Sort columns and rows by frequencies? Else, will
be sorted alphabetically.}
}
\description{
A cross-tabulation function with output similar to STATA, tidy
friendly, with weights if needed. If only one dependent variable
crossval() will be recommended; if two, then both variables will be 
crossed; if three, both variables will be crossed with third as 
weighted values.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{df_str}},
  \code{\link{distr}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{missingness}}, \code{\link{textCloud}},
  \code{\link{textTokenizer}}, \code{\link{tree_var}},
  \code{\link{trendsRelated}}
}
\concept{Exploratory}
