% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crossval}
\alias{crossval}
\title{Weighted Frequency (Counter with Ponder)}
\usage{
crossval(df, ..., order = TRUE, compare = FALSE)
}
\arguments{
\item{df}{Data.frame. It must contain a ponder column.}

\item{...}{Variables. Column(s) to calculate the weighted frequencies.}

\item{order}{Boolean. Do you wish to desc sort frequencies?}

\item{compare}{Boolean. Compare with equal-weighted values?}
}
\description{
This function lets the user calculate the weighted frequency in a
tidy friendly way.
}
