% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer}
\usage{
dalex_explainer(df, model, model_name = "lares::h2o_automl", pred = NA)
}
\arguments{
\item{df}{Dataframe}

\item{model}{Model. Use a H2O Model by default. If another, use pred parameter}

\item{model_name}{Character. Name of your model}

\item{pred}{Function. Custom function for calculation predictions}
}
\description{
DALEX library function to create an explainer object
}
\seealso{
Other Interpretability: \code{\link{dalex_local}},
  \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}
}
\concept{Interpretability}
