% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer for H2O}
\usage{
dalex_explainer(df, model, y = "tag", ignore = NA)
}
\arguments{
\item{df}{Dataframe. Must contain all columns and predictions}

\item{model}{Model object (H2O)}

\item{y}{Character. Variable's column name}

\item{ignore}{Character vector. Which columns should be ignored?}
}
\description{
DALEX function to create an explainer object
}
\examples{
\dontrun{
# TRAIN A SIMPLE MODEL
data(dft) # Titanic dataset
dfm <- h2o_automl(dft, 
                  y = "Survived", 
                  ignore = c("Ticket", "PassengerId", "Cabin")), 
                  max_models = 3)

# EXPLAINER
explainer <- dalex_explainer(df = dfm$datasets$test, model = dfm$model)
explainer$data <- na.omit(explainer$data)

# CATEGORICAL EXAMPLE
class <- dalex_variable(explainer, y = "Pclass")
class$plot

# NUMERICAL EXAMPLE
class <- dalex_variable(explainer, y = "Fare")
class$plot
# x <- ceteris_paribus(explainer, explainer$data, variables = "Fare")
# aux <- aggregate_profiles(x)
# plot(aux) + theme_lares2(legend = "none")

# LOCAL EXAMPLE
print(head(explainer$data, 1))
  Pclass    Sex Age SibSp Parch    Fare Embarked train_test predict    FALSE.     TRUE.
       3 female  27     0     2 11.1333        S       test   FALSE 0.6844502 0.3155498
local <- dalex_local(explainer, row = 1, plot = TRUE)
# OR YOU COULD MANUALLY INPUT THE OBSERVATION
local <- dalex_local(explainer, observation = head(explainer$data, 1))
local$plot
}
}
\seealso{
Other Interpretability: 
\code{\link{dalex_local}()},
\code{\link{dalex_residuals}()},
\code{\link{dalex_variable}()}
}
\concept{Interpretability}
