% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer for H2O}
\usage{
dalex_explainer(df, model, y = "tag", ignore = NA)
}
\arguments{
\item{df}{Dataframe. Must contain all columns and predictions}

\item{model}{Model object}

\item{y}{Character. Variable's column name}

\item{ignore}{Character vector. Which columns should be ignored?}
}
\description{
DALEX function to create an explainer object
}
\seealso{
Other Interpretability: \code{\link{dalex_local}},
  \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}
}
\concept{Interpretability}
