% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer}
\usage{
dalex_explainer(df, model, variable = "tag", model_name = NULL,
  pred = NA)
}
\arguments{
\item{df}{Dataframe. Must contain all columns and predictions}

\item{model}{Model object}

\item{model_name}{Character. Name of your model for plots}

\item{pred}{Function. Custom function for calculating predictions.
Default function will be H2O's function to predict}
}
\description{
DALEX library function to create an explainer object
}
\seealso{
Other Interpretability: \code{\link{dalex_local}},
  \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}
}
\concept{Interpretability}
