% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_explainer}
\alias{dalex_explainer}
\title{DALEX Explainer for H2O}
\usage{
dalex_explainer(df, model, y = "tag", ignore = NA)
}
\arguments{
\item{df}{Dataframe. Must contain all columns and predictions}

\item{model}{Model object (H2O)}

\item{y}{Character. Variable's column name}

\item{ignore}{Character vector. Which columns should be ignored?}
}
\description{
DALEX function to create an explainer object
}
\examples{
\donttest{
data(dft) # Titanic dataset

# TRAIN A SIMPLE MODEL
dfm <- h2o_automl(dft, y = "Survived",
                  ignore = c("Ticket", "PassengerId", "Cabin"),
                  max_models = 1)

# EXPLAINER
explainer <- dalex_explainer(df = dfm$datasets$test, model = dfm$model, y = "Survived")
explainer$data <- na.omit(explainer$data)

# CATEGORICAL EXAMPLE
class <- dalex_variable(explainer, vars = c("Pclass", "Sex"))
class$plot

# NUMERICAL EXAMPLE
num <- dalex_variable(explainer, vars = c("Fare", "Age"))
num$plot

# LOCAL EXAMPLE
print(head(explainer$data, 1))
local <- dalex_local(explainer, row = 1)
# OR YOU COULD MANUALLY INPUT THE OBSERVATION
local <- dalex_local(explainer, observation = head(explainer$data, 1))
local$plot
}
}
\seealso{
Other Interpretability: 
\code{\link{dalex_local}()},
\code{\link{dalex_residuals}()},
\code{\link{dalex_variable}()}
}
\concept{Interpretability}
