% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_local}
\alias{dalex_local}
\title{DALEX Local}
\usage{
dalex_local(explainer, row, plot = TRUE, print = TRUE)
}
\arguments{
\item{explainer}{Object. Result from dalex_explainer function}

\item{row}{Dataframe. An observation which you want to study}

\item{plot}{Boolean. Do you wish to see the results plot?}

\item{print}{Boolean. Do you wish to see the results table?}
}
\description{
DALEX library function for local interpretations
}
\seealso{
Other Interpretability: \code{\link{dalex_explainer}},
  \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}
}
\concept{Interpretability}
