% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_local}
\alias{dalex_local}
\title{DALEX Local}
\usage{
dalex_local(explainer, observation = NA, row = 1, plot = TRUE,
  print = TRUE, alarm = TRUE)
}
\arguments{
\item{explainer}{Object. Result from dalex_explainer function}

\item{observation}{Data.frame. If you want to use an observation
that was not in the original explainer function, add here. Else, use row}

\item{row}{Dataframe. Row number from the data.frame used in explainer}

\item{plot}{Boolean. Do you wish to see the results plot?}

\item{print}{Boolean. Do you wish to see the results table?}

\item{alarm}{Boolean. Ping an alarm when ready! Needs beepr installed}
}
\description{
DALEX function for local interpretations
}
\seealso{
Other Interpretability: \code{\link{dalex_explainer}},
  \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}
}
\concept{Interpretability}
