% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_residuals}
\alias{dalex_residuals}
\title{DALEX Residuals}
\usage{
dalex_residuals(explainer)
}
\arguments{
\item{explainer}{Object. Result from dalex_explainer function}
}
\description{
DALEX function for residuals
}
\seealso{
Other Interpretability: \code{\link{dalex_explainer}},
  \code{\link{dalex_local}}, \code{\link{dalex_variable}}
}
\concept{Interpretability}
