% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_variable}
\alias{dalex_variable}
\title{DALEX Partial Dependency Plots (PDP)}
\usage{
dalex_variable(explainer, variable, force_class = NA)
}
\description{
DALEX library function for creating Partial Dependency Plots and study
variable's responses vs independent vector.
}
