% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_explainers.R
\name{dalex_variable}
\alias{dalex_variable}
\title{DALEX Partial Dependency Plots (PDP)}
\usage{
dalex_variable(explainer, variable, force_class = NA, alarm = TRUE)
}
\arguments{
\item{explainer}{Object. Result from dalex_explainer function}

\item{variable}{Character. Which character do you wish to study?}

\item{force_class}{Character. If you wish to force a class on your 
variable, which one do you need?}

\item{alarm}{Boolean. Ping an alarm when ready! Needs beepr installed}
}
\description{
DALEX function for creating Partial Dependency Plots and study
variable's responses vs independent vector.
}
\seealso{
Other Interpretability: \code{\link{dalex_explainer}},
  \code{\link{dalex_local}}, \code{\link{dalex_residuals}}
}
\concept{Interpretability}
