% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{date_feats}
\alias{date_feats}
\title{One Hot Encoding for Date/Time Variables (Dummy Variables)}
\usage{
date_feats(dates, keep_originals = FALSE, only = NA, features = TRUE,
  holidays = FALSE, country = "Colombia", currency_pair = NA,
  summary = TRUE)
}
\arguments{
\item{dates}{Vector or dataframe. Non-date/time columns will be 
automatically ignored.}

\item{keep_originals}{Boolean. Should the original date/time columns be
kept in the results?}

\item{only}{Character or vector. Which columns do you wish to process? If
non are explicitly defined, all will be processed}

\item{features}{Create features out of date/time columns?}

\item{holidays}{Boolean. Include holidays as new columns?}

\item{country}{Character or vector. For which countries should the holidays
be included?}

\item{currency_pair}{Character. Which currency exchange do you
wish to get the history from? i.e, USD/COP, EUR/USD...}

\item{summary}{Boolean. Print a summary of the operations?}
}
\description{
This function lets the user automatically create new columns out
of a dataframe or vector with date/time variables.
}
