% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{db_download}
\alias{db_download}
\title{Download Dropbox File by File's Name}
\usage{
db_download(filename, xlsx = TRUE, token_dir = NA,
  token_name = "token_pers.rds")
}
\arguments{
\item{filename}{String. File's name}

\item{xlsx}{Boolean. Is it an Excel file? Can be returned as a list
for each tab and not as a file if needed}

\item{token_dir}{Character. RDS with token local directory}

\item{token_name}{Character. RDS file name with your token's data}
}
\description{
This function lets the user download a file from Dropbox, specifying
its name and using a previously created token.
}
