% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{db_download}
\alias{db_download}
\title{Download Dropbox File by File's Name}
\usage{
db_download(filename, xlsx = TRUE, token_dir = NA,
  token_name = "token_pers.rds")
}
\arguments{
\item{filename}{String. File's name}

\item{xlsx}{Boolean. Is it an Excel file? Can be returned as a list
for each tab and not as a file if needed}

\item{token_dir}{Character. RDS with token local directory}

\item{token_name}{Character. RDS file name with your token's data}
}
\description{
This function lets the user download a file from Dropbox, specifying
its name and using a previously created token.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{iter_seeds}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{mailSend}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{updateLares}},
  \code{\link{zerovar}}

Other Connections: \code{\link{bring_api}},
  \code{\link{db_upload}}, \code{\link{f1_contacts}},
  \code{\link{geoAddress}}, \code{\link{li_auth}},
  \code{\link{queryDW}}, \code{\link{queryDummy}},
  \code{\link{queryProduc}},
  \code{\link{typeform_download}}
}
\concept{Connections}
\concept{Tools}
