% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{db_upload}
\alias{db_upload}
\title{Download Dropbox File by File's Name}
\usage{
db_upload(filename, dir, delete_file = FALSE, token_dir = NA)
}
\arguments{
\item{filename}{String. File's name}

\item{dir}{String. Directory you wish to upload the file to}

\item{delete_file}{Boolean. Delete local file after uploading?}

\item{token_dir}{Character. Credential's local directory}
}
\description{
This function lets the user download a file from Dropbox, specifying
its name and using a previously created token.
}
