% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{db_upload}
\alias{db_upload}
\title{Upload Dropbox File}
\usage{
db_upload(filename, dir, delete_file = FALSE, token_dir = NA,
  token_name = "token_pers.rds")
}
\arguments{
\item{filename}{String. File's name}

\item{dir}{String. Directory you wish to upload the file to}

\item{delete_file}{Boolean. Delete local file after uploading?}

\item{token_dir}{Character. Credential's local directory}

\item{token_name}{Character. RDS file name with your token's data}
}
\description{
This function lets the user download a file from Dropbox, specifying
its name and using a previously created token.
}
