% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{df_str}
\alias{df_str}
\title{Dataset columns and rows structure}
\usage{
df_str(df, return = "plot", plot = FALSE, subtitle = NA,
  quiet = FALSE)
}
\arguments{
\item{df}{Dataframe}

\item{return}{Character. Return "skimr" for skim results, "numbers" for
numbers, or "names" for column names of each of the cateogries}

\item{plot}{Boolean. Do you wish to see a plot?}

\item{subtitle}{Character. Add subtitle to plot}

\item{quiet}{Boolean. Silence the messages?}
}
\description{
This function lets the user to check quickly the structure of a
dataset (data.frame). It returns multiple counters for useful metrics,
a plot, and a list of column names for each of the column metrics.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{distr}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{missingness}}, \code{\link{textCloud}},
  \code{\link{textTokenizer}}, \code{\link{trendsRelated}}
}
\concept{Exploratory}
