% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_info.R
\name{df_str}
\alias{df_str}
\title{Dataset columns and rows structure}
\usage{
df_str(df, return = "numbers", plot = TRUE, subtitle = "")
}
\arguments{
\item{df}{Dataframe}

\item{return}{Character. Return "numbers" for counters and percetnages or 
"names" for column names of each of the cateogries}

\item{plot}{Boolean. Do you wish to see a plot?}

\item{subtitle}{Character. Add subtitle to plot}
}
\description{
This function lets the user to check quickly the structure of a
dataset (data.frame). It returns multiple counters for useful metrics,
a plot, and a list of column names for each of the column metrics.
}
