% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{df_str}
\alias{df_str}
\title{Dataset columns and rows structure}
\usage{
df_str(df, return = "skimr", plot = TRUE, subtitle = "")
}
\arguments{
\item{df}{Dataframe}

\item{return}{Character. Return "skimr" for skim results, "numbers" for
numbers, or "names" for column names of each of the cateogries}

\item{plot}{Boolean. Do you wish to see a plot?}

\item{subtitle}{Character. Add subtitle to plot}
}
\description{
This function lets the user to check quickly the structure of a
dataset (data.frame). It returns multiple counters for useful metrics,
a plot, and a list of column names for each of the column metrics.
}
\seealso{
Other Exploratory: \code{\link{corr_var}},
  \code{\link{crosstab}}, \code{\link{dalex_explainer}},
  \code{\link{dalex_local}}, \code{\link{dalex_residuals}},
  \code{\link{dalex_variable}}, \code{\link{distr}},
  \code{\link{freqs}}
}
\concept{Exploratory}
