% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{df_str}
\alias{df_str}
\title{Dataset columns and rows structure}
\usage{
df_str(df, return = "plot", subtitle = NA, quiet = FALSE)
}
\arguments{
\item{df}{Dataframe}

\item{return}{Character. Return "skimr" for skim report, "numbers" for
stats and numbers, "names" for a list with the column names of each of 
the class types, "plot" for a nice plot with "numbers" output, "summary"
for an overall summary plot showing categorical, numeric, and missing values}

\item{subtitle}{Character. Add subtitle to plot}

\item{quiet}{Boolean. Keep quiet or show other options available?}
}
\description{
This function lets the user to check quickly the structure of a
dataset (data.frame). It returns multiple counters for useful metrics,
a plot, and a list of column names for each of the column metrics.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{distr}}, \code{\link{freqs_df}},
  \code{\link{freqs}}, \code{\link{gain_lift}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{plot_cats}}, \code{\link{plot_df}},
  \code{\link{plot_nums}}, \code{\link{summer}},
  \code{\link{tree_var}}, \code{\link{trendsRelated}}
}
\concept{Exploratory}
