% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dfr}
\alias{dfr}
\title{Dataset: Results for AutoML Predictions}
\format{
An object of class \code{"list"} with 3 \code{"data.frame"}
\describe{
  \item{class2}{Predictions for a Binomial Classification Model}
  \item{class3}{Predictions for a Multi-Categorical Classification Model}
  \item{regr}{Predictions for a Continuous Regression Model}
}
}
\usage{
data(dfr)
}
\description{
List with categorical (2 and 3 classes) and continuous predictions, 
generated with \code{h2o_automl()} and the \code{dft} 
Titanic dataset. Data.frames contain the actual values and model's 
predictions over the test set observations.
}
\examples{
data(dfr)
lapply(dfr, head)
}
\seealso{
Other Dataset: 
\code{\link{dfl}},
\code{\link{dft}}
}
\concept{Dataset}
\keyword{datasets}
