% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/datasets_titanic.R
\docType{data}
\name{dft}
\alias{dft}
\title{Dataset: Titanic Sub-dataset por Examples}
\format{An object of class \code{"data.frame"}}
\usage{
data(dft)

data(dft)
}
\description{
Data generated from a Titanic dataset.

Data generated from a Titanic dataset.
}
\examples{
data(dft)
data(dft)
}
\keyword{datasets}
\keyword{titanic}
