% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distr}
\alias{distr}
\title{Compare Variables with their Distributions}
\usage{
distr(
  data,
  ...,
  type = 1,
  ref = TRUE,
  note = NA,
  top = 10,
  breaks = 10,
  na.rm = FALSE,
  force = "none",
  trim = 0,
  clean = FALSE,
  abc = FALSE,
  custom_colours = FALSE,
  plot = TRUE,
  chords = FALSE,
  save = FALSE,
  subdir = NA
)
}
\arguments{
\item{data}{Dataframe}

\item{...}{Variables. Main (target variable) and secondary (values 
variable) to group by}

\item{type}{Integer. 1 for both plots, 2 for counter plot only, 3 por 
percentages plot only.}

\item{ref}{Boolean. Show a reference line if levels = 2? Quite useful 
when data is unbalanced (not 50/50) because a reference line is drawn}

\item{note}{Character. Caption for the plot}

\item{top}{Integer. Filter and plot the most n frequent for categorical values}

\item{breaks}{Integer. Number of splits for numerical values}

\item{na.rm}{Boolean. Ignore NAs if needed}

\item{force}{Character. Force class on the values data. Choose between 'none',
'character', 'numeric', 'date'}

\item{trim}{Integer. Trim labels until the nth character for categorical values 
(applies for both, target and values)}

\item{clean}{Boolean. Use lares::cleanText for categorical values (applies 
for both, target and values)}

\item{abc}{Boolean. Do you wish to sort by alphabetical order?}

\item{custom_colours}{Boolean. Use custom colours function?}

\item{plot}{Boolean. Return a plot? Otherwise, a table with results}

\item{chords}{Boolean. Use a chords plot?}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}
}
\description{
Compare the distribution of a target variable vs another variable. This 
function automatically splits into quantiles for numerical variables.
Custom and tidyverse friendly.
}
\examples{
\dontrun{
library(dplyr)
data(dft)
# Relation for survived vs sex
dft \%>\% distr(Survived, Sex)
# Relation for sex vs survived
dft \%>\% distr(Sex, Survived)
# Relation for survived vs embark gate (categorical)
dft \%>\% distr(Survived, Embarked)
dft \%>\% distr(Embarked, Survived)
# Relation for survived vs embark gate
dft \%>\% distr(Survived, Fare)
# Relation for survived vs fare with custom colours for Survived
dft \%>\% distr(Survived, Fare, custom_colours = TRUE)
# Relation for survived vs fare with ascending order
dft \%>\% distr(Survived, Fare, abc = TRUE)
# Relation for survived vs fare with only 5 splits
dft \%>\% distr(Survived, Fare, abc = TRUE, breaks = 5)
dft \%>\% distr(Survived, Fare, abc = TRUE, top = 5)
dft \%>\% distr(Survived, Embarked, top = 2)
# Relation for survived vs age (notice NA values)
dft \%>\% distr(Survived, Age)
dft \%>\% distr(Survived, Age, na.rm = TRUE, abc = TRUE)
dft \%>\% distr(Survived, Ticket, top = 10)
# Distribution of fares payed
dft \%>\% distr(Fare)
dft \%>\% filter(Fare < 200) \%>\% distr(Fare)
# Distribution (frequency) of survivors
dft \%>\% distr(Survived, force = "char")
# Distribution of log(Fare) vs Fare
dft \%>\% mutate(logFare = log(Fare)) \%>\% distr(Fare, logFare)
dft \%>\% mutate(logFare = log(Fare)) \%>\% 
  filter(Fare < 100) \%>\% distr(Fare, logFare) + 
  geom_point(colour = "yellow")
# Only one plot or the other
dft \%>\% distr(Survived, Age, type = 2)
dft \%>\% distr(Survived, Age, type = 3)
}
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gain_lift}()},
\code{\link{get_tweets}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{summer}()},
\code{\link{tree_var}()},
\code{\link{trendsRelated}()}

Other Visualization: 
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{mplot_conf}()},
\code{\link{mplot_cuts_error}()},
\code{\link{mplot_cuts}()},
\code{\link{mplot_density}()},
\code{\link{mplot_full}()},
\code{\link{mplot_gain}()},
\code{\link{mplot_importance}()},
\code{\link{mplot_lineal}()},
\code{\link{mplot_metrics}()},
\code{\link{mplot_response}()},
\code{\link{mplot_roc}()},
\code{\link{mplot_splits}()},
\code{\link{noPlot}()},
\code{\link{plot_survey}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Exploratory}
\concept{Visualization}
