% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{errors}
\alias{errors}
\title{Calculate Errors}
\usage{
errors(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate all errors and R squared 
simultaneously.
}
\examples{
df <- data.frame(tag = c(0, 0, 0, 1, 1, 1),
                 score = c(0.1, 0.2, -.1, 0.7, 0.9, 1))
errors(df$tag, df$score)
}
\seealso{
Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{loglossBinary}()},
\code{\link{model_metrics}()},
\code{\link{quants}()},
\code{\link{rmse}()}
}
\concept{Calculus}
