% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{errors}
\alias{errors}
\title{Calculate Errors}
\usage{
errors(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate all errors and R squared 
simultaneously.
}
\seealso{
Other Calculus: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{corr}()},
\code{\link{deg2num}()},
\code{\link{dist2d}()},
\code{\link{loglossBinary}()},
\code{\link{mae}()},
\code{\link{mape}()},
\code{\link{model_metrics}()},
\code{\link{mse}()},
\code{\link{quants}()},
\code{\link{rmse}()},
\code{\link{rsqa}()},
\code{\link{rsq}()}
}
\concept{Calculus}
