% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{etf_sector}
\alias{etf_sector}
\title{ETF's Sectors Breakdown}
\usage{
etf_sector(etf = "VTI", verbose = TRUE)
}
\arguments{
\item{etf}{Character Vector. Which ETFs you wish to scrap?}

\item{verbose}{Boolean. Print results and progress while downloading?}
}
\description{
This function scraps etf.com data for sector breakdown on ETFs.
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
