% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{etf_sector}
\alias{etf_sector}
\title{ETF's Sectors Breakdown}
\usage{
etf_sector(etf = "VTI")
}
\arguments{
\item{etf}{Character Vector. Which ETF(s) you wish to scrap?}
}
\description{
This function scraps etf.com to get an ETF's sector breakdown.
}
