% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{etf_sector}
\alias{etf_sector}
\title{ETF's Sectors Breakdown}
\usage{
etf_sector(etf = "VTI", verbose = TRUE)
}
\arguments{
\item{etf}{Character Vector. Which ETFs you wish to scrap?}

\item{verbose}{Boolean. Print results and progress while downloading?}
}
\description{
This function scraps etf.com data for sector breakdown on ETFs.
Use splot_etf() for visualization.
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}
}
\concept{Investment}
