% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{etf_sector}
\alias{etf_sector}
\title{ETF's Sectors Breakdown}
\usage{
etf_sector(etf = "VTI", quiet = FALSE, cache = TRUE)
}
\arguments{
\item{etf}{Character Vector. Which ETFs you wish to scrap?}

\item{quiet}{Boolean. Keep quiet? If not, informative messages will be shown.}

\item{cache}{Boolean. Use daily cache if available?}
}
\value{
data.frame with ETF break.down data by sector
}
\description{
\code{etf_sector()} scraps etf.com data for sector breakdown on ETFs.
Use \code{splot_etf()} for visualization.
}
\examples{
\donttest{
etf_sector(etf = "VTI")
}
}
\seealso{
Other Investment: 
\code{\link{splot_summary}()},
\code{\link{stocks_file}()},
\code{\link{stocks_quote}()}
}
\concept{Investment}
