% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{export_plot}
\alias{export_plot}
\title{Install latest version of H2O}
\usage{
export_plot(p, name = "plot", vars = NA, sep = ".vs.", width = 8,
  height = 6, subdir = NA, quiet = FALSE)
}
\arguments{
\item{p}{ggplot2 object. Plot to export}

\item{name}{Character. File's name or sufix if vars is not null}

\item{vars}{Vector. Variables in plot}

\item{sep}{Character. Separator for variables}

\item{width}{Numeric. Plot's width on file}

\item{height}{Numeric. Plot's height on file}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}

\item{quiet}{Boolean. Display succesful message with filename when saved?}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
