% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{export_plot}
\alias{export_plot}
\title{Install latest version of H2O}
\usage{
export_plot(p, name = "plot", vars = NA, sep = ".vs.", width = 8,
  height = 6, subdir = NA, quiet = FALSE)
}
\arguments{
\item{p}{ggplot2 object. Plot to export}

\item{name}{Character. File's name or sufix if vars is not null}

\item{vars}{Vector. Variables in plot}

\item{sep}{Character. Separator for variables}

\item{width}{Numeric. Plot's width on file}

\item{height}{Numeric. Plot's height on file}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}

\item{quiet}{Boolean. Display succesful message with filename when saved?}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{statusbar}},
  \code{\link{zerovar}}
}
\concept{Tools}
