% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{export_plot}
\alias{export_plot}
\title{Install latest version of H2O}
\usage{
export_plot(
  p,
  name = "plot",
  vars = NA,
  sep = ".vs.",
  width = 8,
  height = 6,
  res = 300,
  dir = getwd(),
  subdir = NA,
  quiet = FALSE
)
}
\arguments{
\item{p}{ggplot2 or gridExtra object. Plot to export}

\item{name}{Character. File's name or sufix if vars is not null}

\item{vars}{Vector. Variables in plot}

\item{sep}{Character. Separator for variables}

\item{width, height, res}{Numeric. Plot's width, height, and res (for grids)}

\item{dir}{Character. In which directory do you wish to save 
the results? Working directory as default.}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}

\item{quiet}{Boolean. Display succesful message with filename when saved?}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
\examples{
\dontrun{
p <- noPlot()
export_plot(p, name = "noplot", width = 10, height = 8, res = 300)
export_plot(p, name = "noplot2", dir = "~/Desktop", subdir = "newplots")
}
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{pass}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}
}
\concept{Tools}
