% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{export_results}
\alias{export_results}
\title{Export h2o_automl's Results}
\usage{
export_results(results, thresh = 10, txt = TRUE, rds = TRUE,
  binary = TRUE, pojo = TRUE, mojo = TRUE, sample_size = 10,
  subdir = NA, save = TRUE)
}
\arguments{
\item{results}{Object. h2o_automl output}

\item{thresh}{Integer. Threshold for selecting binary or regression 
models: this number is the threshold of unique values we should 
have in 'tag' (more than: regression; less than: classification)}

\item{txt}{Boolean. Do you wish to export the txt results?}

\item{rds}{Boolean. Do you wish to export the RDS results?}

\item{binary}{Boolean. Do you wish to export the Binary model?}

\item{pojo}{Boolean. Do you wish to export the POJO model?}

\item{mojo}{Boolean. Do you wish to export the MOJO model?}

\item{sample_size}{Integer. How many example rows do you want to show?}

\item{subdir}{Character. In which directory do you wish to save the results?}

\item{save}{Boolean. Do you wish to save/export results?}
}
\description{
Export RDS, TXT, POJO, MOJO and all results from h2o_automl
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{clusterKmeans}}, \code{\link{conf_mat}},
  \code{\link{gain_lift}}, \code{\link{h2o_automl}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{impute}},
  \code{\link{iter_seeds}}, \code{\link{model_metrics}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{msplit}}

Other Tools: \code{\link{autoline}},
  \code{\link{bindfiles}}, \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{iter_seeds}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{mailSend}}, \code{\link{matrixwd}},
  \code{\link{msplit}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{statusbar}},
  \code{\link{try_require}}, \code{\link{updateLares}},
  \code{\link{zerovar}}
}
\concept{Machine Learning}
\concept{Tools}
