% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_f1.R
\name{f1_contacts}
\alias{f1_contacts}
\title{Hubspot contacts (Somos F1)}
\usage{
f1_contacts(creds = NA, limit = 500)
}
\arguments{
\item{creds}{Character. Credential's user (see get_credentials)}

\item{limit}{Integer. Limit the amount of contacts to return. 
Multiples of 100 or round up will occur.}
}
\description{
This function brings all contacts from Somos F1's Hubspot account. 
It also transforms the API's JSON result data into a relational table
or data.frame. Keep in mind that it will always bring the latest
value for each parameter (column).
}
