% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot_f1.R
\name{f1_contacts}
\alias{f1_contacts}
\title{Hubspot contacts (Somos F1)}
\usage{
f1_contacts(limit = 1000, creds = NA)
}
\arguments{
\item{limit}{Integer. Limit the amount of contacts to return. 
Multiples of 100 or round up will occur.}

\item{creds}{Character. Credential's user (see \code{get_creds()})}
}
\description{
This function brings all contacts from Somos F1's Hubspot account. 
It also transforms the API's JSON result data into a relational table
or data.frame. Keep in mind that it will always bring the latest
value for each parameter (column).
}
\seealso{
Other Credentials: 
\code{\link{db_download}()},
\code{\link{geoAddress}()},
\code{\link{get_credentials}()},
\code{\link{get_tweets}()},
\code{\link{mailSend}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()},
\code{\link{stocks_file}()},
\code{\link{stocks_report}()}
}
\concept{Credentials}
\concept{Hubspot}
