% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_accounts}
\alias{fb_accounts}
\title{Facebook Ad Accounts}
\usage{
fb_accounts(
  token,
  business_id = "904189322962915",
  type = c("owned", "client"),
  limit = 10000,
  api_version = "v3.3"
)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{business_id}{Character. Business ID}

\item{type}{Character vector. Values: owned, client}

\item{limit}{Integer. Query limit}

\item{api_version}{Character. Facebook API version}
}
\description{
This returns all ad accounts for a FB Business Account FB.
For more information on Ad Insights' API, go to the 
\href{https://developers.facebook.com/docs/marketing-api/reference/ad-account/insights/}{original documentaion}
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
