% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_ads}
\alias{fb_ads}
\title{Facebook Ads API}
\usage{
fb_ads(
  token,
  which,
  start = Sys.Date() - 31,
  end = Sys.Date(),
  fields = NA,
  api_version = "v8.0",
  process = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{which}{Character vector. This is the accounts, campaigns, adsets, 
or ads IDs to be queried. Remember: if report_level = "account", you must 
start the ID with `act_`.}

\item{start}{Character. The first full day to report, in the 
format "YYYY-MM-DD"}

\item{end}{Character. The last full day to report, in the 
format "YYYY-MM-DD"}

\item{fields}{Character, json format. Leave `NA` for default fields.}

\item{api_version}{Character. Facebook API version}

\item{process}{Boolean. Process GET results to a more friendly format?}

\item{flatten}{Boolean. Flatten lists within lists?}
}
\description{
This returns all available FB ads for any account, campaign, or ad set id.
For more information on Ad' API, go to the 
\href{https://developers.facebook.com/docs/marketing-api/reference/adgroup}{original documentaion}
}
\details{
This function was based on FBinsightsR.
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
