% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_creatives}
\alias{fb_creatives}
\title{Facebook Creatives API}
\usage{
fb_creatives(
  token,
  which,
  api_version = "v8.0",
  process = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{which}{Character vector. This is the accounts, campaigns, adsets, 
or ads IDs to be queried. Remember: if report_level = "account", you must 
start the ID with `act_`.}

\item{api_version}{Character. Facebook API version}

\item{process}{Boolean. Process GET results to a more friendly format?}

\item{flatten}{Boolean. Flatten lists within lists?}
}
\description{
For more information: \href{https://developers.facebook.com/docs/marketing-api/reference/ad-creative/}{Ad Creative}
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
