% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_insights}
\alias{fb_insights}
\title{Facebook Insights API}
\usage{
fb_insights(
  token,
  which,
  start = Sys.Date() - 7,
  end = Sys.Date(),
  time_increment = "1",
  report_level = "campaign",
  ad_object = "insights",
  breakdowns = NA,
  fields = NA,
  limit = 10000,
  api_version = "v8.0",
  process = TRUE
)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{which}{Character vector. This is the accounts, campaigns, adsets, 
or ads IDs to be queried. Remember: if report_level = "account", you must 
start the ID with `act_`.}

\item{start}{Character. The first full day to report, in the 
format "YYYY-MM-DD"}

\item{end}{Character. The last full day to report, in the 
format "YYYY-MM-DD"}

\item{time_increment}{Character. Group by months ("monthly"), 
everything together ("all_days") or an integer per days [1-90].
Default: each day separatly (i.e. "1").}

\item{report_level}{Character. One of "ad", "adset", "campaign", or "account"}

\item{ad_object}{Character. One of: "insights" (default), "adsets", ...}

\item{breakdowns}{Character Vector. One or more of breakdowns for 
segmentation results. Set to NA for no breakdowns}

\item{fields}{Character, json format. Leave `NA` for default fields.}

\item{limit}{Integer. Query limit}

\item{api_version}{Character. Facebook API version}

\item{process}{Boolean. Process GET results to a more friendly format?}
}
\description{
This returns all available FB insights per day including any given
breakdown to the specified report level, and place into a data frame.
For more information on Ad Insights' API, go to the 
\href{https://developers.facebook.com/docs/marketing-api/insights}{original documentaion}.
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
