% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_insights}
\alias{fb_insights}
\title{Facebook Insights API}
\usage{
fb_insights(token, which, start = Sys.Date() - 7, end = Sys.Date(),
  time_increment = "1", report_level = "campaign", breakdowns = NA,
  api_version = "v3.3")
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{which}{Character vector. This is the accounts, campaigns, adsets, 
or ads IDs to be queried}

\item{start}{Character. The first full day to report, in the 
format "YYYY-MM-DD"}

\item{end}{Character. The last full day to report, in the 
format "YYYY-MM-DD"}

\item{time_increment}{Character. Agrupación por meses (monthly), 
para todo junto (all_days) o un entero para cantidad de días [1-90].
Por defecto, traerá cada día por separado (i.e.=1)}

\item{report_level}{Character. One of ad, adset, campaign, or account}

\item{breakdowns}{Character Vector. One or more of breakdowns for 
segmentation results. Set to NA for no breakdowns}

\item{api_version}{Character. Facebook API version}
}
\description{
This returns all available FB insights per day including any given
breakdown to the specified report level, and place into a data frame.
For more information on Ad Insights' API, go to the 
\href{https://developers.facebook.com/docs/marketing-api/insights}{original documentaion}
}
\details{
This function was based on FBinsightsR.
}
\seealso{
Other Scrapper: \code{\link{fb_accounts}},
  \code{\link{fb_ads}}, \code{\link{fb_posts}},
  \code{\link{fb_post}}, \code{\link{geoStratum}},
  \code{\link{holidays}}, \code{\link{ip_country}},
  \code{\link{trendsRelated}}

Other API: \code{\link{bring_api}},
  \code{\link{fb_accounts}}, \code{\link{fb_ads}},
  \code{\link{fb_posts}}, \code{\link{fb_post}}

Other Facebook: \code{\link{fb_accounts}},
  \code{\link{fb_ads}}, \code{\link{fb_posts}},
  \code{\link{fb_post}}
}
\concept{API}
\concept{Facebook}
\concept{Scrapper}
