% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_posts}
\alias{fb_posts}
\title{####################################################################
#' Facebook Marketing API
#' 
#' For more information on Ad Insights' API, go to the 
#' \href{https://developers.facebook.com/docs/marketing-api/insights}{original documentaion}
#' 
#' @family API
#' @family Facebook
#' @inheritParams fb_process
#' @inheritParams fb_insights
#' @export
fb_marketing <- function(token,
                         which,
                         start = Sys.Date() - 7, 
                         end = Sys.Date(), 
                         ad_object = "adsets",
                         fields = NA,
                         api_version = "v8.0",
                         process = TRUE){
  
  set_config(config(http_version = 0))
  
  which <- v2t(which, quotes = FALSE)
  if (is.na(fields[1]))
    fields <- paste(
      "name,configured_status,effective_status,bid_strategy,",
      "billing_event,targeting")
  
  URL <- glued("https://graph.facebook.com/{api_version}/{which}/{ad_object}")
  
  # Call insights
  import <- GET(
    URL,
    query = list(
      access_token = token,
      time_range = paste0('{\"since\":\"',start,'\",\"until\":\"',end,'\"}'),
      fields = if (length(fields) > 1) 
        vector2text(fields, sep = ",", quotes = FALSE) else fields
    ),
    encode = "json")
  
  if (!process) return(import)
  output <- fb_process(import)
  return(as_tibble(output))
}
Get Facebook's Page Posts (API Graph)}
\usage{
fb_posts(
  token,
  total = 150,
  posts_limit = 100,
  comments = FALSE,
  shares = FALSE,
  reactions = FALSE
)
}
\arguments{
\item{token}{Character. This must be a valid access token with sufficient 
privileges. Visit the Facebook API Graph Explorer to acquire one}

\item{total}{Integer. How many most recent posts do you need?}

\item{posts_limit}{Integer. For each post, hoy many results do you need?}

\item{comments, shares, reactions}{Boolean. Include in your query?}
}
\description{
Connect to an API Graph's token of a given page and get posts, 
comments, shares, and reactions of n posts (with no limits).
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_post}()},
\code{\link{fb_process}()}
}
\concept{API}
\concept{Facebook}
