% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook.R
\name{fb_process}
\alias{fb_process}
\title{Process Facebook's API Objects}
\usage{
fb_process(response, paginate = TRUE)
}
\arguments{
\item{response}{GET's output object, class response}

\item{paginate}{Boolean. Run through all paginations? If not,
only the first one will be processed.}
}
\value{
data.frame with un-nested processed results or NULL if no results found.
}
\description{
Process and paginate raw results from Facebook's API, result of
querying the API with \code{httr::GET}.
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_rf}()},
\code{\link{fb_token}()},
\code{\link{li_auth}()},
\code{\link{li_profile}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other Facebook: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{fb_rf}()},
\code{\link{fb_token}()}
}
\concept{API}
\concept{Facebook}
