% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{file_name}
\alias{file_name}
\title{Get file names without extensions}
\usage{
file_name(filepath)
}
\arguments{
\item{filepath}{Character vector. File path(s) to get file names with.
no extension.}
}
\description{
Get file names without extensions
}
\examples{
file_name("file.aux")
file_name("temp/file.R")
file_name("/temp/temp3/music.mp3")
}
