% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{filesGD}
\alias{filesGD}
\alias{writeGS4}
\title{Google Drive Files (API v4)}
\usage{
filesGD(title, server = FALSE, json = NULL, api_key = NULL, email = NULL)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{server}{Boolean. Force interacting auth process?}

\item{json}{Character. JSON filename with service auth}

\item{email}{Character. If you have multiple pre-authorized
accounts in your machine, you may non-interactively select
which one you wish to use.}

\item{data}{Object (value, vector, dataframe, list)}

\item{reformat}{Boolean. Reformat the affected cells?}
}
\description{
Authenticate and find Google Drive files and IDs by name.
}
\seealso{
Other Scrapper: 
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()},
\code{\link{writeGS}()}

Other Google: 
\code{\link{geoAddress}()},
\code{\link{queryGA}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()},
\code{\link{trendsTime}()},
\code{\link{writeGS}()}
}
\concept{Google}
\concept{Scrapper}
