% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{font_exists}
\alias{font_exists}
\title{Check if Font is Installed}
\usage{
font_exists(font = "Arial Narrow")
}
\arguments{
\item{font}{Character. Which font to check}
}
\description{
This function check if specific font is installed
}
