% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{font_exists}
\alias{font_exists}
\title{Check if Font is Installed}
\usage{
font_exists(font = "Arial Narrow", quiet = FALSE)
}
\arguments{
\item{font}{Character. Which font to check}

\item{quiet}{Boolean. Keep quiet? If not, show message}
}
\value{
Boolean result of the existing fonts check.
}
\description{
This function checks if a font is installed in your machine.
}
\examples{
font_exists(font = "Arial")
font_exists(font = "XOXO")
font_exists(font = "")
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bind_files}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{files_functions}()},
\code{\link{formatColoured}()},
\code{\link{formatHTML}()},
\code{\link{get_credentials}()},
\code{\link{glued}()},
\code{\link{grepm}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{list_cats}()},
\code{\link{listfiles}()},
\code{\link{mail_send}()},
\code{\link{move_files}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{warnifnot}()},
\code{\link{what_size}()}
}
\concept{Tools}
