% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{font_exists}
\alias{font_exists}
\title{Check if Font is Installed}
\usage{
font_exists(font = "Arial Narrow", quiet = FALSE)
}
\arguments{
\item{font}{Character. Which font to check}

\item{quiet}{Boolean. Keep quiet? If not, show message}
}
\value{
Boolean result of the existing fonts check.
}
\description{
This function checks if a font is installed in your machine.
}
\examples{
font_exists(font = "Arial")
font_exists(font = "XOXO")
font_exists(font = "")
}
