% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast_arima}
\alias{forecast_arima}
\title{ARIMA Forecast}
\usage{
forecast_arima(time, values, n_future = 30, ARMA = 8,
  wd_excluded = NA, plot = TRUE, plot_days = 90, project = NA)
}
\arguments{
\item{time}{POSIX. Vector with date values}

\item{values}{Numeric. Vector with numerical values}

\item{n_future}{Integer. How many steps do you wish to forecast?}

\item{ARMA}{Integer. How many days should the model look backfor ARMA?
Between 7 and 10 days recommmended}

\item{wd_excluded}{Character vector. Which weekdays are excluded in 
your training set. If there are, please define know which ones. Example:
c('Sunday','Thursday'). If set to 'auto' then it will detect automatically
which weekdays have no data and forcast without these days.}

\item{plot}{Boolean. If you wish to plot your results}

\item{plot_days}{Integer. How many days back you wish to plot?}

\item{project}{Character. Name of your forecast project}
}
\description{
This function automates the ARIMA iterations and modeling for 
time forecasting. For the moment, units can only be days.
}
