% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{formatNum}
\alias{formatNum}
\title{Nicely Format Numerical Values}
\usage{
formatNum(x, decimals = 2, type = 1)
}
\arguments{
\item{x}{Numerical Vector}

\item{decimals}{Integer. Amount of decimals to display}

\item{type}{Integer. 1 for International standards. 2 for American Standards.}
}
\description{
This function lets the user format numerical values nicely
}
