% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{freqs}
\alias{freqs}
\title{Frequencies Calculations and Plot}
\usage{
freqs(vector, ..., plot = FALSE, rm.na = FALSE, save = FALSE,
  subdir = NA)
}
\arguments{
\item{vector}{Data.frame}

\item{...}{Variables. Variables you wish to process. Order matters.}

\item{plot}{Boolean. Do you want to see a plot? Three variables tops.}

\item{rm.na}{Boolean. Remove NA values in the plot? (not filtered for 
numerical output; use na.omit() or filter() if needed)}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}
}
\description{
This function lets the user group, count, calculate percentages 
and cumulatives. It also plots results if needed. Tidyverse friendly.
}
