% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{freqs}
\alias{freqs}
\title{Frequencies Calculations}
\usage{
freqs(vector, ..., plot = F)
}
\arguments{
\item{vector}{Vector to group, count, and mutate}
}
\description{
This function lets the user group, count and calculate percentages and cumulatives
}
