% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{freqs}
\alias{freqs}
\title{Group by, count and percentages
Group by, count and percentages}
\usage{
freqs(data, ..., plot = F)
}
\description{
Group by, count and percentages
Group by, count and percentages
}
