% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{freqs}
\alias{freqs}
\title{Frequencies Calculations and Plot}
\usage{
freqs(vector, ..., plot = FALSE, rm.na = FALSE)
}
\arguments{
\item{vector}{Data.frame}

\item{...}{Variables. Variables you wish to process. Order matters.}

\item{plot}{Boolean. Do you want to see a plot? Three variables tops.}

\item{rm.na}{Boolean. Remove NA values in the plot? (not in 
numerical output - use dplyr::filter(!is.na(...)) if needed)}
}
\description{
This function lets the user group, count, calculate 
percentages and cumulatives. It also plots if needed. 
Perfect for inside with dplyr's pipes.
}
