% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{freqs_df}
\alias{freqs_df}
\title{All Frequencies on Data Frame}
\usage{
freqs_df(df, var = 0.9, plot = TRUE, save = FALSE, subdir = NA)
}
\arguments{
\item{df}{Data.frame}

\item{var}{Numeric. Top variance accepted. Range: (0-1]}

\item{plot}{Boolean. Do you want to see a plot? Three variables tops.}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to 
save the plot to?}
}
\description{
This function lets the user analize data by visualizing the
frequency of each value of each column from a whole data frame.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{df_str}}, \code{\link{distr}},
  \code{\link{freqs}}, \code{\link{gain_lift}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{textCloud}}, \code{\link{textTokenizer}},
  \code{\link{tree_var}}, \code{\link{trendsRelated}}

Other Visualization: \code{\link{corr_plot}},
  \code{\link{distr}}, \code{\link{freqs}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{textCloud}},
  \code{\link{theme_lares2}}, \code{\link{theme_lares}},
  \code{\link{tree_var}}
}
\concept{Exploratory}
\concept{Visualization}
