% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{freqs_list}
\alias{freqs_list}
\title{Frequencies on Lists and Plot}
\usage{
freqs_list(
  df,
  var,
  wt = NA,
  fx = "mean",
  limit = 10,
  unique = TRUE,
  abc = FALSE,
  title = "",
  plot = TRUE
)
}
\arguments{
\item{df}{Data.frame}

\item{var}{Variable. Variables you wish to process.}

\item{wt}{Variable, numeric. Select a numeric column to use 
in the colour scale, used as sum, mean... of those values for each
of the combinations.}

\item{fx}{Character. Set operation: mean, sum}

\item{limit}{Integer. Show top n combinations and elements. The rest
will be grouped into a single element.}

\item{unique}{Boolean. a,b = b,a?}

\item{abc}{Boolean. Do you wish to sort by alphabetical order?}

\item{title}{Character. Added to the plot.}

\item{plot}{Boolean. Plot viz?}
}
\description{
Visualize frequency of elements on a list or list vector, which
combinations and elements are the most frequent. Your variable can be
a character vector with comma separated values or a list vector.
}
\examples{
options("lares.font"=NA) # Temporal
df <- dplyr::starwars
df \%>\% freqs_list(films, limit = 5)
freqs_list(df, films, wt = height, abc = TRUE, 
           title = "Star Wars:\nCharacter's\nHeights per Film")
}
\seealso{
Other Frequency: 
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()}

Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{lasso_vars}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{summer}()},
\code{\link{tree_var}()},
\code{\link{trendsRelated}()}

Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{gg_pie}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Exploratory}
\concept{Frequency}
\concept{Visualization}
