% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{gain_lift}
\alias{gain_lift}
\title{Cumulative Gain, Lift and Response}
\usage{
gain_lift(tag, score, target = "auto", splits = 10, plot = FALSE,
  quiet = FALSE)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{target}{Value. Which is your target positive value? If 
set to 'auto', the target with largest mean(score) will be 
selected. Change the value to overwrite.}

\item{splits}{Integer. Numer of percentiles to split the data}

\item{plot}{Boolean. Plot results?}

\item{quiet}{Boolean. Do not show message for auto target?}
}
\description{
This function calculates cumulative gain, lift, and response 
values for a predictive score of a specific target. You can use the 
mplot_gain() function to custom plot results.
}
