% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{gain_lift}
\alias{gain_lift}
\title{Cumulative Gain, Lift and Response}
\usage{
gain_lift(tag, score, target = "auto", splits = 10, plot = FALSE,
  quiet = FALSE)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{target}{Value. Which is your target positive value? If 
set to 'auto', the target with largest mean(score) will be 
selected. Change the value to overwrite. Only used when binary
categorical model.}

\item{splits}{Integer. Numer of percentiles to split the data}

\item{plot}{Boolean. Plot results?}

\item{quiet}{Boolean. Do not show message for auto target?}
}
\description{
This function calculates cumulative gain, lift, and response 
values for a predictive score of a specific target. You can use the 
mplot_gain() function to custom plot results.
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{clusterKmeans}}, \code{\link{conf_mat}},
  \code{\link{export_results}}, \code{\link{h2o_automl}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_results}}, \code{\link{h2o_selectmodel}},
  \code{\link{impute}}, \code{\link{iter_seeds}},
  \code{\link{model_metrics}}, \code{\link{mplot_conf}},
  \code{\link{mplot_cuts_error}}, \code{\link{mplot_cuts}},
  \code{\link{mplot_density}}, \code{\link{mplot_full}},
  \code{\link{mplot_gain}}, \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{msplit}}

Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{df_str}}, \code{\link{distr}},
  \code{\link{freqs_df}}, \code{\link{freqs}},
  \code{\link{get_tweets}}, \code{\link{missingness}},
  \code{\link{plot_cats}}, \code{\link{plot_df}},
  \code{\link{plot_nums}}, \code{\link{tree_var}},
  \code{\link{trendsRelated}}
}
\concept{Exploratory}
\concept{Machine Learning}
