% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{geoStratum}
\alias{geoStratum}
\title{Get Colombia's Stratum given the Coordinates}
\usage{
geoStratum(lon, lat, label = NA)
}
\arguments{
\item{lon}{Numeric Vector. Longitudes}

\item{lat}{Numeric Vector. Latitudes}

\item{label}{Character Vector. If you wish to keep an id on each address, set this values}
}
\description{
This function lets the user obtain Colombia's stratum given the coordinates of an address
}
\seealso{
Other Geographical: 
\code{\link{geoAddress}()},
\code{\link{geoGrid}()},
\code{\link{geoMap}()}

Other Scrapper: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()}
}
\concept{Geographical}
\concept{Scrapper}
