% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Load personal parameters and credentials}
\usage{
get_credentials(from = NA, dir = NA, filename = "config.yml")
}
\arguments{
\item{from}{Character. Which account do we need to import}

\item{dir}{Character. Credentials directory for specific user}

\item{filename}{Character. YML with credentials to read}
}
\description{
This function lets the user load parameters and credentials
}
