% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Load personal parameters and credentials}
\usage{
get_credentials(from = NA, dir = NA, filename = "config.yml")
}
\arguments{
\item{from}{Character. Which account do we need to import}

\item{dir}{Character. Credentials directory for specific user}

\item{filename}{Character. YML with credentials to read}
}
\description{
This function lets the user load parameters and credentials
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{export_results}}, \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{iter_seeds}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{mailSend}}, \code{\link{matrixwd}},
  \code{\link{msplit}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{statusbar}},
  \code{\link{try_require}}, \code{\link{updateLares}},
  \code{\link{zerovar}}
}
\concept{Tools}
