% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Load parameters and credentials from YML file}
\usage{
get_credentials(from = NA, dir = NA, filename = "config.yml")
}
\arguments{
\item{from}{Character. Family of values to import from the YML file.
If you don't know these names, you can run 
\code{get_credentials(dir = "yourdir")} and a warning will display these values.}

\item{dir}{Character. Credentials directory where your YML file is.
If used with frequency, set your directory with 
\code{Sys.setenv(LARES_CREDS = "/your/creds/dir")} once and leave `dir` 
as `NA` to fetch this directory automatically. Remember to reset your
session the first time you set this parameter for it to work.}

\item{filename}{Character. YML filename with your credentials.}
}
\description{
This function lets the user load parameters and credentials
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{iter_seeds}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{pass}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{slackSend}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}
}
\concept{Tools}
