% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials}
\alias{get_credentials}
\alias{get_creds}
\title{Load parameters and credentials from YML file}
\usage{
get_credentials(
  from = NA,
  dir = NA,
  filename = "config.yml",
  env = "LARES_CREDS"
)

get_creds(from = NA, dir = NA, filename = "config.yml", env = "LARES_CREDS")
}
\arguments{
\item{from}{Character. Family of values to import from the YML file.
If you don't know these names, you can run 
\code{get_credentials(dir = "your/dir")} and a warning will display these values.}

\item{dir}{Character. Credentials directory where your YML file is.
If used frequently, set your directory by using the `.Renvironment` file. To do so,
leave \code{dir} as \code{NA} and follow the steps. Next time, everytime you need 
to fetch your credentials' YML this directory will be used automatically.
Remember to reset your session the first time you set this parameter.}

\item{filename}{Character. YML filename with your credentials.}

\item{env}{Character. Environment variable name. No need to set differently
for any function that uses this library. Only for external use}
}
\description{
This function lets the user load parameters and credentials
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{iter_seeds}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{pass}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}

Other Credentials: 
\code{\link{db_download}()},
\code{\link{f1_contacts}()},
\code{\link{geoAddress}()},
\code{\link{get_tweets}()},
\code{\link{mailSend}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{stocks_file}()},
\code{\link{stocks_report}()}
}
\concept{Credentials}
\concept{Tools}
