% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{get_stocks}
\alias{get_stocks}
\title{Get Personal Portfolio's Data}
\usage{
get_stocks(filename = NA)
}
\arguments{
\item{filename}{Characeter. Import a local Excel file}
}
\description{
This function lets me download my personal Excel with my Portfolio data
}
