% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{get_stocks}
\alias{get_stocks}
\title{Get Personal Portfolio's Data}
\usage{

  get_stocks(token_dir = "~/Dropbox (Personal)/Documentos/Interactive Brokers/Portfolio",
  dir = NA)
}
\arguments{
\item{token_dir}{Character. Directory containing personal Dropbox token file}
}
\description{
This function lets me download my personal Excel with my Portfolio data
}
