% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{get_stocks}
\alias{get_stocks}
\title{Get Personal Portfolio's Data}
\usage{
get_stocks(filename = NA,
  token_dir = "~/Dropbox (Personal)/Documentos/Docs/Data", auto = TRUE)
}
\arguments{
\item{filename}{Characeter. Import a local Excel file}

\item{token_dir}{Character. Where is my personal token for Dropbox connection?}

\item{auto}{Boolean. Automatically user my local personal file?}
}
\description{
This function lets me download my personal Excel with my Portfolio data
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks_hist}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
