% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_tweets}
\alias{get_tweets}
\title{Get Tweets}
\usage{
get_tweets(q, n = 10000, creds = NA)
}
\arguments{
\item{q}{Query. Check for ?rtweet::search_tweets()}

\item{n}{Integer. Total of tweets to return}

\item{creds}{Character. Credential's user (see \code{get_creds()})}
}
\value{
data.frame with API response results.
}
\description{
This function downloads tweets with personal credentials
}
\seealso{
Other Credentials: 
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{get_credentials}()},
\code{\link{mail_send}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()},
\code{\link{stocks_file}()},
\code{\link{stocks_report}()}
}
\concept{Credentials}
\concept{Twitter}
