% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{gg_colour_customs}
\alias{gg_colour_customs}
\title{Custom colours to use in ggplot as scale_color_manual}
\usage{
gg_colour_customs()
}
\description{
This function lets the user use some default colours
}
