% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lares.R
\name{gg_fill_customs}
\alias{gg_fill_customs}
\alias{gg_colour_customs}
\alias{gg_text_customs}
\title{Custom fill, colour and text colours for ggplot2}
\usage{
gg_fill_customs(layer = "fill", ...)

gg_colour_customs(layer = "colour", ...)

gg_text_customs(layer = "colour", ...)
}
\arguments{
\item{layer}{Select any of "fill" or "colour" to use on
your \code{lares_pal()$labels} palette.}

\item{...}{Alow additional parameters not used.}
}
\value{
Same as \code{scale_fill_manual} but with custom palette.
}
\description{
This function lets the user use pre-defined default colours.
Check your \code{lares_pal()$labels} scale. Feel free to use
\code{lares:::.gg_vals()} to debug colours used in latest plot.
}
\seealso{
Other Auxiliary: 
\code{\link{lares_pal}()},
\code{\link{plot_palette}()}
}
\concept{Auxiliary}
