% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{gg_fill_customs}
\alias{gg_fill_customs}
\title{Custom colours for scale_fill_manual [Deprecated]}
\usage{
gg_fill_customs()
}
\value{
Same as \code{scale_fill_manual} but with custom palette.
}
\description{
This function lets the user use pre-defined default colours
}
\seealso{
Other Auxiliary: 
\code{\link{gg_colour_customs}()},
\code{\link{gg_text_customs}()},
\code{\link{lares_pal}()},
\code{\link{plot_palette}()}
}
\concept{Auxiliary}
