% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{gg_pie}
\alias{gg_pie}
\title{Density plot for discrete and continuous values}
\usage{
gg_pie(df, var, table = FALSE, save = FALSE, subdir = NA)
}
\arguments{
\item{df}{Dataframe Event, role, label, or row.}

\item{var}{Variable to group, count and plot}

\item{table}{Boolean. Print results as table?}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}
}
\description{
This function plots discrete and continuous values results
}
