% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{gg_pie}
\alias{gg_pie}
\title{Density plot for discrete and continuous values}
\usage{
gg_pie(df, var, table = FALSE, ...)

gg_pie(df, var, table = FALSE, ...)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Variable to group, count and plot}

\item{table}{Boolean. Print results as table?}

\item{...}{Further parameters passed to \code{freqs()}}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
This function plots discrete and continuous values results

This function plots discrete and continuous values results
}
\examples{
Sys.unsetenv("LARES_FONT") # Temporal
data(dft) # Titanic dataset
gg_pie(dft, Survived)
gg_pie(dft, Pclass, table = TRUE)
gg_pie(dft, SibSp, table = TRUE, abc = TRUE)
Sys.unsetenv("LARES_FONT") # Temporal
data(dft) # Titanic dataset
gg_pie(dft, Survived)
gg_pie(dft, Pclass, table = TRUE)
gg_pie(dft, SibSp, table = TRUE, abc = TRUE)
}
\seealso{
Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}

Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Visualization}
