% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_automl}
\alias{h2o_automl}
\title{Automated H2O's AutoML}
\usage{
h2o_automl(df, train_test = NA, split = 0.7, weight = NULL,
  balance = TRUE, seed = 0, thresh = 5, max_time = 5 * 60,
  max_models = 25, start_clean = TRUE,
  exclude_algos = c("StackedEnsemble", "DeepLearning"), alarm = TRUE,
  save = FALSE, subdir = NA, project = "Machine Learning Model")
}
\arguments{
\item{df}{Dataframe. Dataframe containing all your data, including 
the independent variable labeled as 'tag'}

\item{train_test}{Character. If needed, df's column name with 'test' 
and 'train' values to split}

\item{split}{Numeric. Value between 0 and 1 to split as train/test 
datasets. Value is for training set.}

\item{weight}{Column with observation weights. Giving some observation a
weight of zero is equivalent to excluding it from the dataset; giving an 
observation a relative weight of 2 is equivalent to repeating that 
row twice. Negative weights are not allowed.}

\item{balance}{Boolean. Auto-balance train dataset with under-sampling?}

\item{seed}{Numeric. Seed for random stuff and reproducibility}

\item{thresh}{Integer. Threshold for selecting binary or regression 
models: this number is the threshold of unique values we should 
have in 'tag' (more than: regression; less than: classification)}

\item{max_time}{Numeric. Max seconds you wish for the function 
to iterate}

\item{max_models}{Numeric. Max models you wish for the function 
to create}

\item{start_clean}{Boolean. Erase everything in the current h2o 
instance before we start to train models?}

\item{exclude_algos}{Vector of character strings. Algorithms to 
skip during the model-building phase.}

\item{alarm}{Boolean. Ping an alarm when ready!}

\item{save}{Boolean. Do you wish to save/export results into your 
working directory?}

\item{subdir}{Character. In which directory do you wish to save 
the results? Working directory as default.}

\item{project}{Character. Your project's name}

\item{plots}{Boolean. Do you want to plot the results metrics?}
}
\description{
This function lets the user create a robust and fast model, using 
H2O's AutoML function. The result is a list with the best model, 
its parameters, datasets, performance metrics, variables 
importances, and other useful metrics.
}
\details{
Full list of algorithms: "DRF" (Distributed Random Forest, including 
Random Forest (RF) and Extremely-Randomized Trees (XRT)), "GLM" 
(Generalized Linear Model), "XGBoost" (eXtreme Grading Boosting), 
"GBM" (Gradient Boosting Machine), "DeepLearning" (Fully-connected 
multi-layer artificial neural network) and "StackedEnsemble".
}
