% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{h2o_predict_model}
\alias{h2o_predict_model}
\title{H2O Predict using H2O Model Object}
\usage{
h2o_predict_model(df, model)
}
\arguments{
\item{df}{Dataframe/Vector. Data to insert into the model.}

\item{model}{h2o model Object}
}
\value{
data.frame with predicted results.
}
\description{
This function lets the user get scores from a H2O Model Object.
}
\seealso{
Other Machine Learning: 
\code{\link{ROC}()},
\code{\link{conf_mat}()},
\code{\link{export_results}()},
\code{\link{gain_lift}()},
\code{\link{h2o_automl}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_selectmodel}()},
\code{\link{impute}()},
\code{\link{iter_seeds}()},
\code{\link{lasso_vars}()},
\code{\link{model_metrics}()},
\code{\link{model_preprocess}()},
\code{\link{msplit}()}

Other H2O: 
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()}
}
\concept{H2O}
\concept{Machine Learning}
