% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_predict_MOJO}
\alias{h2o_predict_MOJO}
\title{H2O Predict using MOJO file}
\usage{
h2o_predict_MOJO(df, model_path, sample = NA)
}
\arguments{
\item{df}{Dataframe. Data to insert into the model}

\item{model_path}{Character. Relative model_path directory}

\item{sample}{Integer. How many rows should the function predict?}
}
\description{
This function lets the user predict using the h2o .zip file 
containing the MOJO files. Note that it works with the files 
generated when using the function export_results()
}
