% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_selectmodel}
\alias{h2o_selectmodel}
\title{Select Model from h2o_automl's Leaderboard}
\usage{
h2o_selectmodel(results, which_model = 1)
}
\arguments{
\item{results}{Object. h2o_automl output}

\item{which_model}{Integer. Which model from the leaderboard you wish to use?}
}
\description{
Select wich model from the h2o_automl function to use
}
