% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{h2o_selectmodel}
\alias{h2o_selectmodel}
\title{Select Model from h2o_automl's Leaderboard}
\usage{
h2o_selectmodel(results, which_model = 1)
}
\arguments{
\item{results}{Object. h2o_automl output}

\item{which_model}{Integer. Which model from the leaderboard you wish to use?}
}
\description{
Select wich model from the h2o_automl function to use
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{clusterKmeans}}, \code{\link{conf_mat}},
  \code{\link{export_results}}, \code{\link{gain_lift}},
  \code{\link{h2o_automl}}, \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}}, \code{\link{impute}},
  \code{\link{iter_seeds}}, \code{\link{model_metrics}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{msplit}}

Other Tools: \code{\link{autoline}},
  \code{\link{bindfiles}}, \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_update}}, \code{\link{haveInternet}},
  \code{\link{importxlsx}}, \code{\link{ip_country}},
  \code{\link{iter_seeds}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{mailSend}},
  \code{\link{matrixwd}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{try_require}},
  \code{\link{updateLares}}, \code{\link{zerovar}}
}
\concept{Machine Learning}
\concept{Tools}
