% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{h2o_update}
\alias{h2o_update}
\title{Install latest version of H2O}
\usage{
h2o_update(run = TRUE)
}
\arguments{
\item{run}{Boolean. Do you want to run and start an H2O cluster?}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{haveInternet}},
  \code{\link{importxlsx}}, \code{\link{ip_country}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\concept{Tools}
