% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{h2o_update}
\alias{h2o_update}
\title{Install latest version of H2O}
\usage{
h2o_update(run = TRUE, lib = .libPaths())
}
\arguments{
\item{run}{Boolean. Do you want to run and start an H2O cluster?}

\item{lib}{Character. Library directories where to install h2o}
}
\description{
This function lets the user un-install the current version of
H2O installed and update to latest stable version.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{iter_seeds}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{mailSend}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{updateLares}},
  \code{\link{zerovar}}
}
\concept{Tools}
