% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{haveInternet}
\alias{haveInternet}
\title{Internet Connection Check}
\usage{
haveInternet(thresh = 3, url = "http://www.google.com")
}
\arguments{
\item{thresh}{Numeric. How many seconds to consider a slow connection?}

\item{url}{Character. URL to test the readLines 1 command}
}
\description{
This function checks if your R session currently have Wifi or 
Internet connection.
}
