% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{haveInternet}
\alias{haveInternet}
\title{Internet Connection Check}
\usage{
haveInternet(thresh = 3, url = "http://www.google.com")
}
\arguments{
\item{thresh}{Numeric. How many seconds to consider a slow connection?}

\item{url}{Character. URL to test the readLines 1 command}
}
\description{
This function checks if your R session currently have Wifi or 
Internet connection.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{importxlsx}}, \code{\link{ip_country}},
  \code{\link{iter_seeds}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{mailSend}},
  \code{\link{matrixwd}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{try_require}},
  \code{\link{updateLares}}, \code{\link{zerovar}}
}
\concept{Tools}
