% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{holidays}
\alias{holidays}
\title{Holidays in your Country}
\usage{
holidays(years = year(Sys.Date()), countries = "Colombia")
}
\arguments{
\item{years}{Integer or vector. For which year do you wish to import
holiday dates?}

\item{countries}{Character or vector. For which countries should the 
holidays be imported?}
}
\description{
This function lets the user automatically scrap holiday dates from
any country and year within +- 5 years. Thanks to timeanddate.com!
}
