% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{holidays}
\alias{holidays}
\title{Holidays in your Country}
\usage{
holidays(countries = "Colombia", years = year(Sys.Date()))
}
\arguments{
\item{countries}{Character or vector. For which country(ies) should the 
holidays be imported?}

\item{years}{Character or vector. For which year(s) do you wish to import
holiday dates?}
}
\description{
This function lets the user automatically scrap holiday dates from
any country and year within +- 5 years. Thanks to timeanddate.com!
}
\examples{
\dontrun{
holidays(countries = "Argentina") 
holidays(countries = c("Argentina", "Venezuela"), years = c(2019, 2020))
}
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}

Other Feature Engineering: 
\code{\link{date_feats}()},
\code{\link{ohse}()}

Other Scrapper: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{ip_country}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()}
}
\concept{Data Wrangling}
\concept{Feature Engineering}
\concept{Scrapper}
