% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{holidays}
\alias{holidays}
\title{Holidays in your Country}
\usage{
holidays(countries = "Colombia", years = year(Sys.Date()))
}
\arguments{
\item{countries}{Character or vector. For which country(ies) should the 
holidays be imported?}

\item{years}{Character or vector. For which year(s) do you wish to import
holiday dates?}
}
\description{
This function lets the user automatically scrap holiday dates from
any country and year within +- 5 years. Thanks to timeanddate.com!
}
