% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{importxlsx}
\alias{importxlsx}
\title{Import Excel File with All Its Tabs}
\usage{
importxlsx(file)
}
\arguments{
\item{file}{String. Excel's name}
}
\description{
This function lets the user import an Excel file's tabs into a list
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{ip_country}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\concept{Tools}
