% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{importxlsx}
\alias{importxlsx}
\title{Import Excel File with All Its Tabs}
\usage{
importxlsx(file)
}
\arguments{
\item{file}{String. Excel's lilename}
}
\description{
This function lets the user import an Excel file's tabs into a list
}
