% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{importxlsx}
\alias{importxlsx}
\title{Import Excel File with All Its Tabs}
\usage{
importxlsx(file)
}
\arguments{
\item{file}{String. Local Excel file name}
}
\value{
List or data.frame. If single tab is found, a data.frame; if
multiple tabs are found on file, a list of data.frames.
}
\description{
This function lets the user import an Excel file's tabs into a list
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}
}
\concept{Tools}
