% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{impute}
\alias{impute}
\title{Impute Missing Values (using MICE)}
\usage{
impute(df, m = 5, iters = 5, seed = 0, quiet = FALSE)
}
\arguments{
\item{df}{Dataframe. Dataframe to transform}

\item{m}{Integer. Number of multiple imputations}

\item{iters}{Integer. Number of iterations}

\item{seed}{Integer. Set a seed for reproducibility}

\item{quiet}{Boolean. Keep quiet? (or print replacements)}
}
\description{
This function uses the MICE methodology to impute missing values.
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}

Other Machine Learning: 
\code{\link{ROC}()},
\code{\link{clusterKmeans}()},
\code{\link{conf_mat}()},
\code{\link{export_results}()},
\code{\link{gain_lift}()},
\code{\link{h2o_automl}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_results}()},
\code{\link{h2o_selectmodel}()},
\code{\link{iter_seeds}()},
\code{\link{lasso_vars}()},
\code{\link{model_metrics}()},
\code{\link{mplot_conf}()},
\code{\link{mplot_cuts_error}()},
\code{\link{mplot_cuts}()},
\code{\link{mplot_density}()},
\code{\link{mplot_full}()},
\code{\link{mplot_gain}()},
\code{\link{mplot_importance}()},
\code{\link{mplot_lineal}()},
\code{\link{mplot_metrics}()},
\code{\link{mplot_response}()},
\code{\link{mplot_roc}()},
\code{\link{mplot_splits}()},
\code{\link{msplit}()}

Other Missing Values: 
\code{\link{missingness}()}
}
\concept{Data Wrangling}
\concept{Machine Learning}
\concept{Missing Values}
