% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{ip_country}
\alias{ip_country}
\title{Find country from a given IP}
\usage{
ip_country(ip)
}
\arguments{
\item{ip}{Vector. Vector with all IP's we wish to search}
}
\description{
This function lets the user find a country from a given IP Address
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{iter_seeds}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{mailSend}},
  \code{\link{matrixwd}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{try_require}},
  \code{\link{updateLares}}, \code{\link{zerovar}}

Other Scrapper: \code{\link{fb_posts}},
  \code{\link{fb_post}}, \code{\link{geoStratum}},
  \code{\link{holidays}}, \code{\link{trendsRelated}}
}
\concept{Scrapper}
\concept{Tools}
