% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{ip_country}
\alias{ip_country}
\title{Find country from a given IP}
\usage{
ip_country(ip)
}
\arguments{
\item{ip}{Vector. Vector with all IP's we wish to search}
}
\description{
This function lets the user find a country from a given IP Address
}
