% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{iter_seeds}
\alias{iter_seeds}
\title{Iterate and Search for Best Seed}
\usage{
iter_seeds(df, tries = 10)
}
\arguments{
\item{df}{Dataframe. Dataframe with all your data you wish to model (see h2o_automl's df)}

\item{tries}{Integer. How many seed do you wish to try?}
}
\description{
This functions lets the user iterate and search for best seed. Note that if
the results change a lot, you are having a high variance in your data.
}
