% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{iter_seeds}
\alias{iter_seeds}
\title{#' Example source: https://github.com/alan-y/objectremover/blob/master/R/object_remove.R
#' model_exporter
#' @description Friendly h2o_automl() results export tool
#'
#' @param model List. Result from h2o_automl()
#' @export
#' @examples
#' if (interactive()) {
#'   model_exporter()
#' }
model_exporter <- function(model) {
  
  try_require("shiny")
  try_require("miniUI")
  try_require("shinyWidgets")
  try_require("h2o")
  
  ui <- miniPage(
    gadgetTitleBar(model$model_name, left = NULL),
    miniContentPanel(
      strong(p("Performance summary:")),
      tableOutput("metrics"),
      tableOutput("metrics_other"),
      checkboxGroupInput("checkGroup",
                         label = strong("Select files to export:"),
                         choices = list("Summary in txt file" = 1L,
                                        "CSV file with datasets and scores" = 5L,
                                        "RDS object" = 2L,
                                        "MOJO files (cross-platform friendly)" = 4L,
                                        "Binary model file" = 3L,
                                        "Plots as PNG" = 6L),
                         selected = c(1, 2, 3, 4, 5)),
      hr(),
      strong(p("Current directory:")),
      p(textOutput("directory")), 
      hr(),
      actionButton("run", icon = icon("save"), label = "Generate files")
    )
  )
  
  server <- function(input, output, session) {
    
    output$metrics <- renderTable(model$metrics$metrics) 
    
    output$directory <- renderPrint(getwd())
    
    observeEvent(input$run, {
      withProgress(message = "Generating files...", value = 1, {
        export_results(model, 
                       txt = 1L %in% input$checkGroup,
                       rds = 2L %in% input$checkGroup,
                       binary = 3L %in% input$checkGroup,
                       mojo = 4L %in% input$checkGroup,
                       csv = 5L %in% input$checkGroup,
                       plots = 6L %in% input$checkGroup)  
      })
      sendSweetAlert(session, title = "Done!", type = "success",
                     text = paste("Succesfully exported results for", 
                                  model$model_name, "into", getwd()))
    })
    
    observeEvent(input$done, {
      stopApp(message("Done exporting. Bye!"))
    })
  }
  runGadget(ui, server)
}
Iterate and Search for Best Seed}
\usage{
iter_seeds(df, tries = 10)
}
\arguments{
\item{df}{Dataframe. Dataframe with all your data you wish to model (see h2o_automl's df)}

\item{tries}{Integer. How many seed do you wish to try?}
}
\description{
This functions lets the user iterate and search for best seed. Note that if
the results change a lot, you are having a high variance in your data.
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{clusterKmeans}}, \code{\link{conf_mat}},
  \code{\link{export_results}}, \code{\link{gain_lift}},
  \code{\link{h2o_automl}}, \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_results}}, \code{\link{h2o_selectmodel}},
  \code{\link{impute}}, \code{\link{model_metrics}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{msplit}}

Other Tools: \code{\link{autoline}},
  \code{\link{bindfiles}}, \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{mailSend}},
  \code{\link{matrixwd}}, \code{\link{msplit}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{try_require}},
  \code{\link{updateLares}}, \code{\link{zerovar}}
}
\concept{Machine Learning}
\concept{Tools}
