% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{json2vector}
\alias{json2vector}
\title{Download Historical Currency Exchange Rate}
\usage{
json2vector(json)
}
\arguments{
\item{json}{Character. JSON string. Example: '{"feat1": null, "feat2": "M"}'}
}
\description{
This function lets the user download historical currency exchange
rate between two currencies
}
