% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{json2vector}
\alias{json2vector}
\title{Convert JSON string to vector (data.frame with 1 row)}
\usage{
json2vector(json)
}
\arguments{
\item{json}{Character. JSON string. Example of a string: '{"feat1": 
null, "feat2": "M"}'}
}
\description{
This function lets the user transform a JSON string into vector 
(data.frame with 1 row). You can also pass a Python's dictionary.
}
