% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares.R
\docType{package}
\name{lares-package}
\alias{lares}
\alias{lares-package}
\title{lares: Analytics, Visualization & Machine Learning Library}
\description{
R Library for Analytics, Data Mining and Machine Learning Tasks
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laresbernardo/lares}
  \item Report bugs at \url{https://github.com/laresbernardo/lares/issues}
}

}
