% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares.R
\docType{package}
\name{lares-package}
\alias{lares}
\alias{lares-package}
\title{lares: R library for Analytics, Visualization & Machine Learning Tasks}
\description{
R library for better/faster analytics, visualization, data mining, and machine learning tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laresbernardo/lares}
  \item Report bugs at \url{https://github.com/laresbernardo/lares/issues}
}

}
