% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{lares_pal}
\alias{lares_pal}
\title{Palette Colours}
\usage{
lares_pal(which = 1)
}
\arguments{
\item{which}{Integer. Which palette?}
}
\description{
This function plots a list of colours on a specific palette
}
