% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso_vars}
\alias{lasso_vars}
\title{Find Most Relevant Features Using Lasso Regression}
\usage{
lasso_vars(
  df,
  variable,
  ignore = NA,
  nlambdas = 100,
  nfolds = 10,
  seed = 123,
  ...
)
}
\arguments{
\item{df}{Dataframe}

\item{variable}{Variable.}

\item{ignore}{Character vector. Variables to exclude from study}

\item{nlambdas}{Integer. Number of lambdas to be used in a search}

\item{nfolds}{Integer. Number of folds for K-fold cross-validation (>= 2)}

\item{seed}{Numeric}

\item{...}{ohse parameters}
}
\description{
Use Lasso regression to identify the most relevant variables that 
can predict/identify another variable. You might want to compare 
with corr_var() results for further understanding. No need to
standardize your data.
}
\examples{
data(dft) # Titanic dataset
m <- lasso_vars(dft, Survived, ignore = c("Ticket", "Cabin"))
print(m$coef)
}
\seealso{
Other Machine Learning: 
\code{\link{ROC}()},
\code{\link{clusterKmeans}()},
\code{\link{conf_mat}()},
\code{\link{export_results}()},
\code{\link{gain_lift}()},
\code{\link{h2o_automl}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_results}()},
\code{\link{h2o_selectmodel}()},
\code{\link{impute}()},
\code{\link{iter_seeds}()},
\code{\link{model_metrics}()},
\code{\link{mplot_conf}()},
\code{\link{mplot_cuts_error}()},
\code{\link{mplot_cuts}()},
\code{\link{mplot_density}()},
\code{\link{mplot_full}()},
\code{\link{mplot_gain}()},
\code{\link{mplot_importance}()},
\code{\link{mplot_lineal}()},
\code{\link{mplot_metrics}()},
\code{\link{mplot_response}()},
\code{\link{mplot_roc}()},
\code{\link{mplot_splits}()},
\code{\link{msplit}()}
}
\concept{Machine Learning}
