% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{left}
\alias{left}
\title{Left: First n characters}
\usage{
left(string, n)
}
\arguments{
\item{string}{String or Vector}

\item{n}{Integer. How many characters from left to right?}
}
\description{
This function lets the user extract the first n characters of a
string or vector of strings.
}
