% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{left}
\alias{left}
\alias{right}
\title{Left or Right N characters of a string}
\usage{
left(string, n = 1)

right(string, n = 1)
}
\arguments{
\item{string}{String or Vector.}

\item{n}{Integer. How many characters starting on right/left?}
}
\value{
Character. Trimmed strings.
}
\description{
This functions lets the user extract the first or last n characters
of a string or vector of strings.
}
\examples{
left("Bernardo", 3)
right(c("Bernardo", "Lares", "V"), 3)
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{file_name}()},
\code{\link{formatHTML}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{normalize}()},
\code{\link{num_abbr}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{quants}()},
\code{\link{removenacols}()},
\code{\link{replaceall}()},
\code{\link{replacefactor}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{zerovar}()}
}
\concept{Data Wrangling}
