% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_auth}
\alias{li_auth}
\title{OAuth Linkedin}
\usage{
li_auth(app_name = NA, client_id = NA, client_secret = NA)
}
\arguments{
\item{app_name}{Character. Your App's given name}

\item{client_id}{Character. Your App's client ID}

\item{client_secret}{Character. Your App's client secret}
}
\description{
This function authenticates and creates a token for LinkedIn's 
API REST
}
