% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_profile}
\alias{li_profile}
\title{Get My Personal LinkedIn Data}
\usage{
li_profile(token = NA)
}
\arguments{
\item{token}{Object. li_auth() output}
}
\description{
This function brings a list with my personal Linkedin data
}
