% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_profile}
\alias{li_profile}
\title{Get My Personal LinkedIn Data}
\usage{
li_profile(token = NA)
}
\arguments{
\item{token}{Object. OAuth Authentication: li_auth()'s output.}
}
\value{
List. Results of your own profile data given the \code{token}.
}
\description{
This function brings a list with your personal LinkedIn data
}
\seealso{
Other API: 
\code{\link{bring_api}()},
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_creatives}()},
\code{\link{fb_insights}()},
\code{\link{fb_process}()},
\code{\link{fb_report_check}()},
\code{\link{fb_rf}()},
\code{\link{fb_token}()},
\code{\link{gemini_ask}()},
\code{\link{gpt_ask}()},
\code{\link{li_auth}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()}

Other LinkedIn: 
\code{\link{li_auth}()}
}
\concept{API}
\concept{LinkedIn}
