% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{list_cats}
\alias{list_cats}
\title{List categorical values for data.frame}
\usage{
list_cats(df, ..., abc = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{...}{Variables to segment counters}

\item{abc}{Boolean. Sort alphabetically?}
}
\description{
Make a list with all categorical values and
}
