% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{list_cats}
\alias{list_cats}
\title{List categorical values for data.frame}
\usage{
list_cats(df, ..., abc = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{...}{Variables to segment counters}

\item{abc}{Boolean. Sort alphabetically?}
}
\value{
List. Length same as number of categorical columns, each with a
frequency data.frame using \code{freqs()}.
}
\description{
Make a list with all categorical values and
}
\examples{
data(dft) # Titanic dataset
df <- dft[, 1:5]
head(df)
list_cats(df)
}
