% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{list_cats}
\alias{list_cats}
\title{List categorical values for data.frame}
\usage{
list_cats(df, ..., abc = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{...}{Variables to segment counters}

\item{abc}{Boolean. Sort alphabetically?}
}
\value{
List. Length same as number of categorical columns, each with a
frequency data.frame using \code{freqs()}.
}
\description{
Make a list with all categorical values and
}
\examples{
data(dft) # Titanic dataset
df <- dft[, 1:5]
head(df)
list_cats(df)
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bind_files}()},
\code{\link{bring_api}()},
\code{\link{chr2num}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{dont_sleep}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{files_functions}()},
\code{\link{font_exists}()},
\code{\link{formatColoured}()},
\code{\link{formatHTML}()},
\code{\link{glued}()},
\code{\link{grepm}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mail_send}()},
\code{\link{markdown2df}()},
\code{\link{move_files}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{warnifnot}()},
\code{\link{what_size}()}
}
\concept{Tools}
