% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{listfiles}
\alias{listfiles}
\title{List files in a directory}
\usage{
listfiles(folder = getwd(), recursive = TRUE, regex = NA,
  images = FALSE, export = FALSE)
}
\arguments{
\item{folder}{Character. Directory which contains files}

\item{recursive}{Boolean. Should the listing recurse into directories?}

\item{regex}{Character. String to use for filtering files}

\item{images}{Boolean. Bring only image files?}

\item{export}{Boolean. Do you wish to export list as txt file?}
}
\description{
This function lets the user list all files on a given directory.
It also lets filter files which contains a string.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{myip}}, \code{\link{pass}},
  \code{\link{quiet}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\concept{Tools}
