% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{listfiles}
\alias{listfiles}
\title{List files in a directory}
\usage{
listfiles(folder, recursive = TRUE, type = NA, export = FALSE)
}
\arguments{
\item{folder}{Character. Directory which contains files}

\item{recursive}{Boolean. Should the listing recurse into directories?}

\item{type}{Character. String to use for filtering files}

\item{export}{Boolean. Do you wish to export a txt file?}
}
\description{
This function lets the user list all files on a given directory.
It also lets filter files which contains a string.
}
