% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{loglossBinary}
\alias{loglossBinary}
\title{Loggarithmic Loss Function for Binary Models}
\usage{
loglossBinary(tag, score, eps = 1e-15)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{eps}{Numeric. Epsilon value}
}
\description{
This function calculates Loggarithmic Loss for Binary Models
You can define a seed to get the same results every time, but has a default value.
You can prevent it from printing the split counter result.
}
