% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{loglossBinary}
\alias{loglossBinary}
\title{Loggarithmic Loss Function for Binary Models
Loggarithmic Loss Function for Binary Models}
\usage{
loglossBinary(tag, score, eps = 1e-15)
}
\value{
LogLoss calculation
}
\description{
Loggarithmic Loss Function for Binary Models
Loggarithmic Loss Function for Binary Models
}
