% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{loglossBinary}
\alias{loglossBinary}
\title{Loggarithmic Loss Function for Binary Models}
\usage{
loglossBinary(tag, score, eps = 0.000000000000001)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{eps}{Numeric. Epsilon value}
}
\description{
This function calculates log loss/cross-entropy loss for binary 
models. NOTE: when result is 0.69315, the classification is neutral; 
it assigns equal probability to both classes.
}
