% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mails.R
\name{mailSend}
\alias{mailSend}
\title{Send Emails with Attachments (POST)}
\usage{
mailSend(
  from = "RMail <laresbernardo@gmail.com>",
  to = "laresbernardo@gmail.com",
  cc = NA,
  bcc = NA,
  subject = "Mail from R",
  text = NA,
  html = NA,
  attachment = NA,
  quiet = FALSE,
  creds = NA
)
}
\arguments{
\item{from, to, cc, bcc}{Character. Emails}

\item{subject}{Character. Subject for the email}

\item{text, html}{Character. Text or HTML to send in the body}

\item{attachment}{Character, plot or data.frame. Will send the file, 
plot as PNG or data.frame as CSV, respectively.}

\item{quiet}{Boolean. Keep quite or display messages?}

\item{creds}{Character. Credential's user (see \code{get_creds()}).
Must contain: url (POST address), api (api key)}
}
\description{
This function lets the user send Emails with Attachments using MailGun's
API service.
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{iter_seeds}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{pass}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}

Other Credentials: 
\code{\link{db_download}()},
\code{\link{f1_contacts}()},
\code{\link{geoAddress}()},
\code{\link{get_credentials}()},
\code{\link{get_tweets}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()},
\code{\link{stocks_file}()},
\code{\link{stocks_report}()}
}
\concept{Credentials}
\concept{Tools}
