% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mails.R
\name{mailSend}
\alias{mailSend}
\title{Send Emails with Attachments (POST)}
\usage{
mailSend(from = "RMail <laresbernardo@gmail.com>",
  to = "laresbernardo@gmail.com", cc = NA, bcc = NA,
  subject = "Mail from R", text = "No text...", html = NA,
  attachment = NA, quiet = FALSE, creds = NA)
}
\arguments{
\item{from, to, cc, bcc}{Character. Emails}

\item{subject}{Character. Subject for the email}

\item{text, html}{Character. Text or HTML to send in the body}

\item{attachment}{Character, plot or data.frame. Will send the file, 
plot as PNG or data.frame as CSV, respectively.}

\item{quiet}{Boolean. Keep quite or display messages?}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user send Emails with Attachments using MailGun
}
