% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mails.R
\name{mailSend}
\alias{mailSend}
\title{Send Emails with Attachments (Sendgrid)}
\usage{
mailSend(body, subject, attachment = NULL, to = credentials$to,
  from = credentials$from, creds = NA, quite = FALSE)
}
\arguments{
\item{body}{Character. Body for the email}

\item{subject}{Character. Subject for the email}

\item{attachment}{Vector. All objects which will be sent as attachment}

\item{to}{Character. Email of the person who will recevie the email}

\item{from}{Character. Email of the person who send the email}

\item{creds}{Character. Credential's user (see get_credentials)}

\item{quite}{Boolean. Keep quite or display messages?}
}
\description{
This function lets the user send Emails with Attachments using Sendgrid
}
