% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{missingness}
\alias{missingness}
\title{Calculate missingness (%) on a data.frame}
\usage{
missingness(df, plot = FALSE, bins = 25)
}
\arguments{
\item{df}{Dataframe. Dataframe to study}

\item{bins}{Integer. Number of breaks on plotted histogram}

\item{print}{Boolean. Do you wish to print results in a simple histogram?}
}
\description{
This function lets the user calculate the porcentage of NAs or
missingness in a data.frame.
}
