% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{missingness}
\alias{missingness}
\title{Calculate missingness percentage on a data.frame}
\usage{
missingness(df, print = FALSE, bins = 10)
}
\arguments{
\item{df}{Dataframe. Dataframe to study}

\item{print}{Boolean. Do you wish to print results in a simple histogram?}

\item{bins}{Integer. Number of breaks on plotted histogram}
}
\description{
This function lets the user calculate the percentage of NAs or
missingness in a data.frame.
}
