% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{missingness}
\alias{missingness}
\title{Calculate missingness percentage on a data.frame}
\usage{
missingness(df, plot = FALSE)
}
\arguments{
\item{df}{Dataframe. Dataframe to study}

\item{plot}{Boolean. Do you wish to plot results?}
}
\description{
This function lets the user calculate the percentage of NAs or
missingness in a data.frame.
}
