% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings.R
\name{missingness}
\alias{missingness}
\title{Calculate and Visualize Missingness}
\usage{
missingness(df, plot = FALSE, full = FALSE, subtitle = NA,
  summary = TRUE)
}
\arguments{
\item{df}{Dataframe. Dataframe to study}

\item{plot}{Boolean. Do you wish to plot results?}

\item{full}{Boolean. Return all variables (or only with missings)?}

\item{subtitle}{Character. Subtitle to show in plot}

\item{summary}{Boolean. Show numerical summary text?}
}
\description{
This function lets the user calculate the percentage of NAs or
missingness in a data.frame. It also plots the results if needed.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{df_str}}, \code{\link{distr}},
  \code{\link{freqs_df}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{plot_cats}}, \code{\link{plot_df}},
  \code{\link{plot_nums}}, \code{\link{summer}},
  \code{\link{tree_var}}, \code{\link{trendsRelated}}
}
\concept{Exploratory}
