% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{model_metrics}
\alias{model_metrics}
\title{Classification Model Metrics}
\usage{
model_metrics(tag, score, multis = NA, thresh = 0.5, plots = TRUE,
  subtitle = NA)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{multis}{Data.frame. Containing columns with each category score 
(only used when more than 2 categories coexist)}

\item{thresh}{Numeric. Value which splits the results for the 
confusion matrix.}

\item{plots}{Boolean. Include plots?}

\item{subtitle}{Character. Subtitle for plots}
}
\description{
This function lets the user get a confusion matrix and accuracy, and 
for for binary classification models: AUC, Precision, Sensitivity, and
Specificity.
}
