% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{move_files}
\alias{move_files}
\title{Move files from A to B}
\usage{
move_files(from, to)
}
\arguments{
\item{from}{Character. File names and directories. All files
will be moved recursively.}

\item{to}{Character. File names for each \code{from} file or
directory. If directory does not exist, it will be created.}
}
\value{
No return value, called for side effects.
}
\description{
Move one or more files from a directory to another using R.
}
