% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_cuts}
\alias{mplot_cuts}
\title{Cuts by quantiles for score plot
Cuts by quantiles for score plot}
\usage{
mplot_cuts(score, splits = 10, model_name = NA, subtitle = NA,
  save = FALSE, subdir = NA, file_name = "viz_ncuts.png")
}
\arguments{
\item{score}{Vector. Predicted value or model's result}

\item{splits}{Integer. Numer of separations to plot}

\item{model_name}{Character. Model's name}

\item{subtitle}{Character. Subtitle to show in plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\value{
Cuts by quantiles for score plot
}
\description{
Cuts by quantiles for score plot
Cuts by quantiles for score plot
}
