% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_cuts_error}
\alias{mplot_cuts_error}
\title{Cuts by quantiles on absolut and percentual errors plot}
\usage{
mplot_cuts_error(tag, score, splits = 10, title = NA,
  model_name = NA, save = FALSE, subdir = NA,
  file_name = "viz_ncuts_error.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{splits}{Integer. Numer of separations to plot}

\item{title}{Character. Title to show in plot}

\item{model_name}{Character. Model's name}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function cuts by quantiles on absolut and percentual errors
}
