% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_full}
\alias{mplot_full}
\title{MPLOTS Score Full Report Plots}
\usage{
mplot_full(tag, score, splits = 8, subtitle = NA, model_name = NA,
  save = FALSE, subdir = NA, file_name = "viz_full.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{splits}{Integer. Numer of separations to plot}

\item{subtitle}{Character. Subitle to show in plot}

\item{model_name}{Character. Model's name}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function plots a whole dashboard with a model's results. It will automatically
detect if it's a categorical or regression's model by checking how many different
unique values the independent variable (tag) has.
}
