% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_gain}
\alias{mplot_gain}
\title{Cumulative Gain Plot}
\usage{
mplot_gain(tag, score, target = TRUE, splits = 10,
  highlight = "auto", caption = NA, save = FALSE, subdir = NA,
  file_name = "viz_gain.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{target}{Value. Which is your target positive value? If 
set to 'auto', the target with largest mean(score) will be 
selected. Change the value to overwrite.}

\item{splits}{Integer. Numer of quantiles to split the data}

\item{highlight}{Character or Integer. Which split should be used
for the automatic conclussion in the plot? Set to "auto" for
best value, "none" to turn off or the number of split.}

\item{caption}{Character. Caption to show in plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
The cumulative gains plot, often named ‘gains plot’, helps us 
answer the question: When we apply the model and select the best 
X deciles, what % of the actual target class observations can we 
expect to target?
}
