% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_importance}
\alias{mplot_importance}
\title{Variables Importances Plot
Variables Importances Plot}
\usage{
mplot_importance(var, imp, colours = NA, limit = 15, model_name = NA,
  subtitle = NA, save = FALSE, subdir = NA,
  file_name = "viz_importance.png")
}
\arguments{
\item{var}{Vector. Variable or column's names}

\item{imp}{Vector. Importance of said variables. Must have same length as var}

\item{colours}{If possitive and negative contribution is known}

\item{limit}{Integer. Limit how many variavles you wish to plot}

\item{model_name}{Character. Model's name}

\item{subtitle}{Character. Subtitle to show in plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
Variables Importances Plot
Variables Importances Plot
}
