% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_importance}
\alias{mplot_importance}
\title{Variables Importances Plot}
\usage{
mplot_importance(var, imp, colours = NA, limit = 15, model_name = NA,
  subtitle = NA, save = FALSE, subdir = NA,
  file_name = "viz_importance.png")
}
\arguments{
\item{var}{Vector. Variable or column's names}

\item{imp}{Vector. Importance of said variables. Must have same length as var}

\item{colours}{If possitive and negative contribution is known}

\item{limit}{Integer. Limit how many variavles you wish to plot}

\item{model_name}{Character. Model's name}

\item{subtitle}{Character. Subtitle to show in plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function plots Variable Importances
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{clusterKmeans}}, \code{\link{conf_mat}},
  \code{\link{export_results}}, \code{\link{gain_lift}},
  \code{\link{h2o_automl}}, \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{iter_seeds}},
  \code{\link{model_metrics}}, \code{\link{mplot_conf}},
  \code{\link{mplot_cuts_error}}, \code{\link{mplot_cuts}},
  \code{\link{mplot_density}}, \code{\link{mplot_full}},
  \code{\link{mplot_gain}}, \code{\link{mplot_lineal}},
  \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{msplit}}

Other Visualization: \code{\link{corr_plot}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{theme_lares2}},
  \code{\link{theme_lares}}
}
\concept{Machine Learning}
\concept{Visualization}
