% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_lineal}
\alias{mplot_lineal}
\title{Linear Regression Results Plot}
\usage{
mplot_lineal(tag, score, regression = FALSE, subtitle = NA,
  model_name = NA, save = FALSE, subdir = NA,
  file_name = "viz_lineal.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{regression}{Boolean. Plot with geom_smooth's lm regression?}

\item{subtitle}{Character. Subitle to show in plot}

\item{model_name}{Character. Model's name}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function plots a Linear Regression Result
}
