% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{mplot_roc}
\alias{mplot_roc}
\title{ROC Curve Plot
ROC Curve with AUC value and 95\% confidence range Plot}
\usage{
mplot_roc(tag, score, model_name = NA, subtitle = NA, interval = 0.2,
  plotly = FALSE, save = FALSE, subdir = NA,
  file_name = "viz_roc.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{model_name}{Character. Model's name}

\item{subtitle}{Character. Subtitle to show in plot}

\item{interval}{Numeric. Interval for breaks in plot}

\item{plotly}{Boolean. Use plotly for plot's output for an interactive plot}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\value{
ROC Curve with AUC value and 95% confidence range Plot
}
\description{
ROC Curve Plot
ROC Curve with AUC value and 95\% confidence range Plot
}
