% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{msplit}
\alias{msplit}
\title{Split a dataframe for training and testing sets}
\usage{
msplit(df, size = 0.7, seed = 0, print = T)
}
\arguments{
\item{df}{Dataframe to split}

\item{size}{Numeric. Split rate value, between 0 and 1. If set to
1, the train and test set will be the same.}

\item{seed}{Seed for random split}

\item{print}{Print summary results}
}
\value{
A list with both datasets, summary, and split rate
}
\description{
This function splits automatically a dataframe into train and 
test datasets. You can define a seed to get the same results 
every time, but has a default value. You can prevent it from 
printing the split counter result.
}
