% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{nas}
\alias{nas}
\title{Analyze NAs in a data.frame}
\usage{
nas(df, print = FALSE)
}
\arguments{
\item{df}{Dataframe. Dataframe to study}

\item{print}{Boolean. Do you wish to print results?}
}
\description{
This function lets the user analyze NAs in a data.frame using 
VIM and funModeling libraries
}
