% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{noPlot}
\alias{noPlot}
\title{Plot Result with Nothing to Plot}
\usage{
noPlot(message = "Nothing to show here!")
}
\arguments{
\item{message}{Character. What message do you wish to show?}
}
\description{
This function lets the user print a plot without plot, with a 
customizable message. It is quite useful for Shiny renderPlot when
using filters and no data is returned.
}
