% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{noPlot}
\alias{noPlot}
\title{Plot Result with Nothing to Plot}
\usage{
noPlot(
  message = "Nothing to show here!",
  size = 4,
  font = getOption("lares.font")
)
}
\arguments{
\item{message}{Character. What message do you wish to show?}

\item{size}{Numeric. Text size.}

\item{font}{Character. Font name}
}
\description{
This function lets the user print a plot without plot, with a 
customizable message. It is quite useful for Shiny renderPlot when
using filters and no data is returned.
}
\examples{
options("lares.font" = NA) # Temporal
noPlot(message = "No plot to show!")
}
\seealso{
Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{gg_pie}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Visualization}
