% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{noPlot}
\alias{noPlot}
\title{Plot Result with Nothing to Plot}
\usage{
noPlot(message = "Nothing to show here!")
}
\arguments{
\item{message}{Character. What message do you wish to show?}
}
\description{
This function lets the user print a plot without plot, with a 
customizable message. It is quite useful for Shiny renderPlot when
using filters and no data is returned.
}
\seealso{
Other Visualization: \code{\link{corr_plot}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{plot_survey}},
  \code{\link{textCloud}}, \code{\link{theme_lares2}},
  \code{\link{theme_lares}}
}
\concept{Visualization}
