% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{normalize}
\alias{normalize}
\title{Normalize values}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric Vector. Numbers to be transformed into 
normalized vector}
}
\description{
This function lets the user normalize numerical values into 
the 0 to 1 range
}
