% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{normalize}
\alias{normalize}
\title{Normalize values}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric Vector. Numbers to be transformed into 
normalized vector}
}
\description{
This function lets the user normalize numerical values into 
the 0 to 1 range
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{date_feats}}, \code{\link{dateformat}},
  \code{\link{formatNum}}, \code{\link{holidays}},
  \code{\link{impute}}, \code{\link{left}},
  \code{\link{numericalonly}}, \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
