% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{num_abbr}
\alias{num_abbr}
\title{Abbreviate numbers}
\usage{
num_abbr(x, n = 3)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Integer. Single numeric value, specifying number of 
significant figures to show. Range 1 to 6.}
}
\value{
Vector of character values that contain converted values
}
\description{
This function converts a numeric vector's values into their
abbreviated character equivalent, i.e. 100,000,000 into 100M.
}
\examples{
num_abbr(rnorm(10) * 1e6)
num_abbr(rnorm(10) * 1e6, n = 1)
}
