% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{numericalonly}
\alias{numericalonly}
\title{Filter only Numerical Values and}
\usage{
numericalonly(df, dropnacols = TRUE, logs = FALSE, natransform = NA)
}
\arguments{
\item{df}{Data.frame}

\item{dropnacols}{Boolean. Drop columns with only NA values?}

\item{logs}{Boolean. Calculate log(x)+1 for numerical columns?}

\item{natransform}{String. "mean" or 0 to impute NA values. If
set to NA no calculation will run.}
}
\description{
This function lets the user remove all rows that have some or
all values as NAs
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{left}}, \code{\link{missingness}},
  \code{\link{normalize}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{removenacols}}, \code{\link{removenarows}},
  \code{\link{replaceall}}, \code{\link{right}},
  \code{\link{vector2text}}, \code{\link{year_month}},
  \code{\link{year_week}}
}
\concept{Data Wrangling}
