% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{ohe}
\alias{ohe}
\title{One Hot Encoding (Dummy Variables)}
\usage{
ohe(df, redundant = FALSE, summary = TRUE, limit = 20, sep = "_")
}
\arguments{
\item{df}{Dataframe}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. It the
column only has two different values, should we keep both new columns?}

\item{summary}{Boolean. Print a summary of the operations?}

\item{limit}{Integer. Limit OHE to columns which have equal or less 
different values than this parameter's value. Default: 20}

\item{sep}{Character. Separator's string}
}
\description{
This function lets the user automatically transform a dataframe with
categorical columns into numerical by one hot encoding technic.
}
