% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotencoding.R
\name{ohse}
\alias{ohse}
\title{One Hot Smart Encoding (Dummy Variables)}
\usage{
ohse(df, redundant = FALSE, drops = TRUE, ignore = NA,
  dates = FALSE, holidays = FALSE, country = "Colombia",
  currency_pair = NA, trim = 0, limit = 10, variance = 0.9,
  other_label = "OTHER", sep = "_", summary = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{redundant}{Boolean. Should we keep redundat columns? i.e. If the
column only has two different values, should we keep both new columns?}

\item{drops}{Boolean. Drop automatically some useless features?}

\item{ignore}{Vector or character. Which column should be ignored?}

\item{dates}{Boolean. Do you want the function to create more features
out of the date/time columns?}

\item{holidays}{Boolean. Include holidays as new columns?}

\item{country}{Character or vector. For which countries should the holidays
be included?}

\item{currency_pair}{Character. Which currency exchange do you
wish to get the history from? i.e, USD/COP, EUR/USD...}

\item{trim}{Integer. Trim names until the nth character}

\item{limit}{Integer. Limit one hot encoding to the n most frequent 
values of each column}

\item{variance}{Numeric. Drop columns with more than n variance. 
Range: 0-1. For example: if a variable contains 91 unique different
values out of 100 observations, this column will be suppressed if 
value is set to 0.9}

\item{other_label}{Character. With which text do you wish to replace 
the filtered values with?}

\item{sep}{Character. Separator's string}

\item{summary}{Boolean. Print a summary of the operations?}
}
\description{
This function lets the user automatically transform a dataframe with
categorical columns into numerical by one hot encoding technic.
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{date_feats}}, \code{\link{dateformat}},
  \code{\link{formatNum}}, \code{\link{holidays}},
  \code{\link{impute}}, \code{\link{left}},
  \code{\link{normalize}}, \code{\link{numericalonly}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}

Other Feature Engineering: \code{\link{date_feats}},
  \code{\link{holidays}}
}
\concept{Data Wrangling}
\concept{Feature Engineering}
