% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{one_hot_encoding_commas}
\alias{one_hot_encoding_commas}
\title{One Hot Encoding for a Vector with Comma Separated Values}
\usage{
one_hot_encoding_commas(df, variables, sep = ",")
}
\arguments{
\item{df}{Dataframe. May contain one or more columns with comma separated
values which will be separated as one hot encoding}

\item{variables}{Character. Which variables should split into new columns?}

\item{sep}{Character. Which character separates the elements?}
}
\description{
This function lets the user do one hot encoding on a variable with 
comma separated values
}
