% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{one_hot_encoding_commas}
\alias{one_hot_encoding_commas}
\title{One Hot Encoding for a Vector with Comma Separated Values}
\usage{
one_hot_encoding_commas(df, variables, sep = ",")
}
\arguments{
\item{df}{Dataframe. May contain one or more columns with comma separated
values which will be separated as one hot encoding}

\item{variables}{Character. Which variables should split into new columns?}

\item{sep}{Character. Which character separates the elements?}
}
\description{
This function lets the user do one hot encoding on a variable with 
comma separated values
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{cleanText}}, \code{\link{date_feats}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{formatTime}}, \code{\link{holidays}},
  \code{\link{impute}}, \code{\link{left}},
  \code{\link{normalize}}, \code{\link{numericalonly}},
  \code{\link{ohse}}, \code{\link{rbind_full}},
  \code{\link{removenacols}}, \code{\link{removenarows}},
  \code{\link{replaceall}}, \code{\link{right}},
  \code{\link{textFeats}}, \code{\link{textTokenizer}},
  \code{\link{vector2text}}, \code{\link{year_month}},
  \code{\link{year_week}}
}
\concept{Data Wrangling}
