% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlier_zscore_plot}
\alias{outlier_zscore_plot}
\title{Outliers: Z-score method plot}
\usage{
outlier_zscore_plot(df, var, thresh = c(2, 3, 5), mad = FALSE, plot = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{var}{Numeric variable}

\item{thresh}{Numeric vector. Z-Score threshold for n standard deviations.}

\item{mad}{Boolean. Use median absolute deviation instead?}

\item{plot}{Boolean. Show plot?}
}
\description{
Test several Z-score thresholds to visualize outliers. Tidyverse
friendly.
}
\seealso{
Other Outliers: 
\code{\link{outlier_turkey}()},
\code{\link{outlier_zscore}()}
}
\concept{Outliers}
