% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{pass}
\alias{pass}
\title{Pass Through a dplyr's Pipeline}
\usage{
pass(df, fun)
}
\arguments{
\item{df}{Dataframe}

\item{fun}{Function. What function do you wish to run? For example:
pass(. \%>\% ncol \%>\% print)}
}
\description{
This function lets the user print, save or do something inside a
pipeline without affecting the output or pipeline.
}
