% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.h2o_automl}
\alias{print.h2o_automl}
\title{Plot methods for lares}
\usage{
\method{print}{h2o_automl}(x, ...)
}
\arguments{
\item{x}{\code{h2o_automl} object}

\item{...}{Additional parameters}
}
\description{
Plot methods for lares
}
