% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{plot_cats}
\alias{plot_cats}
\title{Plot All Categorical Features (Frequencies)}
\usage{
plot_cats(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Plot. Result of \code{df} categorical features.
}
\description{
This function filters categorical columns and plots the frequency
for each value on every feature.
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{lasso_vars}()},
\code{\link{missingness}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{tree_var}()}
}
\concept{Exploratory}
