% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{plot_chord}
\alias{plot_chord}
\title{Chords Plot}
\usage{
plot_chord(origin, dest, weight = 1, title = "Chord Diagram", mg = 7)
}
\arguments{
\item{origin, dest}{Vectors. Origin and destination vectors}

\item{weight}{Vector. Weight for each chor}

\item{title}{Character. Title for the plot}

\item{mg}{Numeric. Margin adjust for plot in case of need}
}
\description{
This function plots discrete and continuous values results
}
