% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{plot_chord}
\alias{plot_chord}
\title{Chords Plot}
\usage{
plot_chord(
  origin,
  dest,
  weight = 1,
  mg = 3,
  title = "Chord Diagram",
  subtitle = "",
  pal = NA
)
}
\arguments{
\item{origin, dest}{Vectors. Origin and destination vectors}

\item{weight}{Vector. Weight for each chor}

\item{mg}{Numeric. Margin adjust for plot in case of need'}

\item{title}{Character. Title for the plot}

\item{subtitle}{Character. Subtitle for the plot}

\item{pal}{Vector. Colour pallete. Order matters.}
}
\description{
This function plots discrete and continuous values results
}
