% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{plot_df}
\alias{plot_df}
\title{Plot Summary of Numerical and Categorical Features}
\usage{
plot_df(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
Plot. Result of \code{df} categorical and numerical features.
}
\description{
This function plots all columns frequencies and boxplots, for
categorical and numerical respectively.
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{lasso_vars}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_nums}()},
\code{\link{tree_var}()},
\code{\link{trendsRelated}()}
}
\concept{Exploratory}
