% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.R
\name{plot_distr}
\alias{plot_distr}
\title{Plot Target's Distribution vs Another Variable}
\usage{
plot_distr(data, target, values, top = 10, breaks = 10, abc = FALSE,
  truncate = 0, clean = FALSE, na.rm = FALSE,
  custom_colours = FALSE, print = FALSE, save = FALSE, subdir = NA)
}
\arguments{
\item{data}{Dataframe}

\item{target}{Character. Name of the Main -target- variable}

\item{values}{Character. Name of the Secondary variable}

\item{top}{Integer. Filter and plot the most n frequent for categorical values}

\item{breaks}{Integer. Number of splits for numerical values}

\item{abc}{Boolean. Do you wish to sort by alphabetical order?}

\item{truncate}{Integer. Truncate until the n character for categorical values}

\item{clean}{Boolean. Use lares::cleanText for categorical values}

\item{na.rm}{Boolean. Ignore NAs if needed}

\item{custom_colours}{Boolean. Use custom colours function?}

\item{print}{Boolean. Print the table's result}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}
}
\description{
Study the distribution of a target variable vs another variable. This
function is quite similar to the funModeling's corrplot function.
}
