% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_str.R
\name{plot_nums}
\alias{plot_nums}
\title{Plot All Numerical Features (Boxplots)}
\usage{
plot_nums(df)
}
\arguments{
\item{df}{Dataframe}
}
\description{
This function filters numerical columns and plots boxplots.
}
\seealso{
Other Exploratory: \code{\link{corr_cross}},
  \code{\link{corr_var}}, \code{\link{crosstab}},
  \code{\link{df_str}}, \code{\link{distr}},
  \code{\link{freqs_df}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{missingness}}, \code{\link{plot_cats}},
  \code{\link{plot_df}}, \code{\link{tree_var}},
  \code{\link{trendsRelated}}
}
\concept{Exploratory}
