% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colours}
\usage{
plot_palette(fill, colour = "black", id = NA)
}
\arguments{
\item{fill}{Vector. List of colours for fills}

\item{colour}{Vector. List of colours for colours}

\item{id}{Vector. ID for each color}
}
\description{
This function plots a list of colours
}
\examples{
pal <- lares_pal()$palette[1:15]
head(pal)
plot_palette(fill = names(pal), colour = as.vector(pal))
}
\seealso{
Other Auxiliary: 
\code{\link{gg_colour_customs}()},
\code{\link{gg_fill_customs}()},
\code{\link{gg_text_customs}()},
\code{\link{lares_pal}()}
}
\concept{Auxiliary}
