% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colours}
\usage{
plot_palette(vector)
}
\arguments{
\item{vector}{Vector. List of colours}
}
\description{
This function plots a list of colours
}
