% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette Colours}
\usage{
plot_palette(fill, colour = "black", id = NA)
}
\arguments{
\item{fill}{Vector. List of colours for fills}

\item{colour}{Vector. List of colours for colours}

\item{id}{Vector. ID for each color}
}
\description{
This function plots a list of colours
}
