% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{plot_survey}
\alias{plot_survey}
\title{Send Emails with Attachments (Sendgrid)}
\usage{
plot_survey(answers, ignore = c(1), title = NA, subtitle = NA)
}
\arguments{
\item{answers}{Dataframe. Answers. Each row a different person. 
Each column a different answer.}

\item{ignore}{Numeric Vector. Which columns are NOT answers?}

\item{title}{Character. Title for your plot}

\item{subtitle}{Character. Subtitle for your plot}
}
\description{
This function lets the user send Emails with Attachments using Sendgrid
}
