% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{plot_survey}
\alias{plot_survey}
\title{Visualize Survey Results}
\usage{
plot_survey(answers, ignore = c(1), title = NA, subtitle = NA)
}
\arguments{
\item{answers}{Dataframe. Answers. Each row a different person. 
Each column a different answer.}

\item{ignore}{Numeric Vector. Which columns are NOT answers?}

\item{title}{Character. Title for your plot}

\item{subtitle}{Character. Subtitle for your plot}
}
\description{
This function lets the user plot a survey's result.
}
\seealso{
Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{mplot_conf}()},
\code{\link{mplot_cuts_error}()},
\code{\link{mplot_cuts}()},
\code{\link{mplot_density}()},
\code{\link{mplot_full}()},
\code{\link{mplot_gain}()},
\code{\link{mplot_importance}()},
\code{\link{mplot_lineal}()},
\code{\link{mplot_metrics}()},
\code{\link{mplot_response}()},
\code{\link{mplot_roc}()},
\code{\link{mplot_splits}()},
\code{\link{noPlot}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Visualization}
