% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Density plot for discrete and continuous values}
\usage{
plot_timeline(event, start, end, label = NA, group = NA,
  title = "Curriculum Vitae Timeline", subtitle = "Bernardo Lares",
  size = 7, colour = "orange", save = FALSE, subdir = NA)
}
\arguments{
\item{event}{Vector. Event, role, label, or row.}

\item{start}{Vector. Start date.}

\item{end}{Vector. End date.}

\item{label}{Vector. Place, institution, or label.}

\item{group}{Vector. Academic, Work, Extracurricular...}

\item{title}{Character. Title for the plot}

\item{subtitle}{Character. Subtitle for the plot}

\item{size}{Numeric. Bars' width}

\item{colour}{Character. Colour when not using type}

\item{save}{Boolean. Save the output plot in our working directory}

\item{subdir}{Character. Into which subdirectory do you wish to save the plot to?}
}
\description{
This function plots discrete and continuous values results
}
