% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_daily}
\alias{portfolio_daily}
\title{Portfolio Daily Performance}
\usage{
portfolio_daily(data, dailys, cash_fix = 0)
}
\arguments{
\item{data}{Dataframe. Result from get_stocks()}

\item{dailys}{Dataframe. Result from get_stocks_hist()}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value 
to your cash balance}
}
\description{
This function lets the user calculate daily portfolio performance
}
