% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{portfolio_total_plot}
\alias{portfolio_total_plot}
\title{Portfolio's Daily Cumulative}
\usage{
portfolio_total_plot(portfolio, save = FALSE)
}
\arguments{
\item{portfolio}{Dataframe. Results from portfolio_daily()}

\item{save}{Boolean. Export plot as an image?}
}
\description{
This function lets the user plot his portfolio's daily cumulative
}
