% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{print.h2o_automl}
\alias{print.h2o_automl}
\title{Print methods for lares}
\usage{
\method{print}{h2o_automl}(x, importance = TRUE, ...)
}
\arguments{
\item{x}{Object}

\item{importance}{Boolean. Print important variables?}

\item{...}{Additional parameters}
}
\description{
Print methods for lares
}
