% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{prophesize}
\alias{prophesize}
\title{Facebook's Prophet Forecast}
\usage{
prophesize(df, n_future = 120, country = "AR", pout = 0.03,
  project = "Prophet Forecast")
}
\arguments{
\item{df}{Data frame. Must contain date/time column and values column}

\item{n_future}{Integer. How many steps do you wish to forecast?}

\item{country}{Character. Country code for holidays}

\item{pout}{Numeric. Get rid of pout \% of outliers}

\item{project}{Character. Name of your forecast project for plot title}
}
\description{
Prophet is Facebook's procedure for forecasting time series data 
based on an additive model where non-linear trends are fit with 
yearly, weekly, and daily seasonality, plus holiday effects. It 
works best with time series that have strong seasonal effects and 
several seasons of historical data. Prophet is robust to missing 
data and shifts in the trend, and typically handles outliers well.
}
\details{
Official documentation: \url{https://github.com/facebook/prophet}
}
