% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{quants}
\alias{quants}
\title{Calculate cuts by quantiles}
\usage{
quants(values, splits = 10)
}
\arguments{
\item{values}{Vector. Values to calculate quantile cuts}

\item{splits}{Integer. How many cuts should split the values?}
}
\description{
This function lets the user quickly calculate cuts for quantiles
}
