% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryDB}
\alias{queryDB}
\title{PostgreSQL Queries on Database (Read)}
\usage{
queryDB(query, from = "dummy", creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{from}{Character. Credential's user (see `get_credentials`)}

\item{creds}{Character. Credential's directory (see `get_credentials`)}
}
\description{
This function lets the user query a PostgreSQL database. Previously
was called `queryDummy()` but was replaced and deprecated for a 
more general function by using the `from` parameter.
}
\seealso{
Other Credentials: 
\code{\link{db_download}()},
\code{\link{f1_contacts}()},
\code{\link{geoAddress}()},
\code{\link{get_credentials}()},
\code{\link{get_tweets}()},
\code{\link{mailSend}()},
\code{\link{queryGA}()},
\code{\link{stocks_report}()}
}
\concept{Credentials}
\concept{Database}
