% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryDW}
\alias{queryDW}
\title{PostgreSQL Queries on Redshift Database (read-write)}
\usage{
queryDW(query, which = "seguros", creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{which}{Character. Which database do you wish to connect to: seguros, tabunga...}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user query our Data Warehouse Database
}
