% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryDW}
\alias{queryDW}
\title{PostgreSQL Queries on Redshift Database (read-write)}
\usage{
queryDW(query, which = "soat", creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{which}{Character. Which database do you wish to connect to: seguros, tabunga...}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user query our Data Warehouse Database
}
\seealso{
Other Connections: 
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{f1_contacts}()},
\code{\link{geoAddress}()},
\code{\link{li_auth}()},
\code{\link{queryDummy}()},
\code{\link{queryProduc}()},
\code{\link{typeform_download}()}
}
\concept{Connections}
