% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics.R
\name{queryGA}
\alias{queryGA}
\title{Queries on Google Analytics}
\usage{
queryGA(account = "comparamejor", creds = NA, token_dir = NA,
  metrics = "sessions", dimensions = "date",
  start = lubridate::floor_date(Sys.Date(), "month"), end = Sys.Date())
}
\arguments{
\item{account}{Character. Personal named accounts}

\item{creds}{Character. Credential's user (see get_credentials)}

\item{metrics}{Character. Which metrics we sih to bring}

\item{dimensions}{Character. Which dimensions we sih to bring}

\item{start}{Date. Start date for the report}

\item{end}{Date. End date for the report}
}
\description{
This function lets the user query Google Analytics with its API
}
