% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_analytics.R
\name{queryGA}
\alias{queryGA}
\title{Queries on Google Analytics}
\usage{
queryGA(
  account = "comparamejor",
  creds = NA,
  token_dir = NA,
  metrics = "sessions",
  dimensions = "date",
  met_filters = NULL,
  dim_filters = NULL,
  start = lubridate::floor_date(Sys.Date(), "month"),
  end = Sys.Date()
)
}
\arguments{
\item{account}{Character. Personal named accounts}

\item{creds}{Character. Credential's user (see get_credentials)}

\item{token_dir}{Character. Credential's user (see get_credentials)}

\item{metrics}{Character. Which metrics we wish to bring}

\item{dimensions}{Character. Which dimensions we wish to bring}

\item{met_filters, dim_filters}{A filter_clause_ga4 for filtering 
metrics/dimensions. Check googleAnalyticsR::google_analytics()}

\item{start}{Date. Start date for the report}

\item{end}{Date. End date for the report}
}
\description{
This function lets the user query Google Analytics with its API.
More about the documentation and parameters in googleAnalyticsR's
google_analytics or Google Analytics'
\url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}{API}
}
\seealso{
Other Connection: 
\code{\link{get_tweets}()},
\code{\link{readGS4}()},
\code{\link{readGS}()},
\code{\link{writeGS}()}
}
\concept{Connection}
