% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querieSQL.R
\name{queryProduc}
\alias{queryProduc}
\title{PostgreSQL Queries on Production Database (read-write)}
\usage{
queryProduc(query, creds = NA)
}
\arguments{
\item{query}{Character. SQL Query}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user query our Production Database
}
\seealso{
Other Connections: \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{f1_contacts}}, \code{\link{geoAddress}},
  \code{\link{li_auth}}, \code{\link{queryDW}},
  \code{\link{queryDummy}}, \code{\link{typeform_download}}
}
\concept{Connections}
