% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{quiet}
\alias{quiet}
\title{Import Excel File with All Its Tabs}
\usage{
quiet(fx)
}
\arguments{
\item{fx}{Function}
}
\description{
This function silences (verbose) output prints. Thanks to Hadley Wickham!
}
