% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{quiet}
\alias{quiet}
\title{Quiet prints and verbose noice}
\usage{
quiet(fx)
}
\arguments{
\item{fx}{Function to quiet}
}
\description{
This function silences (verbose) output prints.
}
\author{
Hadley Wickham
}
