% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{quiet}
\alias{quiet}
\title{Quiet prints and verbose noice}
\usage{
quiet(fx)
}
\arguments{
\item{fx}{Function to quiet}
}
\description{
This function silences (verbose) output prints.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{read.file}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\author{
Hadley Wickham
}
\concept{Tools}
