% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbind.R
\name{rbind_full}
\alias{rbind_full}
\title{Smart rbind}
\usage{
rbind_full(..., list = NA, fill = NA, sep = ":")
}
\arguments{
\item{...}{Dataframes. Dataframes to bind}

\item{list}{List. Containing dataframes to combine}

\item{fill}{Value. To use when 'filling' missing columns}

\item{sep}{Character. String used to separate column names when 
pasting them together.}
}
\description{
Efficient rbind of data frames, even if the column names don't match.
Code based on gtools::smartbind
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
