% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{read.file}
\alias{read.file}
\title{Read Files (Auto-detected)}
\usage{
read.file(filename, current_wd = TRUE)
}
\arguments{
\item{filename}{Character}

\item{current_wd}{Boolean. Use current working directory before
the file's name? Use this param to NOT get absolute root directory.}
}
\description{
This function lets the user import csv, xlsx, xls, sav files.
}
