% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{read.file}
\alias{read.file}
\title{Read Files (Auto-detected)}
\usage{
read.file(filename, current_wd = TRUE)
}
\arguments{
\item{filename}{Character}

\item{current_wd}{Boolean. Use current working directory before
the file's name? Use this param to NOT get absolute root directory.}
}
\description{
This function lets the user import csv, xlsx, xls, sav files.
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{statusbar}}, \code{\link{zerovar}}
}
\concept{Tools}
