% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{read.file}
\alias{read.file}
\title{Read Files (Auto-detected)}
\usage{
read.file(filename, current_wd = TRUE)
}
\arguments{
\item{filename}{Character}

\item{current_wd}{Boolean. Use current working directory before
the file's name? Use this param to NOT get absolute root directory.}
}
\description{
This function lets the user import csv, xlsx, xls, sav files.
}
\seealso{
Other Tools: \code{\link{autoline}},
  \code{\link{bindfiles}}, \code{\link{bring_api}},
  \code{\link{db_download}}, \code{\link{db_upload}},
  \code{\link{export_plot}}, \code{\link{export_results}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}},
  \code{\link{h2o_predict_API}},
  \code{\link{h2o_predict_MOJO}},
  \code{\link{h2o_predict_binary}},
  \code{\link{h2o_predict_model}},
  \code{\link{h2o_selectmodel}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{iter_seeds}},
  \code{\link{json2vector}}, \code{\link{listfiles}},
  \code{\link{mailSend}}, \code{\link{matrixwd}},
  \code{\link{msplit}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{statusbar}}, \code{\link{try_require}},
  \code{\link{updateLares}}, \code{\link{zerovar}}
}
\concept{Tools}
