% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS}
\alias{readGS}
\alias{readGS4}
\title{Google Sheets Reading (API v4)}
\usage{
readGS(title, sheet = "Hoja 1", range = NULL, creds = NULL, ...)

readGS4(title, sheet = "Hoja 1", range = NULL, creds = NULL, ...)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{creds}{Character. JSON filename with service auth}

\item{...}{Further read_sheet parameters}
}
\description{
This function lets the user read any Google Sheet's data
}
\seealso{
Other Connection: 
\code{\link{get_tweets}()},
\code{\link{queryGA}()}

Other Scrapper: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{trendsRelated}()}
}
\concept{Connection}
\concept{Scrapper}
