% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS}
\alias{readGS}
\alias{readGS4}
\title{####################################################################
#' Google Sheets Reading [Deprecated]
#' 
#' This function lets the user read any Google Sheet's data
#' 
#' @family Connection
#' @family Scrapper
#' @param title Character. Textual title of Google Sheet
#' @param ws Character. Working sheet to import
#' @param first_time Boolean. Authenticate manualy
#' @export
readGS <- function(title, ws = "Hoja 1", first_time = FALSE) {
  try_require("googlesheets")
  if (first_time) {
    options(httr_oob_default = TRUE)
    gs_auth(new_user = TRUE)
  }
  gs <- gs_title(title)
  gs <- invisible(gs_read(gs, ws = ws, verbose = FALSE) %>% data.frame())
  return(gs)
}
####################################################################
#' Google Sheets Writing [Deprecated]
#' 
#' This function lets the user write on any Google Sheet
#' 
#' @family Connection
#' @family Scrapper
#' @param data Data Frame. Data to export to Google Sheet 
#' @param title Character. Textual title of Google Sheet
#' @param ws Character. Working sheet to export to
#' @param cell Character. In which cell should we paste the data (upper left cell)
#' @param first_time Boolean. Authenticate manualy
#' @export
writeGS <- function(data, title, ws = "Hoja 1", cell = 'A1', first_time = FALSE) {
  try_require("googlesheets")
  if (first_time) {
    options(httr_oob_default = TRUE)
    gs_auth(new_user = TRUE)
  }
  gs <- gs_title(title)
  invisible(gs_edit_cells(
    gs, ws = ws, input = data, anchor = cell, verbose = FALSE) %>% 
      data.frame())
}
Google Sheets Reading (API v4)}
\usage{
readGS(title, sheet = "Hoja 1", range = NULL, creds = NULL, ...)

readGS4(title, sheet = "Hoja 1", range = NULL, creds = NULL, ...)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{creds}{Character. JSON filename with service auth}

\item{...}{Further read_sheet parameters}
}
\description{
This function lets the user read any Google Sheet's data
}
\seealso{
Other Connection: 
\code{\link{get_tweets}()},
\code{\link{queryGA}()}

Other Scrapper: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{trendsRelated}()}
}
\concept{Connection}
\concept{Scrapper}
