% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS}
\alias{readGS}
\title{Google Sheets Reading}
\usage{
readGS(title, ws = "Hoja 1", first_time = FALSE)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{ws}{Character. Working sheet to import}

\item{first_time}{Boolean. Authenticate manualy}
}
\description{
This function lets the user read any Google Sheet's data
}
\seealso{
Other Connection: \code{\link{get_tweets}},
  \code{\link{queryGA}}, \code{\link{writeGS}}
}
\concept{Connection}
