% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{readGS}
\alias{readGS}
\alias{readGS4}
\title{Google Sheets Reading (API v4)}
\usage{
readGS(
  title,
  sheet = "Hoja 1",
  range = NULL,
  json = NULL,
  email = NULL,
  server = FALSE,
  ...
)
}
\arguments{
\item{title}{Character. Textual title of Google Sheet}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{json}{Character. JSON filename with service auth}

\item{email}{Character. If you have multiple pre-authorized
accounts in your machine, you may non-interactively select
which one you wish to use.}

\item{server}{Boolean. Force interacting auth process?}

\item{...}{Further read_sheet parameters}
}
\description{
Read data from a Google Sheets.
}
\seealso{
Other Scrapper: 
\code{\link{filesGD}()},
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{trendsRelated}()},
\code{\link{writeGS}()}

Other Google: 
\code{\link{filesGD}()},
\code{\link{geoAddress}()},
\code{\link{queryGA}()},
\code{\link{trendsRelated}()},
\code{\link{trendsTime}()},
\code{\link{writeGS}()}
}
\concept{Google}
\concept{Scrapper}
