% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{removenacols}
\alias{removenacols}
\title{Remove/Drop Columns in which ALL or SOME values are NAs}
\usage{
removenacols(df, all = TRUE, ignore = NULL)
}
\arguments{
\item{df}{Data.frame}

\item{all}{Boolean. Remove columns containing ONLY \code{NA} values.
If set to \code{FALSE}, remove columns containing at least one \code{NA}.}

\item{ignore}{Character vector. Column names to ignore validation.}
}
\value{
data.frame with removed columns.
}
\description{
This function lets the user remove all columns that have some or
all values as NAs
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{formatNum}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
