% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{removenarows}
\alias{removenarows}
\title{Remove/Drop Rows in which ALL or SOME values are NAs}
\usage{
removenarows(df, all = TRUE)
}
\arguments{
\item{df}{Data.frame}

\item{all}{Boolean. Remove rows which contains ONLY NA values.
If set to FALSE, rows which contains at least one NA will be removed}
}
\value{
data.frame with removed rows.
}
\description{
This function lets the user remove all rows that have some or
all values as NAs
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{formatNum}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{removenacols}()},
\code{\link{replaceall}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
