% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{replaceall}
\alias{replaceall}
\title{Replace Values With}
\usage{
replaceall(df, original, change)
}
\arguments{
\item{df}{Data.frame}

\item{original}{String. Original text you wish to replace}

\item{change}{String. Values you wish to replace the originals with}
}
\description{
This function lets the user replace all specific values in a 
vector or data.frame into another value.
}
