% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{replaceall}
\alias{replaceall}
\title{Replace Values With}
\usage{
replaceall(df, original, change, quiet = TRUE)
}
\arguments{
\item{df}{Data.frame}

\item{original}{String or Vector. Original text you wish to replace}

\item{change}{String or Vector. Values you wish to replace the originals with}

\item{quiet}{Boolean. Keep quiet? (or print replacements)}
}
\description{
This function lets the user replace all specific values in a 
vector or data.frame into another value. If replacing more than
one value, order matters so they will be replaced in the same
order that you pass them to the function.
}
