% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
rmse(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate Root Mean Squared Error
}
