% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robyn.R
\name{robyn_modelselector}
\alias{robyn_modelselector}
\alias{plot.robyn_modelselector}
\title{Robyn: Model Selection Viz}
\usage{
robyn_modelselector(
  InputCollect,
  OutputCollect,
  n_per_cluster = 5,
  wt = c(2, 1, 1, 1, 0.1),
  quiet = FALSE,
  cache = TRUE,
  ...
)

\method{plot}{robyn_modelselector}(x, ...)
}
\arguments{
\item{InputCollect, OutputCollect}{Robyn output objects.}

\item{n_per_cluster}{Integer. How many models per cluster do you
want to plot? Default: 5. Keep in mind they will all be considered
for the calculations.}

\item{wt}{Vector. Length 5. Pick the weight for each of the normalized
metrics to calculate the score and rank models. In order: R2, Performance
(ROAS or CPA), Potential Performance Boost, Non-Zero Coefficients, Amount
of models in a cluster.}

\item{quiet}{Boolean. Keep quiet? If not, show messages}

\item{cache}{Use cache functionality for allocator's results?}

\item{...}{Additional parameters passed to \code{corr} and \code{cor.test}}

\item{x}{robyn_modelselector object}
}
\value{
list with data.frame and plot.
}
\description{
Consider N models per cluster to select the right one using
several metrics to consider such as potential improvement on budget
allocator and how many non-zero coefficients there are.
}
\seealso{
Other Robyn: 
\code{\link{hyps_builder}()}
}
\concept{Robyn}
