% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{ROC}
\alias{ROC}
\title{ROC Curves}
\usage{
ROC(tag, score, multis = NA)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{multis}{Data.frame. Containing columns with each category score 
(only used when more than 2 categories coexist)}
}
\description{
This function calculates ROC Curves and AUC values with 95\% confidence 
range. It also works for multi-categorical models.
}
