% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtistry.R
\name{rtistry_sphere}
\alias{rtistry_sphere}
\title{Generative Art: Sphere XmodY}
\usage{
rtistry_sphere(eye = c(100, 0, 0), pal = "auto", var = 3)
}
\arguments{
\item{eye, pal, var}{Parameters to change aesthetics and calculations}
}
\value{
ggplot object
}
\description{
Generative Art: Sphere XmodY
}
\concept{rtistry}
