% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_plots.R
\name{scale_x_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{scale_x_dollar}
\alias{scale_y_dollar}
\alias{scale_x_abbr}
\alias{scale_y_abbr}
\title{Axis scales format}
\usage{
scale_x_comma(...)

scale_y_comma(...)

scale_x_percent(...)

scale_y_percent(...)

scale_x_dollar(...)

scale_y_dollar(...)

scale_x_abbr(...)

scale_y_abbr(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::continuous_scale}}
}
\description{
The \code{_comma} ones set comma format for axis text, the \code{_percent} 
ones set percent format for axis text, \code{_dollar} for collar currency,
and \code{_abbr} for abbreviated format.
}
