% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{seq_surnames}
\alias{seq_surnames}
\title{Surnames Order Sequence}
\usage{
seq_surnames(n = 1)
}
\arguments{
\item{n}{Integer. Number of generations to include in the sequence.
Notice it will generate a vector with 2^(n-1) values.}
}
\value{
Integer vector.
}
\description{
Generate a sequence of numbers that determines the order in which
surnames should be listed based on the number of generations of
ancestors you wish to include. This sequence follows the traditional
Latin custom of assigning the father's surname first, followed by
the mother's surname. The same logic extends systematically to higher
generations, ensuring that the order of surnames remains consistent
as you move upward through the family tree.
}
\examples{
seq_surnames(1)
seq_surnames(2)
seq_surnames(3)
seq_surnames(4)
}
