% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley.R
\name{shap_var}
\alias{shap_var}
\title{SHAP-based dependence plots for categorical/numerical features (PDP)}
\usage{
shap_var(x, var, keep_outliers = FALSE)
}
\arguments{
\item{x}{\code{h2o_shap} object}

\item{var}{Variable name}

\item{keep_outliers}{Boolean. Outliers detected with z-score and 3sd
may be suppress or kept in your plot. Keep them?}
}
\description{
Having a \code{h2o_shap} object, plot a dependence plot for any
categorical or numerical feature.
}
\examples{
\dontrun{
# Train a model
model <- h2o_automl(dft, Survived, 
                    max_models = 1, 
                    target = TRUE,
                    ignore = c("Ticket", "Cabin", "PassengerId"))

# Calculate SHAP values
SHAP_values <- h2o_shap(model)
plot(SHAP_values)

# Plot some of the variables (categorical and numerical)
shap_var(SHAP_values, Pclass)
shap_var(SHAP_values, Age)
shap_var(SHAP_values, Fare)
shap_var(SHAP_values, Fare, keep_outliers = TRUE)
}
}
\seealso{
Other SHAP: 
\code{\link{h2o_shap}()},
\code{\link{plot.h2o_automl}()}
}
\concept{SHAP}
