% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{statusbar}
\alias{statusbar}
\title{Progressbar for for loops}
\usage{
statusbar(run, max.run, percent.max = 40L, info = run)
}
\arguments{
\item{run}{Iterator. for loop or an integer with the current loop number}

\item{max.run}{Number. Maximum number of loops}

\item{percent.max}{Integer. Indicates how wide the progress bar is printed}

\item{info}{String. With additionaly information to be printed 
at the end of the line. The default is \code{run}.}
}
\description{
This function lets the user view a progressbar for a 'for' loop. 
Taken from https://github.com/STATWORX/helfRlein/blob/master/R/statusbar.R
}
\seealso{
Other Tools: \code{\link{bindfiles}},
  \code{\link{bring_api}}, \code{\link{db_download}},
  \code{\link{db_upload}}, \code{\link{export_plot}},
  \code{\link{get_credentials}},
  \code{\link{get_currency}}, \code{\link{h2o_update}},
  \code{\link{haveInternet}}, \code{\link{importxlsx}},
  \code{\link{ip_country}}, \code{\link{json2vector}},
  \code{\link{listfiles}}, \code{\link{myip}},
  \code{\link{pass}}, \code{\link{quiet}},
  \code{\link{read.file}}, \code{\link{zerovar}}
}
\concept{Tools}
