% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_daily_plot}
\alias{stocks_daily_plot}
\title{Stocks Daily Plot}
\usage{
stocks_daily_plot(portfolio, daily, weighted = TRUE, group = TRUE,
  save = FALSE)
}
\arguments{
\item{portfolio}{Dataframe. Output of the portfolio_perf function}

\item{daily}{Dataframe. Daily data}

\item{weighted}{Boolean. Should variation values be weighted to the
portfolio (or simply compared with initial value)?}

\item{group}{Boolean. Group stocks by stocks type?}

\item{save}{Boolean. Export plot as an image?}
}
\description{
This function lets the user plot stocks daily change
}
\seealso{
Other Investment: \code{\link{etf_sector}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
