% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_file}
\alias{stocks_file}
\title{Get Personal Portfolio's Data}
\usage{
stocks_file(
  filename = NA,
  creds = NA,
  auto = TRUE,
  sheets = c("Portafolio", "Fondos", "Transacciones"),
  keep_old = TRUE
)
}
\arguments{
\item{filename}{Characeter. Import a local Excel file}

\item{creds}{Character. Where is my personal token for Dropbox connection?}

\item{auto}{Boolean. Automatically use my local personal file?}

\item{sheets}{Character Vector. Names of each sheet containing Portfolio summary,
Cash, and Transactions information}

\item{keep_old}{Boolean. Include sold tickers eventhough not currently in portfolio?}
}
\description{
This function lets me download my personal Excel with my Portfolio 
data, locally or from Dropbox. You can also use it for local files
}
\examples{
\dontrun{
 # Load lares' dummy data
 file <- system.file("inst/docs", "dummyPortfolio.xlsx", package = "lares")
 df <- stocks_file(filename = file, 
                   sheets = c("Portafolio","Fondos","Transacciones"), 
                   keep_old = FALSE)
 # This will create a list with the following 3 dataframes:
 names(df)
 "portfolio" "transactions" "cash" 
}
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()},
\code{\link{stocks_report}()}
}
\concept{Investment}
