% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_hist}
\alias{stocks_hist}
\title{Download Stocks Historical Data}
\usage{
stocks_hist(symbols = c("VTI", "IEMG"), from = Sys.Date() - 365,
  tax = 30, verbose = TRUE)
}
\arguments{
\item{symbols}{Character Vector. List of symbols to download historical data}

\item{from}{Date. Since when do you wish to download historical data? 
If not set, will return since 365 days ago}

\item{tax}{Numeric. How much [0-99] of your dividends are gone with taxes?}

\item{verbose}{Boolean. Print results and progress while downloading?}
}
\description{
This function lets the user download stocks historical data
}
\seealso{
Other Investment: \code{\link{daily_portfolio}},
  \code{\link{daily_stocks}}, \code{\link{etf_sector}},
  \code{\link{splot_change}}, \code{\link{splot_etf}},
  \code{\link{splot_growth}}, \code{\link{splot_roi}},
  \code{\link{splot_summary}}, \code{\link{splot_types}},
  \code{\link{stocks_file}}, \code{\link{stocks_obj}},
  \code{\link{stocks_report}}
}
\concept{Investment}
