% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_html}
\alias{stocks_html}
\title{Portfolio's Full Report in HTML}
\usage{
stocks_html(results)
}
\arguments{
\item{results}{List. Containing the following objects: portf_daily_change, 
portf_stocks_change, portf_stocks_histchange, portf_distribution & portfolio_perf.
You can use simply use the stocks_objects(data) if you didn't mess with the order!}
}
\description{
This function lets the user create his portfolio's full report in HTML using
the library's results
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks_hist}},
  \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
