% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_objects}
\alias{stocks_objects}
\title{Portfolio's Calculations and Plots}
\usage{
stocks_objects(data, cash_fix = 0, tax = 30, expenses = 7,
  sectors = TRUE)
}
\arguments{
\item{data}{List. Containing the following dataframes: portfolio,
transactions, cash. They have to follow the original xlsx format}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value 
to your cash balance}

\item{tax}{Numeric. How much does of your dividends does the taxes take? 
Range from 0 to 99}

\item{expenses}{Numeric. How much does that bank or broker charges per
transaction? Absolute value.}

\item{sectors}{Boolean. Return sectors segmentation for ETFs?}
}
\description{
This function lets the user create his portfolio's calculations and
plots for further study.
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks_hist}},
  \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
