% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_objects}
\alias{stocks_objects}
\title{Portfolio's Calculations and Plots}
\usage{
stocks_objects(data)
}
\arguments{
\item{data}{List. Containing the following dataframes: portfolio,
transactions, cash. They have to follow the original xlsx format}
}
\description{
This function lets the user create his portfolio's calculations and
plots for further study.
}
