% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_performance}
\alias{stocks_performance}
\title{Stocks Overall Performance}
\usage{
stocks_performance(dailys, cash_in, cash_fix = 0)
}
\arguments{
\item{dailys}{Dataframe. Daily values. Structure: "Date", "Symbol", 
"Open", "High", "Low", "Close", "Volume", "Adjusted", "Quant", 
"Value", "Amount", "Expenses", "Stocks", "Div", "DivReal", "DailyDiv", 
"DailyValue", "RelChangeP", 
"RelChangeUSD"}

\item{cash_in}{Dataframe. Deposits and withdrawals. Structure: "ID", "Date", "Cash"}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value 
to your cash balance}
}
\description{
This function lets the user calculate stocks performance
}
\seealso{
Other Investment: \code{\link{etf_sector_plot}},
  \code{\link{etf_sector}}, \code{\link{get_stocks_hist}},
  \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_report}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
