% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_report}
\alias{stocks_report}
\title{Portfolio's Full Report and Email}
\usage{
stocks_report(wd = "personal", cash_fix = 0,
  mail = "laresbernardo@gmail.com", creds = NA)
}
\arguments{
\item{wd}{Character. Where do you wish to save the results (plots and report)?}

\item{cash_fix}{Numeric. If you wish to algebraically sum a value to your cash balance}

\item{mail}{Boolean. Do you wish to send the email? Set to NA to not send email}

\item{creds}{Character. Credential's user (see get_credentials) for sending mail}
}
\description{
This function lets the user create his portfolio's full report with plots and email sent
}
\seealso{
Other Investment: \code{\link{etf_sector}},
  \code{\link{get_stocks_hist}}, \code{\link{get_stocks}},
  \code{\link{portfolio_daily_plot}},
  \code{\link{portfolio_daily}},
  \code{\link{portfolio_distr_plot}},
  \code{\link{portfolio_performance}},
  \code{\link{portfolio_total_plot}},
  \code{\link{stocks_daily_plot}},
  \code{\link{stocks_hist_fix}}, \code{\link{stocks_html}},
  \code{\link{stocks_objects}},
  \code{\link{stocks_performance}},
  \code{\link{stocks_total_plot}}
}
\concept{Investment}
