% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks_report}
\alias{stocks_report}
\title{Portfolio's Full Report and Email}
\usage{
stocks_report(
  data = NA,
  dir = NA,
  mail = FALSE,
  to = "laresbernardo@gmail.com",
  sectors = TRUE,
  creds = NA
)
}
\arguments{
\item{data}{Character. stocks_file() output. If NA, automatic report
parameters will be used}

\item{dir}{Character. Directory for HTML report output. If set to NA, 
current working directory will be used. If mail sent, file will be erased}

\item{mail}{Boolean. Do you want to send an email with the report attached? 
If not, an HTML file will be created in dir}

\item{to}{Character. Email to send the report to}

\item{sectors}{Boolean. Return sectors segmentation for ETFs?}

\item{creds}{Character. Credential's user (see \code{get_creds()}) for 
sending mail and Dropbox interaction}
}
\description{
This function lets the user create his portfolio's full report with 
plots and send it to an email with the HTML report attached
}
\examples{
\dontrun{
list <- stocks_obj()
stocks_report(list, dir = "~/Desktop")
}
}
\seealso{
Other Investment: 
\code{\link{daily_portfolio}()},
\code{\link{daily_stocks}()},
\code{\link{etf_sector}()},
\code{\link{splot_change}()},
\code{\link{splot_etf}()},
\code{\link{splot_growth}()},
\code{\link{splot_roi}()},
\code{\link{splot_summary}()},
\code{\link{splot_types}()},
\code{\link{stocks_file}()},
\code{\link{stocks_hist}()},
\code{\link{stocks_obj}()},
\code{\link{stocks_quote}()}

Other Credentials: 
\code{\link{db_download}()},
\code{\link{f1_contacts}()},
\code{\link{geoAddress}()},
\code{\link{get_credentials}()},
\code{\link{get_tweets}()},
\code{\link{mailSend}()},
\code{\link{queryDB}()},
\code{\link{queryGA}()},
\code{\link{slackSend}()},
\code{\link{stocks_file}()}
}
\concept{Credentials}
\concept{Investment}
