% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks_report.R
\name{stocks_total_plot}
\alias{stocks_total_plot}
\title{Stocks Total Performance Plot}
\usage{
stocks_total_plot(stocks_perf, portfolio_perf, daily, trans, cash)
}
\arguments{
\item{stocks_perf}{Dataframe. Output of the stocks_performance function}

\item{portfolio_perf}{Dataframe. Output of the portfolio_performance function}

\item{daily}{Dataframe. Daily data}

\item{trans}{Dataframe. Transactions data}

\item{cash}{Dataframe. Cash data}
}
\description{
This function lets the user plot his stocks total performance
}
