% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summer.R
\name{summer}
\alias{summer}
\title{Sum Calculations and Plot}
\usage{
summer(df, ..., fun = sum, which = NA, top = NA, plot = FALSE)
}
\arguments{
\item{df}{Data.frame}

\item{...}{Variables. Variables you wish to process. Order matters.
If no variables are passed, the whole data.frame will be considered}

\item{fun}{Function (sum by default)}

\item{which}{Character vector. Select variables to sum. If not set,
all numeric variables will be returned}

\item{top}{Integer. Filter and plot the n largest n values by group}

\item{plot}{Boolean. Do you want to see a plot? Three variables tops.}
}
\description{
This function lets the user group, sum, calculate percentages 
and cumulatives. It also plots results if needed. Tidyverse friendly.
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gain_lift}()},
\code{\link{lasso_vars}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{tree_var}()},
\code{\link{trendsRelated}()}

Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{gg_pie}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()},
\code{\link{tree_var}()}
}
\concept{Exploratory}
\concept{Visualization}
