% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textCloud}
\alias{textCloud}
\title{Study the distribution of a target variable vs another variable. This
function is quite similar to the funModeling's corrplot function.}
\usage{
textCloud(text, lang = "english", exclude = c(), seed = 0,
  print = T, keep_spaces = FALSE)
}
\arguments{
\item{text}{Character vector}

\item{lang}{Character. Language in text (used for stop words)}

\item{exclude}{Character vector. Which word do you wish to exclude?}

\item{seed}{Numeric. Seed for re-producible plots}

\item{print}{Boolean. Plot results as textcloud?}

\item{keep_spaces}{Boolean. If you wish to keep spaces in each line
to keep unique compount words, separated with spaces, set to TRUE. 
For example, 'LA ALAMEDA' will be set as 'LA_ALAMEDA' and treated as
a single word.}
}
\description{
Study the distribution of a target variable vs another variable. This
function is quite similar to the funModeling's corrplot function.
}
