% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textCloud}
\alias{textCloud}
\title{Wordcloud Plot}
\usage{
textCloud(text, lang = "english", exclude = c(), seed = 0,
  keep_spaces = FALSE, min = 2, pal = NA, print = TRUE)
}
\arguments{
\item{text}{Character vector}

\item{lang}{Character. Language in text (used for stop words)}

\item{exclude}{Character vector. Which word do you wish to exclude?}

\item{seed}{Numeric. Seed for re-producible plots}

\item{keep_spaces}{Boolean. If you wish to keep spaces in each line
to keep unique compount words, separated with spaces, set to TRUE. 
For example, 'LA ALAMEDA' will be set as 'LA_ALAMEDA' and treated as
a single word.}

\item{min}{Integer. Words with less frequency will not be plotted}

\item{pal}{Character vector. Which colours do you wish to use}

\item{print}{Boolean. Plot results as textcloud?}
}
\description{
Study the distribution of a target variable vs another variable. This
function is quite similar to the funModeling's corrplot function.
}
\seealso{
Other Visualization: \code{\link{corr_plot}},
  \code{\link{mplot_conf}}, \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_gain}},
  \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{theme_lares2}},
  \code{\link{theme_lares}}

Other Exploratory: \code{\link{corr_var}},
  \code{\link{crosstab}}, \code{\link{df_str}},
  \code{\link{distr}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{missingness}}, \code{\link{textTokenizer}},
  \code{\link{trendsRelated}}

Other Text Mining: \code{\link{cleanText}},
  \code{\link{replaceall}}, \code{\link{textFeats}},
  \code{\link{textTokenizer}}
}
\concept{Exploratory}
\concept{Text Mining}
\concept{Visualization}
