% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textFeats}
\alias{textFeats}
\title{Create features out of text}
\usage{
textFeats(text, auto = TRUE, contains = NA)
}
\arguments{
\item{text}{Character vector}

\item{auto}{Boolean. Auto create some useful parameters?}

\item{contains}{Character vector. Which columns do you wish to add
with a contains (counter) string validator?}
}
\description{
This function creates a data.frame with features based on a text vector
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohse}()},
\code{\link{one_hot_encoding_commas}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()},
\code{\link{year_week}()}

Other Text Mining: 
\code{\link{cleanText}()},
\code{\link{replaceall}()},
\code{\link{sentimentBreakdown}()},
\code{\link{textCloud}()},
\code{\link{textTokenizer}()},
\code{\link{topics_rake}()}
}
\concept{Data Wrangling}
\concept{Text Mining}
