% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{textTokenizer}
\alias{textTokenizer}
\title{Tokenize Vectors into Words}
\usage{
textTokenizer(text, lang = "english", exclude = c(),
  keep_spaces = FALSE, df = FALSE, min = 2)
}
\arguments{
\item{text}{Character vector}

\item{lang}{Character. Language in text (used for stop words)}

\item{exclude}{Character vector. Which word do you wish to exclude?}

\item{keep_spaces}{Boolean. If you wish to keep spaces in each line
to keep unique compount words, separated with spaces, set to TRUE. 
For example, 'LA ALAMEDA' will be set as 'LA_ALAMEDA' and treated as
a single word.}

\item{df}{Boolean. Return a dataframe with a one-hot-encoding kind of
results? Each word is a column and returns if word is contained.}

\item{min}{Integer. If df = TRUE, what is the minimum frequency for
the word to be considered.}
}
\description{
This function transforms texts into words, calculate frequencies,
supress stop words in a given language.
}
\seealso{
Other Exploratory: \code{\link{corr_var}},
  \code{\link{crosstab}}, \code{\link{df_str}},
  \code{\link{distr}}, \code{\link{freqs}},
  \code{\link{gain_lift}}, \code{\link{get_tweets}},
  \code{\link{missingness}}, \code{\link{textCloud}},
  \code{\link{trendsRelated}}

Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{calibrate}}, \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{date_feats}}, \code{\link{dateformat}},
  \code{\link{formatNum}}, \code{\link{holidays}},
  \code{\link{left}}, \code{\link{normalize}},
  \code{\link{numericalonly}}, \code{\link{ohse}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{rbind_full}}, \code{\link{removenacols}},
  \code{\link{removenarows}}, \code{\link{replaceall}},
  \code{\link{right}}, \code{\link{vector2text}},
  \code{\link{year_month}}, \code{\link{year_week}}

Other Text Mining: \code{\link{cleanText}},
  \code{\link{replaceall}}, \code{\link{textCloud}}
}
\concept{Data Wrangling}
\concept{Exploratory}
\concept{Text Mining}
