% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{theme_lares}
\alias{theme_lares}
\title{lares Theme for ggplot2}
\usage{
theme_lares(labels = TRUE, colours = TRUE)
}
\arguments{
\item{labels}{Boolean. Add labels to plot?}

\item{colours}{Boolean. Personalize colour palettes?}
}
\description{
This function sets some default values into ggplot2 outputs. This
function is almost as using gg_colour_customs() + gg_fill_customs() 
+ gg_text_customs, but these won't be mantained any longer.
}
