% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{theme_lares}
\alias{theme_lares}
\title{Old theme for ggplot2 [Deprecated]}
\usage{
theme_lares(
  labels = FALSE,
  colours = TRUE,
  cont = FALSE,
  xcommas = FALSE,
  ycommas = FALSE
)
}
\arguments{
\item{labels}{Boolean. Add labels to plot?}

\item{colours}{Boolean. Personalize colour palettes?}

\item{cont}{Boolean. Is the value continuous? Discrete by default}

\item{xcommas}{Boolean. Nice format for x continuous values?}

\item{ycommas}{Boolean. Nice format for y continuous values?}
}
\description{
This function sets some default values into ggplot2 outputs. This
function is almost as using gg_colour_customs() + gg_fill_customs() 
+ gg_text_customs = but these won't be mantained any longer.
}
\details{
This function is lo longer mantained: use theme_lares2() instead!
}
\seealso{
Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{gg_pie}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{tree_var}()}
}
\concept{Visualization}
