% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{theme_lares}
\alias{theme_lares}
\title{lares Theme for ggplot2}
\usage{
theme_lares(labels = FALSE, colours = TRUE, cont = FALSE,
  xcommas = FALSE, ycommas = FALSE)
}
\arguments{
\item{labels}{Boolean. Add labels to plot?}

\item{colours}{Boolean. Personalize colour palettes?}

\item{cont}{Boolean. Is the value continuous? Discrete by default}

\item{xcommas}{Boolean. Nice format for x continuous values?}

\item{ycommas}{Boolean. Nice format for y continuous values?}
}
\description{
This function sets some default values into ggplot2 outputs. This
function is almost as using gg_colour_customs() + gg_fill_customs() 
+ gg_text_customs = but these won't be mantained any longer.
}
\seealso{
Other Visualization: \code{\link{corr_plot}},
  \code{\link{noPlot}}
}
\concept{Visualization}
