% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lares2.R
\name{theme_lares2}
\alias{theme_lares2}
\title{lares Theme for ggplot2}
\usage{
theme_lares2(font = "Arial Narrow", size = 12.5,
  main_colour = "darkorange3", hard_colour = "black",
  soft_colour = "grey30", bg_colour = "white",
  panel_colour = "transparent", legend = "right", grid = TRUE,
  mg = 14, pal = 0, which = "fct")
}
\arguments{
\item{font, size}{Character and numeric. Base font family and base size for texts}

\item{main_colour, hard_colour, soft_colour, bg_colour, panel_colour}{Character. Main colours for your theme}

\item{legend}{Character. Legend position: top, right, bottom, left}

\item{grid}{Character (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{mg}{Numeric. External margin}

\item{pal}{Integer. 1 for fill and colour palette, 2 for only colour palette,
3 for only fill palette, 4 for personal labels-colour palette. 0 or else for nothing.}

\item{which}{Character. When pal = 3, select which colours should be
added with the custom colours pelette: fill, colour, text (fct) - first letters}
}
\description{
Based on hrbrthemes' theme_ipsum and customized for lares' use.
}
\section{Why Arial Narrow?}{

First and foremost, Arial Narrow is generally installed by default or readily
available on any modern system, so it's "free"-ish; plus, it is a condensed font
with solid default kerning pairs and geometric numbers.
}

\seealso{
Other Visualization: \code{\link{corr_plot}},
  \code{\link{distr}}, \code{\link{freqs_df}},
  \code{\link{freqs}}, \code{\link{mplot_conf}},
  \code{\link{mplot_cuts_error}}, \code{\link{mplot_cuts}},
  \code{\link{mplot_density}}, \code{\link{mplot_full}},
  \code{\link{mplot_gain}}, \code{\link{mplot_importance}},
  \code{\link{mplot_lineal}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_response}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}, \code{\link{noPlot}},
  \code{\link{plot_survey}}, \code{\link{summer}},
  \code{\link{theme_lares}}, \code{\link{tree_var}}
}
\concept{Visualization}
