% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lares2.R
\name{theme_lares2}
\alias{theme_lares2}
\title{lares Theme for ggplot2}
\usage{
theme_lares2(font = "Arial Narrow", base_size = 12,
  main_colour = "darkorange3", second_colour = "deepskyblue3",
  soft_colour = "grey30", bg_colour = "transparent",
  legend = "right", mg = 15, pal = 0)
}
\arguments{
\item{font, base_size}{Character and numeric. Base font family and size}

\item{main_colour, second_colour, soft_colour, bg_colour}{Character.
Main colours for your theme}

\item{legend}{Character. Legend position: top, right, bottom, left}

\item{mg}{Numeric. External margin}

\item{pal}{Boolean. 1 for fill and colour palette, 2 for only colour palette,
3 for personal labels-colour palette}
}
\description{
Based on hrbrthemes' theme_ipsum: A precise & pristine \link{ggplot2} theme
with opinionated defaults and an emphasis on typography.
}
\section{Why Arial Narrow?}{

First and foremost, Arial Narrow is generally installed by default or readily
available on any modern system, so it's "free"-ish; plus, it is a condensed font
with solid default kerning pairs and geometric numbers.
}

