% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{tic}
\alias{tic}
\alias{toc}
\alias{resettictoc}
\title{Stopwatch to measure R Timings}
\usage{
tic(id = 1, quiet = TRUE)

toc(id = 1, msg = "Elapsed time:", units = TRUE, signif = 3, quiet = FALSE)

resettictoc(which = "both")
}
\arguments{
\item{id}{Define ID if multiple tic() & toc() are being used}

\item{quiet}{Boolean. Quiet messages?}

\item{msg}{Character. Custom message shown}

\item{units}{Boolean. Do you want nice format for the time units? 
If not, seconds elapsed as numerical values}

\item{signif}{Integer. Significant digits}

\item{which}{Character. Select: both, tic, toc}
}
\value{
\code{toc} returns an (invisible) list containing the timestamps
\code{tic} and \code{toc}, \code{time} in seconds and the message \code{msg}.
}
\description{
Start a stopwatch.

Stop a stopwatch.

Reset all tic and toc values in your environment.
}
\examples{
# Basic use (global stopwatch)
tic()
Sys.sleep(0.1)
toc()

# Multiple tic tocs
tic(id = "two", quiet = FALSE)
Sys.sleep(0.2)
toc(id = "two")

# Global is still working (id = 1)
toc(msg = "The function finished its work in")
}
\concept{Time}
