% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{tic}
\alias{tic}
\alias{toc}
\title{Stopwatch to measure timings in R}
\usage{
tic(id = 1, quiet = TRUE)

toc(id = 1, msg = "Elapsed time:", units = TRUE, signif = 3, quiet = FALSE)
}
\arguments{
\item{id}{Define ID if multiple \code{tic} & \code{toc} are being used}

\item{quiet}{Boolean. Quiet messages?}

\item{msg}{Character. Custom message shown}

\item{units}{Boolean. Do you want nice format for the time units? 
If not, seconds elapsed as numerical values}

\item{signif}{Integer. Significant digits}
}
\value{
\code{toc} returns an (invisible) list containing the timestamps
\code{tic} and \code{toc}, \code{time} in seconds and the message \code{msg}.
}
\description{
Start a stopwatch.

Stop a stopwatch.
}
\examples{
# Basic use (global stopwatch)
tic()
Sys.sleep(0.1)
toc()

# Multiple tic tocs
tic(id = "two", quiet = FALSE)
Sys.sleep(0.2)
toc(id = "two")

# Global is still working (id = 1)
toc(msg = "The function finished its work in")
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{zerovar}()}
}
\concept{Time}
\concept{Tools}
