% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{tic}
\alias{tic}
\alias{toc}
\alias{resettictoc}
\title{Stopwatch Start}
\usage{
tic(id = 1, quiet = TRUE)

toc(id = 1, msg = "Elapsed time:", quiet = FALSE)

resettictoc(which = "both")
}
\arguments{
\item{id}{Define ID if multiple tic() & toc() are being used}

\item{quiet}{Boolean. Quiet messages?}

\item{msg}{Character. Custom message shown}

\item{which}{Character. Select: both, tic, toc}
}
\value{
\code{toc} returns an (invisible) list containing the timestamps
\code{tic} and \code{toc}, \code{time} in seconds and the message \code{msg}.
}
\description{
Measure timings between a tic and a toc.

Measure timings between a tic and a toc.

Measure timings between a tic and a toc. Reset all values
}
\examples{
\dontrun{
# Basic use
tic(quiet = FALSE)
toc()
getOption("tic")

# Multiple tic tocs
tic(id = 123, quiet = FALSE)
toc(id = 123)
toc(id = 1)
}
}
\concept{Time}
