% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_var}
\alias{tree_var}
\title{Recursive Partitioning and Regression Trees}
\usage{
tree_var(
  df,
  target,
  max = 3,
  min = 20,
  cp = 0,
  size = 0.7,
  ohse = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{target}{Variable}

\item{max}{Integer. Maximal depth of the tree}

\item{min}{Integer. The minimum number of observations that must 
exist in a node in order for a split to be attempted}

\item{cp}{Numeric. Complexity parameter}

\item{size}{Numeric. Textsize of plot}

\item{ohse}{Boolean. Auto generate One Hot Smart Encoding?}

\item{plot}{Boolean. Return a plot? If not, rpart object}

\item{...}{rpart.plot custom parameters}
}
\description{
Fit and plot a rpart model for exploratory purposes using rpart
and rpart.plot libraries. Idea from explore library.
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs}()},
\code{\link{gain_lift}()},
\code{\link{get_tweets}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{summer}()},
\code{\link{trendsRelated}()}

Other Visualization: 
\code{\link{corr_plot}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs}()},
\code{\link{mplot_conf}()},
\code{\link{mplot_cuts_error}()},
\code{\link{mplot_cuts}()},
\code{\link{mplot_density}()},
\code{\link{mplot_full}()},
\code{\link{mplot_gain}()},
\code{\link{mplot_importance}()},
\code{\link{mplot_lineal}()},
\code{\link{mplot_metrics}()},
\code{\link{mplot_response}()},
\code{\link{mplot_roc}()},
\code{\link{mplot_splits}()},
\code{\link{noPlot}()},
\code{\link{plot_survey}()},
\code{\link{summer}()},
\code{\link{theme_lares2}()},
\code{\link{theme_lares}()}
}
\concept{Exploratory}
\concept{Visualization}
