% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_var}
\alias{tree_var}
\title{Recursive Partitioning and Regression Trees}
\usage{
tree_var(
  df,
  y,
  max = 3,
  min = 15,
  cp = 0,
  ohse = TRUE,
  plot = TRUE,
  title = NA,
  subtitle = NULL,
  caption = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame}

\item{y}{Variable or Character. Name of the independent variable.}

\item{max}{Integer. Maximal depth of the tree.}

\item{min}{Integer. The minimum number of observations that must 
exist in a node in order for a split to be attempted.}

\item{cp}{Numeric. Complexity parameter.}

\item{ohse}{Boolean. Auto generate One Hot Smart Encoding?}

\item{plot}{Boolean. Return a plot? If not, \code{rpart} object.}

\item{title, subtitle}{Character. Title and subtitle to include in plot.
Set to \code{NULL} to ignore.}

\item{caption}{Boolean. Add interpretation text below plot?}

\item{...}{Additional parameters passed to \code{rpart.plot()}.}
}
\value{
List with plot (function), model, performance metrics, and
interpret auxiliary text.
}
\description{
Fit and plot a \code{rpart} model for exploratory purposes using
\code{rpart} and \code{rpart.plot} libraries.
}
\examples{
data(dft)
tree <- tree_var(dft, Fare, subtitle = "Titanic dataset")
tree$plot() # tree plot
tree$model # rpart model
tree$performance # metrics
tree_var(dft, Survived_TRUE)$plot()
}
\seealso{
Other Exploratory: 
\code{\link{corr_cross}()},
\code{\link{corr_var}()},
\code{\link{crosstab}()},
\code{\link{df_str}()},
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{lasso_vars}()},
\code{\link{missingness}()},
\code{\link{plot_cats}()},
\code{\link{plot_df}()},
\code{\link{plot_nums}()},
\code{\link{trendsRelated}()}

Other Visualization: 
\code{\link{distr}()},
\code{\link{freqs_df}()},
\code{\link{freqs_list}()},
\code{\link{freqs_plot}()},
\code{\link{freqs}()},
\code{\link{gg_bars}()},
\code{\link{gg_pie}()},
\code{\link{noPlot}()},
\code{\link{plot_chord}()},
\code{\link{plot_survey}()},
\code{\link{plot_timeline}()},
\code{\link{theme_lares}()}
}
\concept{Exploratory}
\concept{Visualization}
