% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeform.R
\name{typeform_download}
\alias{typeform_download}
\title{Download typeform data}
\usage{
typeform_download(account = "tabunga", form, creds = NA)
}
\arguments{
\item{account}{Character. Which of the accounts in config.yml 
do you wish to use?}

\item{form}{Character. Define which form you wish to import. Should
be the same as the form's name at the online platform}

\item{creds}{Character. Credential's user (see get_credentials)}
}
\description{
This function lets the user download surveys results from a
Typeform account.
}
