% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateLares}
\alias{updateLares}
\title{Update the library}
\usage{
updateLares(local = FALSE)
}
\arguments{
\item{local}{If set to TRUE then it will install package with local files}
}
\description{
This function lets the user update from repository or local source.
}
