% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateLares}
\alias{updateLares}
\title{Update the library}
\usage{
updateLares(local = FALSE, force = FALSE, restart = FALSE)
}
\arguments{
\item{local}{Boolean. Install package with local files (TRUE) or Github repository}

\item{local}{Boolean. Force install if needed}
}
\description{
This function lets the user update from repository or local source.
}
