% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateLares}
\alias{updateLares}
\title{Update the library}
\usage{
updateLares(local = FALSE, force = FALSE, n = TRUE, fb = FALSE, all = FALSE)
}
\arguments{
\item{local}{Boolean. Install package with local files (TRUE) or Github repository}

\item{force}{Boolean. Force install if needed}

\item{n}{Boolean. Notify update for lares' stats}

\item{fb}{Boolean. From FB instance?}

\item{all}{Boolean. Install other recommended libraries? Kinda Docker install!}
}
\description{
This function lets the user update from repository or local source.
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bindfiles}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{get_credentials}()},
\code{\link{h2o_predict_API}()},
\code{\link{h2o_predict_MOJO}()},
\code{\link{h2o_predict_binary}()},
\code{\link{h2o_predict_model}()},
\code{\link{h2o_selectmodel}()},
\code{\link{h2o_update}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_country}()},
\code{\link{iter_seeds}()},
\code{\link{json2vector}()},
\code{\link{listfiles}()},
\code{\link{mailSend}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{pass}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{slackSend}()},
\code{\link{statusbar}()},
\code{\link{try_require}()},
\code{\link{zerovar}()}
}
\concept{Tools}
