% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{vector2text}
\alias{vector2text}
\title{Convert a vector into a comma separated text}
\usage{
vector2text(vector, sep = ", ", quotes = TRUE)
}
\arguments{
\item{vector}{Vector. Vector with more than 1 observation}

\item{sep}{Character. String text wished to insert between values}

\item{quotes}{Boolean. Bring simple quotes for each 
observation (useful for SQL)}
}
\description{
Convert a vector into a comma separated text
}
\seealso{
Other Data Wrangling: \code{\link{balance_data}},
  \code{\link{categ_reducer}},
  \code{\link{categoryCounter}}, \code{\link{cleanText}},
  \code{\link{dateformat}}, \code{\link{formatNum}},
  \code{\link{left}}, \code{\link{missingness}},
  \code{\link{normalize}}, \code{\link{numericalonly}},
  \code{\link{one_hot_encoding_commas}},
  \code{\link{removenacols}}, \code{\link{removenarows}},
  \code{\link{replaceall}}, \code{\link{right}},
  \code{\link{year_month}}, \code{\link{year_week}}
}
\concept{Data Wrangling}
