% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{vector2text}
\alias{vector2text}
\title{Convert a vector into a comma separated text}
\usage{
vector2text(vector, sep = ", ")
}
\arguments{
\item{vector}{Vector. Vector with more than 1 observation}

\item{sep}{Character. String text wished to insert between values}
}
\description{
Convert a vector into a comma separated text
}
