% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{vector2text}
\alias{vector2text}
\title{Convert a vector into a comma separated text}
\usage{
vector2text(vector, sep = ", ", quotes = TRUE, and = "")
}
\arguments{
\item{vector}{Vector. Vector with more than 1 observation}

\item{sep}{Character. String text wished to insert between values}

\item{quotes}{Boolean. Bring simple quotes for each observation}

\item{and}{Character. Add 'and' or something before last observation. 
Not boolean variable so it can be used on other languages}
}
\description{
Convert a vector into a comma separated text
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohse}()},
\code{\link{one_hot_encoding_commas}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{year_month}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
