% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{warnifnot}
\alias{warnifnot}
\title{Test the Truth of R Expressions and Warn}
\usage{
warnifnot(...)
}
\arguments{
\item{...}{any R expression, which should evaluate to TRUE}
}
\description{
If the expression in ... is not \code{TRUE}, \code{warning} is called,
producing a warning message indicating the expression which was not true.
}
\examples{
warnifnot(TRUE)
warnifnot(FALSE)
warnifnot(1 + 1 == 3)
}
