% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{warnifnot}
\alias{warnifnot}
\title{Test the Truth of R Expressions and Warn}
\usage{
warnifnot(...)
}
\arguments{
\item{...}{any R expression, which should evaluate to TRUE}
}
\description{
If the expression in ... is not \code{TRUE}, \code{warning} is called,
producing a warning message indicating the expression which was not true.
}
\examples{
warnifnot(TRUE)
warnifnot(FALSE)
warnifnot(1 + 1 == 3)
}
\seealso{
Other Tools: 
\code{\link{autoline}()},
\code{\link{bind_files}()},
\code{\link{bring_api}()},
\code{\link{db_download}()},
\code{\link{db_upload}()},
\code{\link{export_plot}()},
\code{\link{export_results}()},
\code{\link{files_functions}()},
\code{\link{font_exists}()},
\code{\link{formatColoured}()},
\code{\link{formatHTML}()},
\code{\link{get_credentials}()},
\code{\link{glued}()},
\code{\link{grepm}()},
\code{\link{h2o_selectmodel}()},
\code{\link{haveInternet}()},
\code{\link{image_metadata}()},
\code{\link{importxlsx}()},
\code{\link{ip_data}()},
\code{\link{json2vector}()},
\code{\link{list_cats}()},
\code{\link{listfiles}()},
\code{\link{mail_send}()},
\code{\link{move_files}()},
\code{\link{msplit}()},
\code{\link{myip}()},
\code{\link{quiet}()},
\code{\link{read.file}()},
\code{\link{statusbar}()},
\code{\link{tic}()},
\code{\link{try_require}()},
\code{\link{updateLares}()},
\code{\link{what_size}()}
}
\concept{Tools}
