% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{writeGS}
\alias{writeGS}
\alias{writeGS4}
\title{Google Sheets Writing (API v4)}
\usage{
writeGS(
  data,
  title,
  sheet = "Hoja 1",
  range = "A1",
  reformat = FALSE,
  creds = NULL,
  server = FALSE,
  ...
)
}
\arguments{
\item{data}{Object (value, vector, dataframe, list)}

\item{title}{Character. Textual title of Google Sheet}

\item{sheet}{Character. Working sheet to import}

\item{range}{Character. A cell range to read from}

\item{reformat}{Boolean. Reformat the affected cells?}

\item{creds}{Character. JSON filename with service auth}

\item{server}{Boolean. Force interacting auth process?}

\item{...}{Further read_sheet parameters}
}
\description{
Write data into Google Sheets.
}
\seealso{
Other Scrapper: 
\code{\link{geoStratum}()},
\code{\link{get_mp3}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()}

Other Google: 
\code{\link{geoAddress}()},
\code{\link{queryGA}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()},
\code{\link{trendsTime}()}
}
\concept{Google}
\concept{Scrapper}
