% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{writeGS}
\alias{writeGS}
\title{Google Sheets Writing}
\usage{
writeGS(data, title, ws = "Hoja 1", cell = "A1", first_time = FALSE)
}
\arguments{
\item{data}{Data Frame. Data to export to Google Sheet}

\item{title}{Character. Textual title of Google Sheet}

\item{ws}{Character. Working sheet to export to}

\item{cell}{Character. In which cell should we paste the data (upper left cell)}

\item{first_time}{Boolean. Authenticate manualy}
}
\description{
This function lets the user write on any Google Sheet
}
\seealso{
Other Connection: \code{\link{get_tweets}},
  \code{\link{queryGA}}, \code{\link{readGS4}},
  \code{\link{readGS}}
}
\concept{Connection}
