% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{year_month}
\alias{year_month}
\title{Convert year month format YYYY-MM}
\usage{
year_month(date)
}
\arguments{
\item{date}{Date. Date we wish to transform into format YYYY-MM format}
}
\description{
This function lets the user convert a year month format into YYYY-MM
}
