% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{year_month}
\alias{year_month}
\title{Convert Date into Year-Month (YYYY-MM)}
\usage{
year_month(date)
}
\arguments{
\item{date}{Date vector. Date to transform format.}
}
\value{
Vector with dates reformatted
}
\description{
This function lets the user convert a date into YYYY-MM format
}
\examples{
year_month(Sys.Date())
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{formatNum}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohe_commas}()},
\code{\link{ohse}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_week}()}
}
\concept{Data Wrangling}
