% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{year_week}
\alias{year_week}
\title{Convert Date into Year-Week (YYYY-WW)}
\usage{
year_week(date)
}
\arguments{
\item{date}{Date. Date we wish to transform}
}
\description{
This function lets the user convert a date into YYYY-WW format
}
\seealso{
Other Data Wrangling: 
\code{\link{balance_data}()},
\code{\link{calibrate}()},
\code{\link{categ_reducer}()},
\code{\link{cleanText}()},
\code{\link{date_cuts}()},
\code{\link{date_feats}()},
\code{\link{dateformat}()},
\code{\link{formatNum}()},
\code{\link{formatTime}()},
\code{\link{holidays}()},
\code{\link{impute}()},
\code{\link{left}()},
\code{\link{normalize}()},
\code{\link{numericalonly}()},
\code{\link{ohse}()},
\code{\link{one_hot_encoding_commas}()},
\code{\link{rbind_full}()},
\code{\link{removenacols}()},
\code{\link{removenarows}()},
\code{\link{replaceall}()},
\code{\link{right}()},
\code{\link{textFeats}()},
\code{\link{textTokenizer}()},
\code{\link{vector2text}()},
\code{\link{year_month}()}
}
\concept{Data Wrangling}
