% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{yt2mp3}
\alias{yt2mp3}
\title{Get Personal Portfolio's Data}
\usage{
yt2mp3(
  url = "https://www.youtube.com/watch?v=lrlKcCdVw9Q",
  filename = NA,
  start = 0
)
}
\arguments{
\item{url}{Characeter. YouTube URL}

\item{filename}{Character. Specify output filename. If NA, video's name 
will be automatically used}

\item{start}{Numeric. Seeks in the input file to position (Must be 
greater than 0) and crops output}
}
\description{
This function lets me download my personal Excel with my Portfolio 
data, locally or from Dropbox. You can also use it for local files
}
\seealso{
Other Scrapper: 
\code{\link{fb_accounts}()},
\code{\link{fb_ads}()},
\code{\link{fb_insights}()},
\code{\link{fb_posts}()},
\code{\link{fb_post}()},
\code{\link{geoStratum}()},
\code{\link{holidays}()},
\code{\link{ip_country}()},
\code{\link{readGS4}()},
\code{\link{readGS}()},
\code{\link{trendsRelated}()},
\code{\link{writeGS}()}
}
\concept{Audio}
\concept{Scrapper}
