% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{zerovar}
\alias{zerovar}
\title{Zero Variance Columns}
\usage{
zerovar(df)
}
\arguments{
\item{df}{Dataframe}
}
\description{
This function quickly detectes which columns have the same value
for each observation
}
